main chank of moving about part of refactor done, now to apply all Pauls functionality and keep deleting all the mess I have created
