+++ client/src/components/MapPolygon.js
@@ -1,0 +1,11 @@
+import { Polygon } from 'react-leaflet'
+
+const MapPolygon = ({ coordinates }) => {
+
+    const purpleOptions = { color: 'purple' } 
+
+    return (
+        <Polygon pathOptions={purpleOptions} positions={coordinates} />
+    )
+}
+export default MapPolygon
