--- client/src/components/SimpleMap.js
@@ -1,8297 +1,0 @@
-import {
-    Circle,
-    CircleMarker,
-    MapContainer,
-    Polyline,
-    Polygon,
-    Popup,
-    Rectangle,
-    TileLayer,
-    Marker,
-    useMap
-  } from 'react-leaflet'
-  import MarkerElement from './MarkerElement'
-//   import entities from '../data/test_data.json'
-
-
-
-
-
-const SimpleMap = () => {
-    const coordinates = [
-        [-1.677834, 53.762191],
-        [-1.677364, 53.762155],
-        [-1.677035, 53.762073],
-        [-1.676895, 53.762002],
-        [-1.676687, 53.761811],
-        [-1.673832, 53.761461],
-        [-1.673837, 53.761264],
-        [-1.672088, 53.761514],
-        [-1.67169, 53.761008],
-        [-1.670415, 53.759716],
-        [-1.670675, 53.759619],
-        [-1.670715, 53.759388],
-        [-1.670706, 53.75888],
-        [-1.670669, 53.75867],
-        [-1.670577, 53.758455],
-        [-1.670456, 53.758363],
-        [-1.669943, 53.758721],
-        [-1.668563, 53.757947],
-        [-1.667977, 53.757652],
-        [-1.667412, 53.757421],
-        [-1.667769, 53.757035],
-        [-1.667228, 53.756808],
-        [-1.667519, 53.756598],
-        [-1.667897, 53.756231],
-        [-1.666764, 53.755848],
-        [-1.667086, 53.755545],
-        [-1.669205, 53.75551],
-        [-1.668678, 53.75498],
-        [-1.669364, 53.754418],
-        [-1.669363, 53.754359],
-        [-1.66939, 53.754351],
-        [-1.66954, 53.754508],
-        [-1.671224, 53.754022],
-        [-1.671297, 53.754103],
-        [-1.673855, 53.753289],
-        [-1.673834, 53.753325],
-        [-1.673968, 53.753362],
-        [-1.673944, 53.753392],
-        [-1.675215, 53.753652],
-        [-1.675349, 53.753323],
-        [-1.675417, 53.753282],
-        [-1.67503, 53.753062],
-        [-1.674379, 53.752787],
-        [-1.676959, 53.751996],
-        [-1.677887, 53.752267],
-        [-1.678126, 53.752363],
-        [-1.679982, 53.753546],
-        [-1.680099, 53.753611],
-        [-1.6802, 53.753561],
-        [-1.680461, 53.753683],
-        [-1.680733, 53.753746],
-        [-1.680754, 53.753777],
-        [-1.680724, 53.753828],
-        [-1.682253, 53.754245],
-        [-1.683436, 53.754005],
-        [-1.68336, 53.754183],
-        [-1.683411, 53.754501],
-        [-1.683698, 53.754874],
-        [-1.684037, 53.75571],
-        [-1.684607, 53.755705],
-        [-1.685463, 53.755581],
-        [-1.686718, 53.755671],
-        [-1.686829, 53.755715],
-        [-1.686896, 53.755654],
-        [-1.68785, 53.756081],
-        [-1.688271, 53.756232],
-        [-1.687267, 53.757342],
-        [-1.686089, 53.757216],
-        [-1.685939, 53.757667],
-        [-1.685782, 53.757861],
-        [-1.684472, 53.757485],
-        [-1.684187, 53.757895],
-        [-1.685187, 53.75815],
-        [-1.684884, 53.758562],
-        [-1.685261, 53.75869],
-        [-1.684469, 53.759231],
-        [-1.684548, 53.759279],
-        [-1.684549, 53.759318],
-        [-1.68448, 53.759302],
-        [-1.684355, 53.759389],
-        [-1.684411, 53.759417],
-        [-1.68403, 53.759695],
-        [-1.683963, 53.759696],
-        [-1.683848, 53.759599],
-        [-1.683775, 53.759577],
-        [-1.683095, 53.759647],
-        [-1.683172, 53.759878],
-        [-1.683051, 53.759891],
-        [-1.683033, 53.75984],
-        [-1.682866, 53.759859],
-        [-1.682888, 53.759927],
-        [-1.682817, 53.759942],
-        [-1.681876, 53.76005],
-        [-1.682533, 53.760796],
-        [-1.682754, 53.761003],
-        [-1.682246, 53.761325],
-        [-1.680161, 53.76102],
-        [-1.679704, 53.761076],
-        [-1.678169, 53.762068],
-        [-1.677834, 53.762191],
-      ]
-
-      const entities = [
-        {
-          "entry-date": "2017-09-29",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400000,
-          "name": "Battle of Adwalton Moor 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000000",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.677834, 53.762191],
-                    [-1.677364, 53.762155],
-                    [-1.677035, 53.762073],
-                    [-1.676895, 53.762002],
-                    [-1.676687, 53.761811],
-                    [-1.673832, 53.761461],
-                    [-1.673837, 53.761264],
-                    [-1.672088, 53.761514],
-                    [-1.67169, 53.761008],
-                    [-1.670415, 53.759716],
-                    [-1.670675, 53.759619],
-                    [-1.670715, 53.759388],
-                    [-1.670706, 53.75888],
-                    [-1.670669, 53.75867],
-                    [-1.670577, 53.758455],
-                    [-1.670456, 53.758363],
-                    [-1.669943, 53.758721],
-                    [-1.668563, 53.757947],
-                    [-1.667977, 53.757652],
-                    [-1.667412, 53.757421],
-                    [-1.667769, 53.757035],
-                    [-1.667228, 53.756808],
-                    [-1.667519, 53.756598],
-                    [-1.667897, 53.756231],
-                    [-1.666764, 53.755848],
-                    [-1.667086, 53.755545],
-                    [-1.669205, 53.75551],
-                    [-1.668678, 53.75498],
-                    [-1.669364, 53.754418],
-                    [-1.669363, 53.754359],
-                    [-1.66939, 53.754351],
-                    [-1.66954, 53.754508],
-                    [-1.671224, 53.754022],
-                    [-1.671297, 53.754103],
-                    [-1.673855, 53.753289],
-                    [-1.673834, 53.753325],
-                    [-1.673968, 53.753362],
-                    [-1.673944, 53.753392],
-                    [-1.675215, 53.753652],
-                    [-1.675349, 53.753323],
-                    [-1.675417, 53.753282],
-                    [-1.67503, 53.753062],
-                    [-1.674379, 53.752787],
-                    [-1.676959, 53.751996],
-                    [-1.677887, 53.752267],
-                    [-1.678126, 53.752363],
-                    [-1.679982, 53.753546],
-                    [-1.680099, 53.753611],
-                    [-1.6802, 53.753561],
-                    [-1.680461, 53.753683],
-                    [-1.680733, 53.753746],
-                    [-1.680754, 53.753777],
-                    [-1.680724, 53.753828],
-                    [-1.682253, 53.754245],
-                    [-1.683436, 53.754005],
-                    [-1.68336, 53.754183],
-                    [-1.683411, 53.754501],
-                    [-1.683698, 53.754874],
-                    [-1.684037, 53.75571],
-                    [-1.684607, 53.755705],
-                    [-1.685463, 53.755581],
-                    [-1.686718, 53.755671],
-                    [-1.686829, 53.755715],
-                    [-1.686896, 53.755654],
-                    [-1.68785, 53.756081],
-                    [-1.688271, 53.756232],
-                    [-1.687267, 53.757342],
-                    [-1.686089, 53.757216],
-                    [-1.685939, 53.757667],
-                    [-1.685782, 53.757861],
-                    [-1.684472, 53.757485],
-                    [-1.684187, 53.757895],
-                    [-1.685187, 53.75815],
-                    [-1.684884, 53.758562],
-                    [-1.685261, 53.75869],
-                    [-1.684469, 53.759231],
-                    [-1.684548, 53.759279],
-                    [-1.684549, 53.759318],
-                    [-1.68448, 53.759302],
-                    [-1.684355, 53.759389],
-                    [-1.684411, 53.759417],
-                    [-1.68403, 53.759695],
-                    [-1.683963, 53.759696],
-                    [-1.683848, 53.759599],
-                    [-1.683775, 53.759577],
-                    [-1.683095, 53.759647],
-                    [-1.683172, 53.759878],
-                    [-1.683051, 53.759891],
-                    [-1.683033, 53.75984],
-                    [-1.682866, 53.759859],
-                    [-1.682888, 53.759927],
-                    [-1.682817, 53.759942],
-                    [-1.681876, 53.76005],
-                    [-1.682533, 53.760796],
-                    [-1.682754, 53.761003],
-                    [-1.682246, 53.761325],
-                    [-1.680161, 53.76102],
-                    [-1.679704, 53.761076],
-                    [-1.678169, 53.762068],
-                    [-1.677834, 53.762191]
-                  ]
-                ],
-                [
-                  [
-                    [-1.66393, 53.75358],
-                    [-1.663179, 53.753069],
-                    [-1.663617, 53.752851],
-                    [-1.663389, 53.752695],
-                    [-1.663482, 53.752646],
-                    [-1.663408, 53.752597],
-                    [-1.663269, 53.75267],
-                    [-1.663105, 53.75256],
-                    [-1.662722, 53.752759],
-                    [-1.661833, 53.752161],
-                    [-1.658843, 53.752551],
-                    [-1.65823, 53.752709],
-                    [-1.658192, 53.752684],
-                    [-1.660469, 53.75147],
-                    [-1.660718, 53.751391],
-                    [-1.660736, 53.751333],
-                    [-1.661379, 53.750992],
-                    [-1.661519, 53.750948],
-                    [-1.661513, 53.750922],
-                    [-1.665076, 53.749026],
-                    [-1.66539, 53.74905],
-                    [-1.666104, 53.749563],
-                    [-1.666951, 53.75004],
-                    [-1.667907, 53.750724],
-                    [-1.667869, 53.750819],
-                    [-1.66745, 53.751043],
-                    [-1.667383, 53.751373],
-                    [-1.667332, 53.751449],
-                    [-1.665919, 53.752515],
-                    [-1.665955, 53.752664],
-                    [-1.666164, 53.752848],
-                    [-1.66625, 53.753022],
-                    [-1.666379, 53.753173],
-                    [-1.6668, 53.753276],
-                    [-1.667535, 53.753652],
-                    [-1.666725, 53.753883],
-                    [-1.666639, 53.753931],
-                    [-1.666649, 53.75412],
-                    [-1.666584, 53.754377],
-                    [-1.666735, 53.754937],
-                    [-1.666842, 53.755023],
-                    [-1.666828, 53.755059],
-                    [-1.666151, 53.755325],
-                    [-1.666027, 53.75534],
-                    [-1.665956, 53.75531],
-                    [-1.664789, 53.754283],
-                    [-1.66393, 53.75358]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2017-09-29",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400000,
-              "name": "Battle of Adwalton Moor 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000000",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000000"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000000"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400001,
-          "name": "Battle of Barnet 1471",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000001",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.207293, 51.668637],
-                    [-0.20593, 51.668622],
-                    [-0.204521, 51.668428],
-                    [-0.203372, 51.66818],
-                    [-0.203183, 51.668066],
-                    [-0.202208, 51.667769],
-                    [-0.198978, 51.668172],
-                    [-0.198913, 51.667882],
-                    [-0.19891, 51.667727],
-                    [-0.198921, 51.667514],
-                    [-0.199035, 51.666983],
-                    [-0.197991, 51.666841],
-                    [-0.197954, 51.666575],
-                    [-0.197663, 51.666471],
-                    [-0.196691, 51.666368],
-                    [-0.196094, 51.666353],
-                    [-0.196716, 51.665485],
-                    [-0.192361, 51.664684],
-                    [-0.191825, 51.664483],
-                    [-0.191381, 51.664213],
-                    [-0.191115, 51.663911],
-                    [-0.191036, 51.663774],
-                    [-0.190979, 51.663604],
-                    [-0.190949, 51.663374],
-                    [-0.192122, 51.662897],
-                    [-0.192028, 51.662779],
-                    [-0.19183, 51.662412],
-                    [-0.193015, 51.662119],
-                    [-0.193063, 51.6619],
-                    [-0.192878, 51.661901],
-                    [-0.192716, 51.661833],
-                    [-0.192674, 51.661721],
-                    [-0.192678, 51.66163],
-                    [-0.192834, 51.661558],
-                    [-0.192977, 51.661438],
-                    [-0.193201, 51.661105],
-                    [-0.193475, 51.660976],
-                    [-0.193613, 51.660739],
-                    [-0.19347, 51.660282],
-                    [-0.193401, 51.659835],
-                    [-0.193147, 51.659129],
-                    [-0.193075, 51.658792],
-                    [-0.193029, 51.658326],
-                    [-0.193189, 51.658263],
-                    [-0.193177, 51.658077],
-                    [-0.193151, 51.657914],
-                    [-0.193095, 51.657836],
-                    [-0.193053, 51.657509],
-                    [-0.192953, 51.65718],
-                    [-0.192809, 51.656831],
-                    [-0.192724, 51.656782],
-                    [-0.192729, 51.656755],
-                    [-0.192605, 51.656642],
-                    [-0.192289, 51.65623],
-                    [-0.192158, 51.655977],
-                    [-0.192135, 51.65585],
-                    [-0.192051, 51.655797],
-                    [-0.191861, 51.655518],
-                    [-0.19184, 51.655359],
-                    [-0.191788, 51.655269],
-                    [-0.191801, 51.655215],
-                    [-0.192085, 51.65529],
-                    [-0.192593, 51.655331],
-                    [-0.194427, 51.655044],
-                    [-0.195004, 51.655065],
-                    [-0.195987, 51.654827],
-                    [-0.198284, 51.658001],
-                    [-0.199193, 51.658463],
-                    [-0.199391, 51.658234],
-                    [-0.19944, 51.658251],
-                    [-0.199521, 51.658169],
-                    [-0.199657, 51.658138],
-                    [-0.200365, 51.658386],
-                    [-0.200952, 51.658525],
-                    [-0.200875, 51.658665],
-                    [-0.201509, 51.658697],
-                    [-0.201418, 51.659229],
-                    [-0.201349, 51.65943],
-                    [-0.201423, 51.659449],
-                    [-0.201354, 51.659715],
-                    [-0.202446, 51.659783],
-                    [-0.202426, 51.660294],
-                    [-0.204507, 51.66025],
-                    [-0.205867, 51.660069],
-                    [-0.20973, 51.665379],
-                    [-0.211143, 51.666996],
-                    [-0.207796, 51.668659],
-                    [-0.207293, 51.668637]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400001,
-              "name": "Battle of Barnet 1471",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000001",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000001"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000001"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400002,
-          "name": "Battle of Blore Heath 1459",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000002",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.415877, 52.91023],
-                    [-2.417167, 52.908741],
-                    [-2.417654, 52.909003],
-                    [-2.418084, 52.909126],
-                    [-2.418493, 52.90919],
-                    [-2.41864, 52.908975],
-                    [-2.418717, 52.908812],
-                    [-2.418687, 52.908802],
-                    [-2.419267, 52.907984],
-                    [-2.424812, 52.909004],
-                    [-2.425552, 52.909197],
-                    [-2.430361, 52.909807],
-                    [-2.430318, 52.909829],
-                    [-2.430318, 52.909893],
-                    [-2.430427, 52.909971],
-                    [-2.430542, 52.909972],
-                    [-2.430555, 52.909949],
-                    [-2.431222, 52.910204],
-                    [-2.431604, 52.910317],
-                    [-2.432117, 52.91044],
-                    [-2.432414, 52.910479],
-                    [-2.436491, 52.911607],
-                    [-2.438989, 52.912918],
-                    [-2.440041, 52.914206],
-                    [-2.440732, 52.916197],
-                    [-2.439277, 52.916982],
-                    [-2.434672, 52.917778],
-                    [-2.429701, 52.918144],
-                    [-2.427279, 52.919438],
-                    [-2.423154, 52.918344],
-                    [-2.417437, 52.916972],
-                    [-2.415137, 52.914022],
-                    [-2.414435, 52.913586],
-                    [-2.414671, 52.913302],
-                    [-2.414642, 52.912068],
-                    [-2.415877, 52.91023]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400002,
-              "name": "Battle of Blore Heath 1459",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000002",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000002"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000002"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400003,
-          "name": "Battle of Boroughbridge 1322",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000003",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.385521, 54.101375],
-                    [-1.384802, 54.097911],
-                    [-1.390304, 54.096849],
-                    [-1.391264, 54.096633],
-                    [-1.391699, 54.097646],
-                    [-1.391743, 54.097814],
-                    [-1.391733, 54.097911],
-                    [-1.391912, 54.098249],
-                    [-1.392614, 54.097889],
-                    [-1.393279, 54.097624],
-                    [-1.393932, 54.097463],
-                    [-1.394522, 54.097272],
-                    [-1.395203, 54.097169],
-                    [-1.395579, 54.097153],
-                    [-1.397987, 54.097339],
-                    [-1.398052, 54.097516],
-                    [-1.398164, 54.097662],
-                    [-1.398708, 54.098065],
-                    [-1.399441, 54.098422],
-                    [-1.399275, 54.098469],
-                    [-1.397985, 54.098493],
-                    [-1.397962, 54.098535],
-                    [-1.398074, 54.098692],
-                    [-1.396687, 54.098809],
-                    [-1.396514, 54.098866],
-                    [-1.396203, 54.099486],
-                    [-1.393502, 54.099736],
-                    [-1.392346, 54.099883],
-                    [-1.391328, 54.100095],
-                    [-1.391475, 54.100618],
-                    [-1.389821, 54.100942],
-                    [-1.389071, 54.100924],
-                    [-1.388715, 54.101013],
-                    [-1.38726, 54.101494],
-                    [-1.38678, 54.101751],
-                    [-1.386247, 54.101975],
-                    [-1.385751, 54.101526],
-                    [-1.385521, 54.101375]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400003,
-              "name": "Battle of Boroughbridge 1322",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000003",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000003"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000003"
-          }
-        },
-        {
-          "entry-date": "2013-06-07",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400004,
-          "name": "Battle of Bosworth (Field) 1485",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000004",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.460157, 52.583608],
-                    [-1.459831, 52.583777],
-                    [-1.459346, 52.583921],
-                    [-1.458748, 52.584069],
-                    [-1.458655, 52.584063],
-                    [-1.458632, 52.584094],
-                    [-1.457274, 52.584415],
-                    [-1.451638, 52.585534],
-                    [-1.449173, 52.586682],
-                    [-1.454228, 52.590091],
-                    [-1.45415, 52.59014],
-                    [-1.449565, 52.591672],
-                    [-1.44941, 52.591619],
-                    [-1.449389, 52.591743],
-                    [-1.449101, 52.591826],
-                    [-1.448563, 52.591879],
-                    [-1.447796, 52.591905],
-                    [-1.447686, 52.5916],
-                    [-1.445858, 52.591743],
-                    [-1.445733, 52.591351],
-                    [-1.44383, 52.591693],
-                    [-1.441136, 52.59143],
-                    [-1.439424, 52.591142],
-                    [-1.439019, 52.592054],
-                    [-1.438931, 52.592519],
-                    [-1.438904, 52.593049],
-                    [-1.436987, 52.593187],
-                    [-1.436754, 52.593688],
-                    [-1.436313, 52.594356],
-                    [-1.437017, 52.594704],
-                    [-1.436251, 52.595327],
-                    [-1.4363, 52.595381],
-                    [-1.436206, 52.595452],
-                    [-1.436234, 52.595476],
-                    [-1.436051, 52.595506],
-                    [-1.436092, 52.595565],
-                    [-1.434935, 52.596289],
-                    [-1.434459, 52.596191],
-                    [-1.434398, 52.596229],
-                    [-1.432288, 52.595241],
-                    [-1.432053, 52.595198],
-                    [-1.431914, 52.595497],
-                    [-1.431466, 52.595945],
-                    [-1.431216, 52.596306],
-                    [-1.431141, 52.596323],
-                    [-1.430753, 52.596976],
-                    [-1.430776, 52.59708],
-                    [-1.429345, 52.597293],
-                    [-1.428079, 52.59763],
-                    [-1.428013, 52.59768],
-                    [-1.428004, 52.597763],
-                    [-1.42804, 52.597854],
-                    [-1.427977, 52.597876],
-                    [-1.42798, 52.5979],
-                    [-1.427838, 52.597918],
-                    [-1.426901, 52.598236],
-                    [-1.426707, 52.598188],
-                    [-1.426491, 52.598222],
-                    [-1.425776, 52.598595],
-                    [-1.425736, 52.598638],
-                    [-1.425759, 52.598693],
-                    [-1.425063, 52.598837],
-                    [-1.424564, 52.59901],
-                    [-1.424124, 52.599108],
-                    [-1.423535, 52.599295],
-                    [-1.423085, 52.599476],
-                    [-1.423123, 52.5995],
-                    [-1.423067, 52.599529],
-                    [-1.423034, 52.599511],
-                    [-1.422656, 52.599703],
-                    [-1.421948, 52.60018],
-                    [-1.420628, 52.600815],
-                    [-1.42059, 52.600865],
-                    [-1.420516, 52.600837],
-                    [-1.419852, 52.601168],
-                    [-1.418231, 52.601825],
-                    [-1.41791, 52.601823],
-                    [-1.415611, 52.602004],
-                    [-1.415501, 52.601976],
-                    [-1.414357, 52.602022],
-                    [-1.412465, 52.602006],
-                    [-1.395444, 52.600629],
-                    [-1.395245, 52.600629],
-                    [-1.390696, 52.601314],
-                    [-1.389087, 52.601285],
-                    [-1.388673, 52.601202],
-                    [-1.388035, 52.600799],
-                    [-1.388166, 52.600691],
-                    [-1.388351, 52.600456],
-                    [-1.388612, 52.600553],
-                    [-1.388713, 52.600456],
-                    [-1.388657, 52.600439],
-                    [-1.389023, 52.600077],
-                    [-1.388523, 52.599978],
-                    [-1.387967, 52.599786],
-                    [-1.387961, 52.599752],
-                    [-1.387883, 52.599728],
-                    [-1.38813, 52.599375],
-                    [-1.386908, 52.598997],
-                    [-1.387023, 52.598835],
-                    [-1.386712, 52.598705],
-                    [-1.386669, 52.598743],
-                    [-1.386316, 52.59864],
-                    [-1.386116, 52.598542],
-                    [-1.386048, 52.59859],
-                    [-1.38603, 52.59873],
-                    [-1.385985, 52.598767],
-                    [-1.385753, 52.598808],
-                    [-1.385584, 52.598874],
-                    [-1.385219, 52.598676],
-                    [-1.384212, 52.597874],
-                    [-1.383588, 52.598178],
-                    [-1.384246, 52.598684],
-                    [-1.383195, 52.599524],
-                    [-1.383183, 52.599564],
-                    [-1.382952, 52.599762],
-                    [-1.382214, 52.59962],
-                    [-1.381892, 52.599672],
-                    [-1.381792, 52.599802],
-                    [-1.381162, 52.59966],
-                    [-1.37991, 52.599256],
-                    [-1.379832, 52.599277],
-                    [-1.379761, 52.599208],
-                    [-1.378916, 52.59901],
-                    [-1.379266, 52.598511],
-                    [-1.379498, 52.598027],
-                    [-1.378912, 52.597793],
-                    [-1.377254, 52.596993],
-                    [-1.377347, 52.596846],
-                    [-1.377537, 52.596721],
-                    [-1.377576, 52.596643],
-                    [-1.377636, 52.596614],
-                    [-1.377844, 52.596385],
-                    [-1.37802, 52.596306],
-                    [-1.378017, 52.596163],
-                    [-1.378072, 52.59606],
-                    [-1.378081, 52.595887],
-                    [-1.378837, 52.594693],
-                    [-1.379006, 52.594528],
-                    [-1.379142, 52.594312],
-                    [-1.380287, 52.594596],
-                    [-1.380297, 52.59454],
-                    [-1.383306, 52.595328],
-                    [-1.384572, 52.595606],
-                    [-1.386314, 52.593368],
-                    [-1.389428, 52.594198],
-                    [-1.391105, 52.594083],
-                    [-1.391179, 52.594095],
-                    [-1.39117, 52.594131],
-                    [-1.392059, 52.594312],
-                    [-1.395516, 52.595152],
-                    [-1.396416, 52.595302],
-                    [-1.397194, 52.595354],
-                    [-1.398099, 52.595518],
-                    [-1.398998, 52.593814],
-                    [-1.399138, 52.591186],
-                    [-1.400395, 52.59132],
-                    [-1.402693, 52.590438],
-                    [-1.404786, 52.589735],
-                    [-1.407061, 52.588722],
-                    [-1.407232, 52.5888],
-                    [-1.407759, 52.588475],
-                    [-1.408675, 52.587809],
-                    [-1.408258, 52.587794],
-                    [-1.407672, 52.587583],
-                    [-1.407422, 52.587581],
-                    [-1.406893, 52.587472],
-                    [-1.405286, 52.587022],
-                    [-1.40409, 52.586824],
-                    [-1.403147, 52.586761],
-                    [-1.402562, 52.586776],
-                    [-1.402273, 52.586867],
-                    [-1.402111, 52.586841],
-                    [-1.402075, 52.586617],
-                    [-1.402023, 52.586572],
-                    [-1.401982, 52.586467],
-                    [-1.401895, 52.586415],
-                    [-1.402188, 52.586268],
-                    [-1.404075, 52.583472],
-                    [-1.404268, 52.58328],
-                    [-1.404294, 52.583223],
-                    [-1.404277, 52.583184],
-                    [-1.405453, 52.581445],
-                    [-1.405394, 52.581415],
-                    [-1.40591, 52.580996],
-                    [-1.406526, 52.580517],
-                    [-1.407407, 52.579994],
-                    [-1.408026, 52.579667],
-                    [-1.408002, 52.579605],
-                    [-1.408055, 52.579477],
-                    [-1.409866, 52.57896],
-                    [-1.411238, 52.578949],
-                    [-1.411278, 52.578143],
-                    [-1.411313, 52.577978],
-                    [-1.411196, 52.576901],
-                    [-1.41107, 52.576176],
-                    [-1.411133, 52.575943],
-                    [-1.411286, 52.575748],
-                    [-1.411412, 52.57566],
-                    [-1.411677, 52.575557],
-                    [-1.411948, 52.575503],
-                    [-1.412556, 52.57546],
-                    [-1.414506, 52.575653],
-                    [-1.41502, 52.575676],
-                    [-1.414991, 52.575499],
-                    [-1.414796, 52.574894],
-                    [-1.4149, 52.574796],
-                    [-1.414905, 52.574626],
-                    [-1.414776, 52.573749],
-                    [-1.414633, 52.573219],
-                    [-1.414581, 52.57313],
-                    [-1.414572, 52.573022],
-                    [-1.414705, 52.573012],
-                    [-1.41464, 52.572739],
-                    [-1.415531, 52.572693],
-                    [-1.415541, 52.572366],
-                    [-1.417127, 52.57235],
-                    [-1.417131, 52.572245],
-                    [-1.417873, 52.57223],
-                    [-1.41801, 52.572196],
-                    [-1.41787, 52.571936],
-                    [-1.418311, 52.571943],
-                    [-1.418286, 52.571528],
-                    [-1.418376, 52.571526],
-                    [-1.418365, 52.571435],
-                    [-1.418829, 52.571441],
-                    [-1.418868, 52.571744],
-                    [-1.419459, 52.571683],
-                    [-1.419989, 52.571741],
-                    [-1.420036, 52.571708],
-                    [-1.420067, 52.571613],
-                    [-1.420165, 52.571548],
-                    [-1.42069, 52.571575],
-                    [-1.420849, 52.571701],
-                    [-1.421183, 52.571849],
-                    [-1.421275, 52.571851],
-                    [-1.421339, 52.571976],
-                    [-1.421327, 52.571947],
-                    [-1.421531, 52.571909],
-                    [-1.421429, 52.571724],
-                    [-1.421308, 52.571286],
-                    [-1.422338, 52.571355],
-                    [-1.422345, 52.571301],
-                    [-1.423532, 52.571356],
-                    [-1.423638, 52.571314],
-                    [-1.424084, 52.571292],
-                    [-1.424878, 52.571306],
-                    [-1.42584, 52.571268],
-                    [-1.427302, 52.571147],
-                    [-1.427742, 52.57107],
-                    [-1.428626, 52.57125],
-                    [-1.430398, 52.571746],
-                    [-1.430758, 52.571824],
-                    [-1.430995, 52.571832],
-                    [-1.430668, 52.569535],
-                    [-1.430578, 52.569137],
-                    [-1.431659, 52.568218],
-                    [-1.432647, 52.568581],
-                    [-1.434273, 52.566929],
-                    [-1.435181, 52.567327],
-                    [-1.436744, 52.568107],
-                    [-1.436763, 52.568091],
-                    [-1.437776, 52.568382],
-                    [-1.440995, 52.570143],
-                    [-1.441298, 52.570384],
-                    [-1.441437, 52.570458],
-                    [-1.441556, 52.570366],
-                    [-1.441614, 52.570386],
-                    [-1.441651, 52.570221],
-                    [-1.441734, 52.570135],
-                    [-1.442035, 52.569881],
-                    [-1.442552, 52.56952],
-                    [-1.443823, 52.570115],
-                    [-1.444477, 52.570501],
-                    [-1.445179, 52.570848],
-                    [-1.447186, 52.571648],
-                    [-1.447777, 52.571048],
-                    [-1.448096, 52.571194],
-                    [-1.449923, 52.570474],
-                    [-1.450642, 52.571172],
-                    [-1.450731, 52.571139],
-                    [-1.452262, 52.572979],
-                    [-1.452765, 52.57286],
-                    [-1.453654, 52.572467],
-                    [-1.453784, 52.572484],
-                    [-1.453933, 52.572434],
-                    [-1.454922, 52.57196],
-                    [-1.456103, 52.573559],
-                    [-1.456838, 52.574611],
-                    [-1.457449, 52.57556],
-                    [-1.457504, 52.575697],
-                    [-1.458144, 52.57547],
-                    [-1.458595, 52.576633],
-                    [-1.459775, 52.576596],
-                    [-1.459295, 52.578709],
-                    [-1.458417, 52.578464],
-                    [-1.458359, 52.578522],
-                    [-1.457922, 52.579317],
-                    [-1.457277, 52.58022],
-                    [-1.457778, 52.580384],
-                    [-1.45731, 52.58075],
-                    [-1.456762, 52.581101],
-                    [-1.457502, 52.581925],
-                    [-1.458281, 52.582707],
-                    [-1.458521, 52.582848],
-                    [-1.460157, 52.583608]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2013-06-07",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400004,
-              "name": "Battle of Bosworth (Field) 1485",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000004",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000004"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000004"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400005,
-          "name": "Battle of Braddock Down 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000005",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-4.573112, 50.441127],
-                    [-4.571395, 50.441315],
-                    [-4.571125, 50.441383],
-                    [-4.569054, 50.441628],
-                    [-4.56718, 50.441777],
-                    [-4.565958, 50.441802],
-                    [-4.564843, 50.441676],
-                    [-4.564743, 50.441132],
-                    [-4.564317, 50.440102],
-                    [-4.564219, 50.439775],
-                    [-4.564392, 50.439161],
-                    [-4.564928, 50.438416],
-                    [-4.565552, 50.437724],
-                    [-4.566125, 50.437248],
-                    [-4.567849, 50.436179],
-                    [-4.568693, 50.435785],
-                    [-4.569483, 50.435571],
-                    [-4.56975, 50.435453],
-                    [-4.570019, 50.435523],
-                    [-4.572635, 50.435898],
-                    [-4.572706, 50.435847],
-                    [-4.573277, 50.435708],
-                    [-4.575788, 50.435168],
-                    [-4.576038, 50.435751],
-                    [-4.576334, 50.436947],
-                    [-4.576442, 50.437656],
-                    [-4.57652, 50.438498],
-                    [-4.576455, 50.439267],
-                    [-4.57627, 50.440345],
-                    [-4.576134, 50.441491],
-                    [-4.576121, 50.441718],
-                    [-4.576188, 50.441922],
-                    [-4.575349, 50.441767],
-                    [-4.573112, 50.441127]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400005,
-              "name": "Battle of Braddock Down 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000005",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000005"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000005"
-          }
-        },
-        {
-          "entry-date": "2019-07-17",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400006,
-          "name": "Battle of Chalgrove 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000006",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.061652, 51.680526],
-                    [-1.060801, 51.679481],
-                    [-1.058497, 51.680953],
-                    [-1.056966, 51.679343],
-                    [-1.059627, 51.678461],
-                    [-1.059192, 51.678076],
-                    [-1.059182, 51.678033],
-                    [-1.059223, 51.677941],
-                    [-1.060758, 51.676696],
-                    [-1.061104, 51.676492],
-                    [-1.061028, 51.676428],
-                    [-1.061204, 51.676144],
-                    [-1.061503, 51.675697],
-                    [-1.062235, 51.674791],
-                    [-1.062751, 51.674063],
-                    [-1.06301, 51.673597],
-                    [-1.063922, 51.672632],
-                    [-1.065425, 51.67133],
-                    [-1.067298, 51.66991],
-                    [-1.067307, 51.669814],
-                    [-1.06702, 51.669651],
-                    [-1.067283, 51.66949],
-                    [-1.06771, 51.669485],
-                    [-1.073721, 51.67054],
-                    [-1.073684, 51.670609],
-                    [-1.07367, 51.670825],
-                    [-1.073768, 51.671963],
-                    [-1.074468, 51.678543],
-                    [-1.072124, 51.679513],
-                    [-1.071371, 51.680478],
-                    [-1.066296, 51.680352],
-                    [-1.06559, 51.680868],
-                    [-1.06418, 51.681635],
-                    [-1.063741, 51.681985],
-                    [-1.063401, 51.682188],
-                    [-1.062906, 51.681837],
-                    [-1.062519, 51.681506],
-                    [-1.061652, 51.680526]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2019-07-17",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400006,
-              "name": "Battle of Chalgrove 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000006",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000006"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000006"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400007,
-          "name": "Battle of Cheriton 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000007",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.145457, 51.068138],
-                    [-1.145387, 51.067714],
-                    [-1.144832, 51.067688],
-                    [-1.143705, 51.067815],
-                    [-1.142555, 51.067796],
-                    [-1.142154, 51.067755],
-                    [-1.14153, 51.067631],
-                    [-1.139371, 51.067023],
-                    [-1.137052, 51.066524],
-                    [-1.133069, 51.065999],
-                    [-1.132769, 51.066014],
-                    [-1.130171, 51.066501],
-                    [-1.129561, 51.06654],
-                    [-1.12901, 51.066499],
-                    [-1.128284, 51.066146],
-                    [-1.127165, 51.065508],
-                    [-1.126532, 51.065193],
-                    [-1.126296, 51.064999],
-                    [-1.124701, 51.064231],
-                    [-1.12419, 51.063881],
-                    [-1.124083, 51.063747],
-                    [-1.123969, 51.063429],
-                    [-1.123758, 51.062997],
-                    [-1.123621, 51.062298],
-                    [-1.123641, 51.062142],
-                    [-1.123742, 51.062133],
-                    [-1.124003, 51.061891],
-                    [-1.1244, 51.061268],
-                    [-1.124733, 51.060902],
-                    [-1.12499, 51.060675],
-                    [-1.126553, 51.059743],
-                    [-1.127648, 51.059248],
-                    [-1.127433, 51.058995],
-                    [-1.126575, 51.058212],
-                    [-1.127939, 51.05758],
-                    [-1.128848, 51.05725],
-                    [-1.129873, 51.056766],
-                    [-1.131359, 51.056245],
-                    [-1.132365, 51.055818],
-                    [-1.132423, 51.055834],
-                    [-1.132621, 51.055768],
-                    [-1.133421, 51.055467],
-                    [-1.134712, 51.054913],
-                    [-1.134847, 51.054887],
-                    [-1.13728, 51.05393],
-                    [-1.139284, 51.053407],
-                    [-1.142926, 51.053261],
-                    [-1.148002, 51.053003],
-                    [-1.14954, 51.052872],
-                    [-1.152962, 51.052694],
-                    [-1.154505, 51.052681],
-                    [-1.159168, 51.052861],
-                    [-1.159745, 51.052913],
-                    [-1.160437, 51.053032],
-                    [-1.161192, 51.053215],
-                    [-1.161473, 51.053323],
-                    [-1.161759, 51.053483],
-                    [-1.161786, 51.053552],
-                    [-1.162219, 51.053951],
-                    [-1.162564, 51.054352],
-                    [-1.162953, 51.054933],
-                    [-1.163578, 51.05598],
-                    [-1.163815, 51.056206],
-                    [-1.163947, 51.056267],
-                    [-1.164024, 51.056252],
-                    [-1.164106, 51.05632],
-                    [-1.164654, 51.056416],
-                    [-1.166049, 51.056552],
-                    [-1.167399, 51.05662],
-                    [-1.167788, 51.056683],
-                    [-1.167759, 51.056761],
-                    [-1.167787, 51.056775],
-                    [-1.167798, 51.056873],
-                    [-1.167619, 51.057058],
-                    [-1.16738, 51.057171],
-                    [-1.166972, 51.057185],
-                    [-1.166702, 51.057535],
-                    [-1.166829, 51.058028],
-                    [-1.168498, 51.057889],
-                    [-1.168808, 51.057921],
-                    [-1.169029, 51.058092],
-                    [-1.169456, 51.058643],
-                    [-1.169684, 51.05887],
-                    [-1.169758, 51.058873],
-                    [-1.170433, 51.058656],
-                    [-1.170868, 51.059169],
-                    [-1.171746, 51.06058],
-                    [-1.171889, 51.060896],
-                    [-1.172129, 51.061573],
-                    [-1.172475, 51.062174],
-                    [-1.173143, 51.063614],
-                    [-1.172101, 51.063945],
-                    [-1.166343, 51.066045],
-                    [-1.166196, 51.066048],
-                    [-1.166085, 51.065986],
-                    [-1.165951, 51.066028],
-                    [-1.165366, 51.066328],
-                    [-1.163622, 51.066995],
-                    [-1.161582, 51.067611],
-                    [-1.160325, 51.067915],
-                    [-1.158669, 51.067927],
-                    [-1.157757, 51.068019],
-                    [-1.156692, 51.067995],
-                    [-1.156469, 51.067727],
-                    [-1.15596, 51.067258],
-                    [-1.155814, 51.067225],
-                    [-1.155039, 51.067522],
-                    [-1.153798, 51.067901],
-                    [-1.153715, 51.067899],
-                    [-1.153675, 51.067851],
-                    [-1.153664, 51.067373],
-                    [-1.153566, 51.067118],
-                    [-1.153184, 51.066764],
-                    [-1.152864, 51.066658],
-                    [-1.1523, 51.066822],
-                    [-1.149624, 51.067829],
-                    [-1.148334, 51.068015],
-                    [-1.148238, 51.067993],
-                    [-1.146005, 51.06844],
-                    [-1.145559, 51.068493],
-                    [-1.145457, 51.068138]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400007,
-              "name": "Battle of Cheriton 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000007",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000007"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000007"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400008,
-          "name": "Battle of Cropredy Bridge 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000008",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.307271, 52.103804],
-                    [-1.307516, 52.103531],
-                    [-1.304751, 52.103634],
-                    [-1.303085, 52.103537],
-                    [-1.30299, 52.103658],
-                    [-1.302423, 52.104731],
-                    [-1.301388, 52.104291],
-                    [-1.300782, 52.104799],
-                    [-1.300667, 52.10474],
-                    [-1.300545, 52.104601],
-                    [-1.300357, 52.104493],
-                    [-1.299458, 52.104232],
-                    [-1.299079, 52.104054],
-                    [-1.298327, 52.103503],
-                    [-1.297922, 52.103098],
-                    [-1.297759, 52.103007],
-                    [-1.297856, 52.10282],
-                    [-1.298047, 52.102628],
-                    [-1.299149, 52.101998],
-                    [-1.300902, 52.101109],
-                    [-1.301835, 52.100766],
-                    [-1.303148, 52.100098],
-                    [-1.303746, 52.099876],
-                    [-1.304465, 52.0997],
-                    [-1.305595, 52.099485],
-                    [-1.305997, 52.099325],
-                    [-1.306289, 52.099115],
-                    [-1.307212, 52.098298],
-                    [-1.307891, 52.097587],
-                    [-1.308444, 52.097135],
-                    [-1.309177, 52.097441],
-                    [-1.311818, 52.097464],
-                    [-1.311906, 52.09759],
-                    [-1.316918, 52.097846],
-                    [-1.317583, 52.097897],
-                    [-1.31761, 52.09795],
-                    [-1.317835, 52.09803],
-                    [-1.317403, 52.098289],
-                    [-1.315456, 52.099782],
-                    [-1.314393, 52.100558],
-                    [-1.312966, 52.101533],
-                    [-1.312848, 52.101672],
-                    [-1.312503, 52.101875],
-                    [-1.311083, 52.102862],
-                    [-1.310929, 52.102985],
-                    [-1.310851, 52.103094],
-                    [-1.310766, 52.103306],
-                    [-1.310754, 52.10354],
-                    [-1.30927, 52.103982],
-                    [-1.308767, 52.103811],
-                    [-1.308144, 52.103836],
-                    [-1.308019, 52.105224],
-                    [-1.307799, 52.105213],
-                    [-1.307808, 52.105092],
-                    [-1.307738, 52.104874],
-                    [-1.307094, 52.104416],
-                    [-1.307036, 52.104299],
-                    [-1.307032, 52.104158],
-                    [-1.307271, 52.103804]
-                  ]
-                ],
-                [
-                  [
-                    [-1.292421, 52.124483],
-                    [-1.292186, 52.124436],
-                    [-1.291819, 52.124446],
-                    [-1.291497, 52.124407],
-                    [-1.289292, 52.124059],
-                    [-1.288219, 52.123996],
-                    [-1.28773, 52.124006],
-                    [-1.287635, 52.123967],
-                    [-1.287631, 52.123899],
-                    [-1.287998, 52.123323],
-                    [-1.288192, 52.12288],
-                    [-1.288412, 52.121934],
-                    [-1.28787, 52.12052],
-                    [-1.287828, 52.120199],
-                    [-1.287895, 52.11948],
-                    [-1.287863, 52.119389],
-                    [-1.287431, 52.118841],
-                    [-1.287187, 52.118617],
-                    [-1.289523, 52.116943],
-                    [-1.290633, 52.117312],
-                    [-1.291565, 52.11702],
-                    [-1.292125, 52.116759],
-                    [-1.292626, 52.116323],
-                    [-1.292817, 52.116118],
-                    [-1.293138, 52.11572],
-                    [-1.293444, 52.115217],
-                    [-1.290587, 52.114606],
-                    [-1.290234, 52.114995],
-                    [-1.288086, 52.11427],
-                    [-1.288992, 52.113595],
-                    [-1.286761, 52.112864],
-                    [-1.288837, 52.111044],
-                    [-1.290243, 52.109718],
-                    [-1.290826, 52.109209],
-                    [-1.292523, 52.107918],
-                    [-1.29313, 52.107208],
-                    [-1.293292, 52.106762],
-                    [-1.293376, 52.106592],
-                    [-1.293518, 52.106417],
-                    [-1.293918, 52.106092],
-                    [-1.294287, 52.105868],
-                    [-1.295962, 52.105029],
-                    [-1.297338, 52.105911],
-                    [-1.297444, 52.106007],
-                    [-1.297561, 52.106033],
-                    [-1.297753, 52.106841],
-                    [-1.29785, 52.107571],
-                    [-1.298583, 52.107745],
-                    [-1.298729, 52.107816],
-                    [-1.298689, 52.107937],
-                    [-1.298714, 52.108404],
-                    [-1.298869, 52.10894],
-                    [-1.299302, 52.108886],
-                    [-1.299811, 52.108871],
-                    [-1.301499, 52.109062],
-                    [-1.301796, 52.10917],
-                    [-1.302205, 52.109265],
-                    [-1.303007, 52.109389],
-                    [-1.303108, 52.109469],
-                    [-1.303592, 52.109665],
-                    [-1.303798, 52.109717],
-                    [-1.305834, 52.110108],
-                    [-1.306374, 52.110267],
-                    [-1.307041, 52.110524],
-                    [-1.307362, 52.110683],
-                    [-1.308257, 52.111287],
-                    [-1.309538, 52.111945],
-                    [-1.312487, 52.113694],
-                    [-1.314052, 52.112296],
-                    [-1.314086, 52.112207],
-                    [-1.315991, 52.112218],
-                    [-1.316285, 52.112244],
-                    [-1.316287, 52.112323],
-                    [-1.316152, 52.112659],
-                    [-1.31618, 52.112809],
-                    [-1.316328, 52.113029],
-                    [-1.316304, 52.113158],
-                    [-1.316058, 52.113382],
-                    [-1.315676, 52.113355],
-                    [-1.315514, 52.113455],
-                    [-1.315377, 52.113631],
-                    [-1.315298, 52.113896],
-                    [-1.315074, 52.114205],
-                    [-1.315025, 52.114459],
-                    [-1.314932, 52.114599],
-                    [-1.314698, 52.114745],
-                    [-1.314692, 52.114764],
-                    [-1.314841, 52.114792],
-                    [-1.315032, 52.11489],
-                    [-1.31514, 52.114989],
-                    [-1.315171, 52.115105],
-                    [-1.315155, 52.115303],
-                    [-1.315104, 52.115393],
-                    [-1.314695, 52.115716],
-                    [-1.314739, 52.116098],
-                    [-1.314255, 52.116604],
-                    [-1.313777, 52.117218],
-                    [-1.313662, 52.117333],
-                    [-1.313419, 52.117475],
-                    [-1.313128, 52.117486],
-                    [-1.312894, 52.117424],
-                    [-1.312257, 52.117433],
-                    [-1.311994, 52.1175],
-                    [-1.31116, 52.117503],
-                    [-1.310939, 52.117442],
-                    [-1.31055, 52.117468],
-                    [-1.310384, 52.117436],
-                    [-1.310225, 52.117362],
-                    [-1.309932, 52.117339],
-                    [-1.308035, 52.117259],
-                    [-1.307854, 52.117299],
-                    [-1.306526, 52.118112],
-                    [-1.300859, 52.116127],
-                    [-1.299372, 52.118122],
-                    [-1.298484, 52.118914],
-                    [-1.297105, 52.120436],
-                    [-1.292459, 52.119569],
-                    [-1.291691, 52.120468],
-                    [-1.292126, 52.121158],
-                    [-1.292444, 52.121816],
-                    [-1.293161, 52.122602],
-                    [-1.293327, 52.12295],
-                    [-1.29343, 52.123479],
-                    [-1.293532, 52.123721],
-                    [-1.29377, 52.124095],
-                    [-1.29284, 52.124642],
-                    [-1.292689, 52.124541],
-                    [-1.292421, 52.124483]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400008,
-              "name": "Battle of Cropredy Bridge 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000008",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000008"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000008"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400009,
-          "name": "Battle of Edgehill 1642",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000009",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.481666, 52.153972],
-                    [-1.481642, 52.153528],
-                    [-1.481175, 52.1533],
-                    [-1.480899, 52.153296],
-                    [-1.479693, 52.15294],
-                    [-1.479627, 52.152886],
-                    [-1.479662, 52.152702],
-                    [-1.479836, 52.152404],
-                    [-1.479752, 52.152397],
-                    [-1.47977, 52.152444],
-                    [-1.479692, 52.152467],
-                    [-1.479531, 52.152474],
-                    [-1.479354, 52.152419],
-                    [-1.479227, 52.152217],
-                    [-1.479247, 52.152147],
-                    [-1.479073, 52.15202],
-                    [-1.479061, 52.1519],
-                    [-1.479011, 52.151854],
-                    [-1.478696, 52.151791],
-                    [-1.47827, 52.151847],
-                    [-1.478124, 52.151839],
-                    [-1.478027, 52.151792],
-                    [-1.477946, 52.15169],
-                    [-1.477911, 52.151554],
-                    [-1.477712, 52.151429],
-                    [-1.477611, 52.151406],
-                    [-1.477569, 52.151377],
-                    [-1.477575, 52.151326],
-                    [-1.477321, 52.151342],
-                    [-1.477157, 52.151446],
-                    [-1.476575, 52.15207],
-                    [-1.476314, 52.152484],
-                    [-1.47465, 52.152033],
-                    [-1.473574, 52.151816],
-                    [-1.473877, 52.151484],
-                    [-1.474218, 52.151191],
-                    [-1.472344, 52.150441],
-                    [-1.471547, 52.150084],
-                    [-1.470912, 52.150232],
-                    [-1.470799, 52.150323],
-                    [-1.470642, 52.150368],
-                    [-1.470001, 52.150479],
-                    [-1.46962, 52.150398],
-                    [-1.469148, 52.150453],
-                    [-1.468992, 52.150603],
-                    [-1.468549, 52.15075],
-                    [-1.467887, 52.150895],
-                    [-1.467688, 52.150895],
-                    [-1.467316, 52.1508],
-                    [-1.467078, 52.150823],
-                    [-1.466933, 52.150703],
-                    [-1.466834, 52.150699],
-                    [-1.466635, 52.150733],
-                    [-1.466227, 52.150713],
-                    [-1.4658, 52.150791],
-                    [-1.465337, 52.150956],
-                    [-1.464906, 52.150909],
-                    [-1.464825, 52.150874],
-                    [-1.463792, 52.150784],
-                    [-1.463442, 52.150709],
-                    [-1.462343, 52.150733],
-                    [-1.462216, 52.150695],
-                    [-1.462061, 52.150587],
-                    [-1.461884, 52.150353],
-                    [-1.46145, 52.150257],
-                    [-1.460859, 52.150068],
-                    [-1.460309, 52.150011],
-                    [-1.460224, 52.150053],
-                    [-1.460051, 52.150052],
-                    [-1.45956, 52.149922],
-                    [-1.459309, 52.149906],
-                    [-1.459184, 52.149848],
-                    [-1.458796, 52.149769],
-                    [-1.458537, 52.149795],
-                    [-1.45828, 52.14954],
-                    [-1.457933, 52.149386],
-                    [-1.45786, 52.149296],
-                    [-1.457814, 52.149166],
-                    [-1.457708, 52.149112],
-                    [-1.457229, 52.149014],
-                    [-1.456297, 52.14892],
-                    [-1.456081, 52.148861],
-                    [-1.455639, 52.148906],
-                    [-1.455561, 52.14894],
-                    [-1.455449, 52.149074],
-                    [-1.454916, 52.14905],
-                    [-1.454288, 52.148938],
-                    [-1.453945, 52.148948],
-                    [-1.453509, 52.148911],
-                    [-1.452433, 52.148548],
-                    [-1.45187, 52.148392],
-                    [-1.451603, 52.148282],
-                    [-1.450678, 52.148091],
-                    [-1.44959, 52.147958],
-                    [-1.449236, 52.147667],
-                    [-1.44804, 52.14733],
-                    [-1.447749, 52.147326],
-                    [-1.447555, 52.146399],
-                    [-1.447289, 52.145694],
-                    [-1.446392, 52.144279],
-                    [-1.446258, 52.143919],
-                    [-1.446286, 52.143764],
-                    [-1.446278, 52.143489],
-                    [-1.446242, 52.143353],
-                    [-1.445959, 52.142848],
-                    [-1.445871, 52.142302],
-                    [-1.445493, 52.14183],
-                    [-1.445487, 52.141673],
-                    [-1.445896, 52.140935],
-                    [-1.446362, 52.139944],
-                    [-1.446502, 52.139813],
-                    [-1.44675, 52.13973],
-                    [-1.447864, 52.139617],
-                    [-1.448289, 52.139545],
-                    [-1.448496, 52.139487],
-                    [-1.450099, 52.138825],
-                    [-1.451148, 52.138725],
-                    [-1.451012, 52.138361],
-                    [-1.450831, 52.138073],
-                    [-1.450973, 52.138019],
-                    [-1.452057, 52.137204],
-                    [-1.454474, 52.135176],
-                    [-1.455823, 52.135649],
-                    [-1.456257, 52.135657],
-                    [-1.457642, 52.134442],
-                    [-1.457701, 52.134412],
-                    [-1.458171, 52.134378],
-                    [-1.457967, 52.134028],
-                    [-1.45794, 52.133783],
-                    [-1.459304, 52.133108],
-                    [-1.459537, 52.132916],
-                    [-1.459924, 52.132707],
-                    [-1.460338, 52.132972],
-                    [-1.460821, 52.132737],
-                    [-1.460836, 52.132681],
-                    [-1.461162, 52.132668],
-                    [-1.461511, 52.132429],
-                    [-1.460569, 52.131886],
-                    [-1.460653, 52.130778],
-                    [-1.46045, 52.13065],
-                    [-1.460474, 52.130599],
-                    [-1.460975, 52.130308],
-                    [-1.461096, 52.130267],
-                    [-1.461564, 52.130196],
-                    [-1.462484, 52.129957],
-                    [-1.462807, 52.129719],
-                    [-1.463329, 52.129789],
-                    [-1.463315, 52.129913],
-                    [-1.463687, 52.129928],
-                    [-1.463797, 52.129862],
-                    [-1.464086, 52.129913],
-                    [-1.464575, 52.129359],
-                    [-1.465864, 52.127754],
-                    [-1.46594, 52.127515],
-                    [-1.465879, 52.127129],
-                    [-1.466002, 52.126837],
-                    [-1.466253, 52.126552],
-                    [-1.466671, 52.126229],
-                    [-1.467122, 52.12592],
-                    [-1.467677, 52.125671],
-                    [-1.466994, 52.124799],
-                    [-1.466878, 52.124622],
-                    [-1.466887, 52.124536],
-                    [-1.46711, 52.124447],
-                    [-1.467749, 52.123978],
-                    [-1.469362, 52.122969],
-                    [-1.469521, 52.122732],
-                    [-1.470367, 52.122522],
-                    [-1.470706, 52.122603],
-                    [-1.471219, 52.122796],
-                    [-1.471509, 52.122954],
-                    [-1.472128, 52.123197],
-                    [-1.472966, 52.123278],
-                    [-1.473361, 52.12324],
-                    [-1.474319, 52.123338],
-                    [-1.474616, 52.123351],
-                    [-1.474781, 52.123324],
-                    [-1.475632, 52.123867],
-                    [-1.475796, 52.124022],
-                    [-1.476418, 52.124025],
-                    [-1.477362, 52.124359],
-                    [-1.477537, 52.124384],
-                    [-1.478018, 52.124316],
-                    [-1.478879, 52.12432],
-                    [-1.47908, 52.1244],
-                    [-1.479318, 52.124339],
-                    [-1.480432, 52.124855],
-                    [-1.481282, 52.124899],
-                    [-1.482416, 52.124851],
-                    [-1.482656, 52.124809],
-                    [-1.482814, 52.124999],
-                    [-1.483071, 52.124987],
-                    [-1.483092, 52.12506],
-                    [-1.483159, 52.125109],
-                    [-1.483407, 52.12514],
-                    [-1.483532, 52.125218],
-                    [-1.483691, 52.125201],
-                    [-1.483889, 52.125257],
-                    [-1.48403, 52.125216],
-                    [-1.484066, 52.125235],
-                    [-1.484047, 52.125316],
-                    [-1.484075, 52.125346],
-                    [-1.484251, 52.125333],
-                    [-1.484405, 52.125239],
-                    [-1.484541, 52.125268],
-                    [-1.4846, 52.125367],
-                    [-1.48471, 52.125455],
-                    [-1.484772, 52.125571],
-                    [-1.484796, 52.125724],
-                    [-1.484711, 52.125802],
-                    [-1.484974, 52.125851],
-                    [-1.485122, 52.125826],
-                    [-1.4852, 52.12588],
-                    [-1.485516, 52.125929],
-                    [-1.48974, 52.129228],
-                    [-1.489896, 52.129306],
-                    [-1.490148, 52.129368],
-                    [-1.490562, 52.129396],
-                    [-1.490717, 52.129314],
-                    [-1.490808, 52.129321],
-                    [-1.491501, 52.129547],
-                    [-1.491976, 52.129588],
-                    [-1.492257, 52.129662],
-                    [-1.492551, 52.129571],
-                    [-1.492863, 52.129578],
-                    [-1.493136, 52.129517],
-                    [-1.493251, 52.129538],
-                    [-1.493277, 52.129574],
-                    [-1.493888, 52.129613],
-                    [-1.494394, 52.129478],
-                    [-1.494397, 52.129572],
-                    [-1.494491, 52.129654],
-                    [-1.494366, 52.129734],
-                    [-1.494503, 52.129856],
-                    [-1.494629, 52.129893],
-                    [-1.494799, 52.129805],
-                    [-1.494979, 52.129774],
-                    [-1.495018, 52.129818],
-                    [-1.495019, 52.129901],
-                    [-1.495059, 52.129933],
-                    [-1.495303, 52.12995],
-                    [-1.495525, 52.1301],
-                    [-1.495569, 52.130214],
-                    [-1.495737, 52.1303],
-                    [-1.495681, 52.130397],
-                    [-1.495734, 52.130462],
-                    [-1.495628, 52.130502],
-                    [-1.495598, 52.130576],
-                    [-1.495603, 52.130648],
-                    [-1.495732, 52.130766],
-                    [-1.495646, 52.130841],
-                    [-1.495719, 52.130996],
-                    [-1.495811, 52.131025],
-                    [-1.495894, 52.130937],
-                    [-1.495958, 52.13093],
-                    [-1.496009, 52.130975],
-                    [-1.496005, 52.131031],
-                    [-1.496174, 52.131048],
-                    [-1.496253, 52.131091],
-                    [-1.496435, 52.131292],
-                    [-1.49643, 52.131348],
-                    [-1.496332, 52.131424],
-                    [-1.496361, 52.131479],
-                    [-1.496512, 52.131477],
-                    [-1.496755, 52.131539],
-                    [-1.496889, 52.131637],
-                    [-1.497069, 52.131635],
-                    [-1.497263, 52.131769],
-                    [-1.497273, 52.131794],
-                    [-1.497199, 52.131897],
-                    [-1.497097, 52.131975],
-                    [-1.496993, 52.13233],
-                    [-1.497042, 52.132351],
-                    [-1.497136, 52.132321],
-                    [-1.497207, 52.132388],
-                    [-1.497656, 52.132451],
-                    [-1.497715, 52.132505],
-                    [-1.497772, 52.132666],
-                    [-1.497929, 52.132913],
-                    [-1.498285, 52.133166],
-                    [-1.49838, 52.133187],
-                    [-1.498452, 52.133128],
-                    [-1.4986, 52.13311],
-                    [-1.498872, 52.13319],
-                    [-1.498946, 52.133113],
-                    [-1.49903, 52.133081],
-                    [-1.49919, 52.133143],
-                    [-1.499412, 52.133127],
-                    [-1.499476, 52.133157],
-                    [-1.499465, 52.133238],
-                    [-1.49958, 52.133262],
-                    [-1.499474, 52.133323],
-                    [-1.499478, 52.133379],
-                    [-1.499721, 52.133578],
-                    [-1.4998, 52.133539],
-                    [-1.500076, 52.133536],
-                    [-1.500137, 52.133434],
-                    [-1.500111, 52.133358],
-                    [-1.500261, 52.13335],
-                    [-1.500485, 52.133437],
-                    [-1.500544, 52.133557],
-                    [-1.500479, 52.133648],
-                    [-1.500496, 52.133678],
-                    [-1.500711, 52.133633],
-                    [-1.500769, 52.133749],
-                    [-1.500861, 52.133769],
-                    [-1.500911, 52.13373],
-                    [-1.501112, 52.133777],
-                    [-1.50108, 52.133845],
-                    [-1.501105, 52.133882],
-                    [-1.50122, 52.133803],
-                    [-1.501367, 52.133879],
-                    [-1.501614, 52.133911],
-                    [-1.501662, 52.133912],
-                    [-1.501635, 52.133859],
-                    [-1.501691, 52.133844],
-                    [-1.501847, 52.133921],
-                    [-1.501996, 52.133852],
-                    [-1.502095, 52.133867],
-                    [-1.502233, 52.134064],
-                    [-1.502701, 52.13422],
-                    [-1.502764, 52.134365],
-                    [-1.502711, 52.134473],
-                    [-1.502764, 52.134534],
-                    [-1.502808, 52.134558],
-                    [-1.503008, 52.134531],
-                    [-1.503106, 52.134558],
-                    [-1.503128, 52.134621],
-                    [-1.503192, 52.134671],
-                    [-1.5032, 52.134766],
-                    [-1.503154, 52.134852],
-                    [-1.503293, 52.134939],
-                    [-1.503249, 52.135034],
-                    [-1.503262, 52.135089],
-                    [-1.503362, 52.13512],
-                    [-1.503617, 52.135122],
-                    [-1.503642, 52.135174],
-                    [-1.503585, 52.135218],
-                    [-1.503592, 52.135253],
-                    [-1.503727, 52.135283],
-                    [-1.503816, 52.135427],
-                    [-1.503905, 52.135466],
-                    [-1.504013, 52.135464],
-                    [-1.504114, 52.135293],
-                    [-1.504363, 52.135319],
-                    [-1.504541, 52.135496],
-                    [-1.504405, 52.135643],
-                    [-1.504415, 52.135682],
-                    [-1.504746, 52.135793],
-                    [-1.504853, 52.135732],
-                    [-1.505127, 52.13573],
-                    [-1.505232, 52.135681],
-                    [-1.505603, 52.135772],
-                    [-1.505658, 52.135745],
-                    [-1.505688, 52.135664],
-                    [-1.506135, 52.135642],
-                    [-1.506214, 52.135459],
-                    [-1.506638, 52.135462],
-                    [-1.507097, 52.13533],
-                    [-1.507401, 52.135337],
-                    [-1.508761, 52.137363],
-                    [-1.509827, 52.138738],
-                    [-1.509931, 52.138835],
-                    [-1.510004, 52.138862],
-                    [-1.51046, 52.13875],
-                    [-1.510842, 52.139164],
-                    [-1.511002, 52.139864],
-                    [-1.511078, 52.140446],
-                    [-1.511071, 52.140757],
-                    [-1.510694, 52.142222],
-                    [-1.510547, 52.14268],
-                    [-1.510371, 52.142659],
-                    [-1.509856, 52.143387],
-                    [-1.509355, 52.143959],
-                    [-1.508973, 52.144497],
-                    [-1.507963, 52.145564],
-                    [-1.508604, 52.146017],
-                    [-1.509804, 52.146683],
-                    [-1.511736, 52.14819],
-                    [-1.511967, 52.148339],
-                    [-1.512938, 52.14945],
-                    [-1.512716, 52.149573],
-                    [-1.512164, 52.149992],
-                    [-1.51218, 52.150969],
-                    [-1.512302, 52.151518],
-                    [-1.512699, 52.151955],
-                    [-1.512961, 52.152096],
-                    [-1.512925, 52.152122],
-                    [-1.51301, 52.152155],
-                    [-1.512985, 52.15219],
-                    [-1.513035, 52.152251],
-                    [-1.512911, 52.152518],
-                    [-1.512546, 52.153733],
-                    [-1.512255, 52.154175],
-                    [-1.512064, 52.154375],
-                    [-1.512003, 52.154331],
-                    [-1.511603, 52.154289],
-                    [-1.511065, 52.154143],
-                    [-1.510083, 52.154162],
-                    [-1.509865, 52.154202],
-                    [-1.50944, 52.154323],
-                    [-1.508947, 52.154531],
-                    [-1.508602, 52.154711],
-                    [-1.508158, 52.154884],
-                    [-1.507751, 52.155109],
-                    [-1.50703, 52.155294],
-                    [-1.506642, 52.155529],
-                    [-1.506552, 52.155553],
-                    [-1.505675, 52.155297],
-                    [-1.504967, 52.155026],
-                    [-1.503888, 52.154479],
-                    [-1.502949, 52.153817],
-                    [-1.502585, 52.153707],
-                    [-1.502247, 52.153705],
-                    [-1.50205, 52.153793],
-                    [-1.501913, 52.153936],
-                    [-1.501942, 52.153994],
-                    [-1.501345, 52.154075],
-                    [-1.499509, 52.154109],
-                    [-1.499225, 52.154022],
-                    [-1.499051, 52.154065],
-                    [-1.498944, 52.154038],
-                    [-1.498789, 52.1538],
-                    [-1.498608, 52.153766],
-                    [-1.498389, 52.153798],
-                    [-1.498321, 52.153877],
-                    [-1.498136, 52.153944],
-                    [-1.497804, 52.154002],
-                    [-1.497267, 52.153975],
-                    [-1.49707, 52.153909],
-                    [-1.496365, 52.153977],
-                    [-1.495842, 52.153958],
-                    [-1.493147, 52.15358],
-                    [-1.492438, 52.153362],
-                    [-1.492088, 52.15339],
-                    [-1.491927, 52.153613],
-                    [-1.491112, 52.153734],
-                    [-1.489162, 52.153839],
-                    [-1.488152, 52.153939],
-                    [-1.486649, 52.154012],
-                    [-1.486518, 52.154067],
-                    [-1.486239, 52.154333],
-                    [-1.486, 52.154467],
-                    [-1.486062, 52.15455],
-                    [-1.485833, 52.154746],
-                    [-1.484996, 52.155143],
-                    [-1.484375, 52.155496],
-                    [-1.484142, 52.155379],
-                    [-1.483733, 52.15511],
-                    [-1.483339, 52.154801],
-                    [-1.482737, 52.154431],
-                    [-1.481668, 52.153997],
-                    [-1.481666, 52.153972]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400009,
-              "name": "Battle of Edgehill 1642",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000009",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000009"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000009"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400010,
-          "name": "Battle of Evesham 1265",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000010",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.94084, 52.110768],
-                    [-1.938725, 52.110255],
-                    [-1.932315, 52.110444],
-                    [-1.932667, 52.109766],
-                    [-1.933047, 52.108866],
-                    [-1.933101, 52.108482],
-                    [-1.93498, 52.104346],
-                    [-1.935093, 52.104315],
-                    [-1.935517, 52.103394],
-                    [-1.935921, 52.102662],
-                    [-1.937582, 52.099212],
-                    [-1.937813, 52.09883],
-                    [-1.938181, 52.098462],
-                    [-1.938805, 52.098027],
-                    [-1.940236, 52.098298],
-                    [-1.940297, 52.09823],
-                    [-1.940403, 52.098224],
-                    [-1.940415, 52.098114],
-                    [-1.941627, 52.098115],
-                    [-1.941546, 52.098443],
-                    [-1.941358, 52.098425],
-                    [-1.94129, 52.098734],
-                    [-1.941329, 52.098742],
-                    [-1.941315, 52.0988],
-                    [-1.940798, 52.098759],
-                    [-1.940771, 52.098883],
-                    [-1.941188, 52.098923],
-                    [-1.941151, 52.09906],
-                    [-1.941358, 52.099074],
-                    [-1.940949, 52.100878],
-                    [-1.941189, 52.100923],
-                    [-1.941207, 52.101361],
-                    [-1.941237, 52.101394],
-                    [-1.9417, 52.101363],
-                    [-1.941751, 52.101683],
-                    [-1.940983, 52.101738],
-                    [-1.940917, 52.102467],
-                    [-1.941189, 52.102472],
-                    [-1.941088, 52.103421],
-                    [-1.941088, 52.103889],
-                    [-1.940496, 52.103984],
-                    [-1.940456, 52.104014],
-                    [-1.940429, 52.104124],
-                    [-1.940492, 52.104717],
-                    [-1.940504, 52.10547],
-                    [-1.940635, 52.106029],
-                    [-1.940824, 52.106504],
-                    [-1.94043, 52.10655],
-                    [-1.940193, 52.107746],
-                    [-1.940208, 52.107911],
-                    [-1.940124, 52.108727],
-                    [-1.942788, 52.108737],
-                    [-1.942969, 52.108692],
-                    [-1.944211, 52.108649],
-                    [-1.944062, 52.107728],
-                    [-1.945157, 52.107436],
-                    [-1.944939, 52.105906],
-                    [-1.944886, 52.105379],
-                    [-1.944944, 52.105359],
-                    [-1.944925, 52.105197],
-                    [-1.944872, 52.104967],
-                    [-1.944599, 52.104986],
-                    [-1.94463, 52.104536],
-                    [-1.945025, 52.104561],
-                    [-1.945068, 52.104038],
-                    [-1.945153, 52.103832],
-                    [-1.944805, 52.102918],
-                    [-1.945165, 52.102857],
-                    [-1.945313, 52.102047],
-                    [-1.946464, 52.09949],
-                    [-1.946703, 52.099301],
-                    [-1.947129, 52.099302],
-                    [-1.947734, 52.09967],
-                    [-1.948247, 52.099415],
-                    [-1.949269, 52.100147],
-                    [-1.949539, 52.10041],
-                    [-1.948582, 52.100652],
-                    [-1.949677, 52.104143],
-                    [-1.952053, 52.103792],
-                    [-1.952465, 52.104453],
-                    [-1.952871, 52.104946],
-                    [-1.953343, 52.105437],
-                    [-1.953593, 52.105691],
-                    [-1.955276, 52.107156],
-                    [-1.955885, 52.107886],
-                    [-1.956668, 52.108592],
-                    [-1.955963, 52.108768],
-                    [-1.95471, 52.109295],
-                    [-1.953889, 52.109962],
-                    [-1.953469, 52.110181],
-                    [-1.952907, 52.110033],
-                    [-1.94965, 52.112318],
-                    [-1.9467, 52.11097],
-                    [-1.943193, 52.110887],
-                    [-1.94084, 52.110768]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400010,
-              "name": "Battle of Evesham 1265",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000010",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000010"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000010"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400011,
-          "name": "Battle of Flodden 1513",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000011",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.150358, 55.632246],
-                    [-2.149553, 55.630652],
-                    [-2.150348, 55.629381],
-                    [-2.150008, 55.627519],
-                    [-2.153353, 55.625843],
-                    [-2.154027, 55.625724],
-                    [-2.154384, 55.6254],
-                    [-2.160299, 55.621004],
-                    [-2.164024, 55.620505],
-                    [-2.168949, 55.61951],
-                    [-2.178082, 55.617736],
-                    [-2.178262, 55.618077],
-                    [-2.178369, 55.618362],
-                    [-2.178785, 55.61892],
-                    [-2.178781, 55.619038],
-                    [-2.179134, 55.619171],
-                    [-2.179364, 55.619499],
-                    [-2.179479, 55.619533],
-                    [-2.179503, 55.619573],
-                    [-2.179484, 55.619649],
-                    [-2.179767, 55.619775],
-                    [-2.179922, 55.619769],
-                    [-2.180068, 55.619817],
-                    [-2.180117, 55.619869],
-                    [-2.180244, 55.619915],
-                    [-2.180299, 55.620064],
-                    [-2.180398, 55.62017],
-                    [-2.180684, 55.620375],
-                    [-2.180762, 55.620477],
-                    [-2.180936, 55.620542],
-                    [-2.180983, 55.620606],
-                    [-2.181139, 55.620691],
-                    [-2.181317, 55.620858],
-                    [-2.18159, 55.621239],
-                    [-2.181791, 55.62138],
-                    [-2.181829, 55.621482],
-                    [-2.181923, 55.62149],
-                    [-2.181956, 55.621566],
-                    [-2.182244, 55.621765],
-                    [-2.182452, 55.621974],
-                    [-2.182718, 55.622109],
-                    [-2.182878, 55.622129],
-                    [-2.18301, 55.622191],
-                    [-2.183998, 55.62286],
-                    [-2.185152, 55.623492],
-                    [-2.185324, 55.623629],
-                    [-2.186122, 55.624099],
-                    [-2.18649, 55.624269],
-                    [-2.186487, 55.624301],
-                    [-2.186576, 55.62437],
-                    [-2.187864, 55.625],
-                    [-2.18807, 55.625137],
-                    [-2.187498, 55.62551],
-                    [-2.18723, 55.625645],
-                    [-2.183033, 55.627237],
-                    [-2.182471, 55.627309],
-                    [-2.18241, 55.627275],
-                    [-2.181891, 55.627431],
-                    [-2.181187, 55.627704],
-                    [-2.180746, 55.627944],
-                    [-2.180138, 55.628629],
-                    [-2.180103, 55.628838],
-                    [-2.179939, 55.628986],
-                    [-2.179752, 55.628992],
-                    [-2.179661, 55.629028],
-                    [-2.179368, 55.628994],
-                    [-2.179058, 55.629162],
-                    [-2.178965, 55.629273],
-                    [-2.179013, 55.629396],
-                    [-2.179095, 55.62942],
-                    [-2.178865, 55.629497],
-                    [-2.178871, 55.629591],
-                    [-2.178764, 55.629651],
-                    [-2.178273, 55.629693],
-                    [-2.178221, 55.629782],
-                    [-2.178078, 55.629794],
-                    [-2.178048, 55.629849],
-                    [-2.177895, 55.629936],
-                    [-2.17789, 55.629983],
-                    [-2.177948, 55.63006],
-                    [-2.177924, 55.63017],
-                    [-2.177842, 55.630233],
-                    [-2.177487, 55.630208],
-                    [-2.177303, 55.630288],
-                    [-2.177051, 55.630285],
-                    [-2.176813, 55.630391],
-                    [-2.176607, 55.630441],
-                    [-2.176163, 55.630644],
-                    [-2.176131, 55.63069],
-                    [-2.176022, 55.630685],
-                    [-2.175911, 55.630732],
-                    [-2.175748, 55.630856],
-                    [-2.175657, 55.630876],
-                    [-2.175476, 55.630984],
-                    [-2.175304, 55.630994],
-                    [-2.17489, 55.63124],
-                    [-2.174877, 55.631436],
-                    [-2.174828, 55.631514],
-                    [-2.174584, 55.631744],
-                    [-2.174368, 55.631863],
-                    [-2.170581, 55.632912],
-                    [-2.165022, 55.633753],
-                    [-2.164407, 55.634038],
-                    [-2.161989, 55.634744],
-                    [-2.160913, 55.634962],
-                    [-2.159484, 55.635161],
-                    [-2.159057, 55.635327],
-                    [-2.158906, 55.635432],
-                    [-2.158695, 55.635609],
-                    [-2.15851, 55.63583],
-                    [-2.158012, 55.63656],
-                    [-2.157678, 55.63666],
-                    [-2.155368, 55.636887],
-                    [-2.150729, 55.632998],
-                    [-2.150358, 55.632246]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400011,
-              "name": "Battle of Flodden 1513",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000011",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000011"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000011"
-          }
-        },
-        {
-          "entry-date": "2016-10-18",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400012,
-          "name": "The Site of The Battle of Halidon Hill 1333",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000012",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.052897, 55.802639],
-                    [-2.052611, 55.802523],
-                    [-2.051538, 55.800312],
-                    [-2.051044, 55.79942],
-                    [-2.050858, 55.798623],
-                    [-2.050821, 55.798585],
-                    [-2.049423, 55.798446],
-                    [-2.048734, 55.798283],
-                    [-2.047719, 55.798161],
-                    [-2.047479, 55.798144],
-                    [-2.046322, 55.798224],
-                    [-2.045927, 55.798209],
-                    [-2.045947, 55.797728],
-                    [-2.046302, 55.797449],
-                    [-2.046518, 55.797146],
-                    [-2.045643, 55.796872],
-                    [-2.044447, 55.796668],
-                    [-2.041991, 55.796022],
-                    [-2.041548, 55.796019],
-                    [-2.040961, 55.796079],
-                    [-2.039027, 55.794423],
-                    [-2.039696, 55.794206],
-                    [-2.039324, 55.793073],
-                    [-2.039203, 55.792492],
-                    [-2.035552, 55.791008],
-                    [-2.039821, 55.787656],
-                    [-2.0405, 55.787799],
-                    [-2.041849, 55.78529],
-                    [-2.044421, 55.78569],
-                    [-2.045964, 55.782887],
-                    [-2.048866, 55.783286],
-                    [-2.058832, 55.784361],
-                    [-2.060131, 55.784545],
-                    [-2.061163, 55.784659],
-                    [-2.061745, 55.784684],
-                    [-2.06278, 55.784823],
-                    [-2.063042, 55.784872],
-                    [-2.063703, 55.785111],
-                    [-2.066, 55.786097],
-                    [-2.068408, 55.786902],
-                    [-2.067507, 55.787726],
-                    [-2.067109, 55.7883],
-                    [-2.066359, 55.789118],
-                    [-2.068381, 55.791663],
-                    [-2.068957, 55.793525],
-                    [-2.068912, 55.793533],
-                    [-2.068994, 55.79385],
-                    [-2.069048, 55.793853],
-                    [-2.06909, 55.793996],
-                    [-2.069085, 55.794105],
-                    [-2.069068, 55.794158],
-                    [-2.066928, 55.795785],
-                    [-2.068768, 55.796764],
-                    [-2.069246, 55.797055],
-                    [-2.069994, 55.797712],
-                    [-2.070028, 55.797711],
-                    [-2.067087, 55.798517],
-                    [-2.067175, 55.798982],
-                    [-2.067144, 55.80025],
-                    [-2.066656, 55.80067],
-                    [-2.06547, 55.801514],
-                    [-2.064941, 55.801838],
-                    [-2.063324, 55.802406],
-                    [-2.062233, 55.802735],
-                    [-2.061121, 55.802714],
-                    [-2.06055, 55.802664],
-                    [-2.060033, 55.802579],
-                    [-2.055929, 55.802926],
-                    [-2.054962, 55.802787],
-                    [-2.052897, 55.802639]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2016-10-18",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400012,
-              "name": "The Site of The Battle of Halidon Hill 1333",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000012",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000012"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000012"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400013,
-          "name": "Battle of Hastings 1066",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000013",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [0.47895, 50.917092],
-                    [0.479352, 50.917062],
-                    [0.479427, 50.917021],
-                    [0.481212, 50.916879],
-                    [0.48154, 50.916875],
-                    [0.481867, 50.916928],
-                    [0.481929, 50.916909],
-                    [0.48188, 50.916844],
-                    [0.482494, 50.916535],
-                    [0.482515, 50.916551],
-                    [0.482696, 50.916437],
-                    [0.483236, 50.915934],
-                    [0.483695, 50.915134],
-                    [0.483698, 50.91508],
-                    [0.484797, 50.914994],
-                    [0.484792, 50.914944],
-                    [0.485331, 50.914923],
-                    [0.485658, 50.915],
-                    [0.485678, 50.914951],
-                    [0.48631, 50.915051],
-                    [0.486308, 50.915073],
-                    [0.48674, 50.915042],
-                    [0.487782, 50.914846],
-                    [0.489716, 50.914398],
-                    [0.49042, 50.914291],
-                    [0.490299, 50.914147],
-                    [0.490353, 50.913923],
-                    [0.490288, 50.913725],
-                    [0.490606, 50.913161],
-                    [0.490653, 50.913106],
-                    [0.490745, 50.913122],
-                    [0.490936, 50.912948],
-                    [0.491145, 50.912859],
-                    [0.491636, 50.912799],
-                    [0.491767, 50.912832],
-                    [0.491844, 50.912641],
-                    [0.491834, 50.912423],
-                    [0.491796, 50.912307],
-                    [0.491827, 50.912155],
-                    [0.491872, 50.912086],
-                    [0.492244, 50.911646],
-                    [0.492491, 50.91144],
-                    [0.493007, 50.911088],
-                    [0.492546, 50.910619],
-                    [0.492176, 50.910328],
-                    [0.491691, 50.909772],
-                    [0.491315, 50.909219],
-                    [0.490997, 50.90844],
-                    [0.490262, 50.907869],
-                    [0.490198, 50.90775],
-                    [0.49018, 50.907578],
-                    [0.489834, 50.907658],
-                    [0.489253, 50.907567],
-                    [0.488996, 50.907497],
-                    [0.488572, 50.907307],
-                    [0.487869, 50.907092],
-                    [0.486917, 50.907007],
-                    [0.486726, 50.906938],
-                    [0.486263, 50.906928],
-                    [0.486085, 50.906985],
-                    [0.486082, 50.907069],
-                    [0.485701, 50.907215],
-                    [0.485722, 50.907234],
-                    [0.485069, 50.907521],
-                    [0.483458, 50.908041],
-                    [0.481515, 50.908777],
-                    [0.481454, 50.908856],
-                    [0.481736, 50.909691],
-                    [0.48207, 50.910503],
-                    [0.481619, 50.910615],
-                    [0.479673, 50.910851],
-                    [0.477557, 50.911011],
-                    [0.477498, 50.911265],
-                    [0.477476, 50.911905],
-                    [0.477531, 50.912733],
-                    [0.477355, 50.913968],
-                    [0.477359, 50.914442],
-                    [0.477433, 50.915045],
-                    [0.477429, 50.915341],
-                    [0.477465, 50.915455],
-                    [0.477435, 50.915502],
-                    [0.477478, 50.916126],
-                    [0.47756, 50.916353],
-                    [0.477785, 50.916771],
-                    [0.477875, 50.916888],
-                    [0.478031, 50.916997],
-                    [0.478197, 50.917074],
-                    [0.478406, 50.91711],
-                    [0.47895, 50.917092]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400013,
-              "name": "Battle of Hastings 1066",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000013",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000013"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000013"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400014,
-          "name": "Battle of Homildon Hill 1402",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000014",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.048654, 55.560657],
-                    [-2.047137, 55.560461],
-                    [-2.044119, 55.560166],
-                    [-2.040561, 55.559758],
-                    [-2.035555, 55.559128],
-                    [-2.035644, 55.558089],
-                    [-2.035019, 55.557575],
-                    [-2.035634, 55.557578],
-                    [-2.03617, 55.557473],
-                    [-2.036912, 55.557411],
-                    [-2.03803, 55.557472],
-                    [-2.037074, 55.553839],
-                    [-2.037443, 55.553864],
-                    [-2.038096, 55.554002],
-                    [-2.038613, 55.553973],
-                    [-2.038884, 55.553925],
-                    [-2.039203, 55.553818],
-                    [-2.039329, 55.553735],
-                    [-2.039733, 55.553336],
-                    [-2.039749, 55.552874],
-                    [-2.039813, 55.552574],
-                    [-2.040074, 55.551991],
-                    [-2.040069, 55.55187],
-                    [-2.040021, 55.551762],
-                    [-2.039705, 55.551334],
-                    [-2.03918, 55.551086],
-                    [-2.039306, 55.550938],
-                    [-2.039544, 55.550538],
-                    [-2.03959, 55.550238],
-                    [-2.03976, 55.549987],
-                    [-2.039682, 55.549921],
-                    [-2.040567, 55.549511],
-                    [-2.040785, 55.549462],
-                    [-2.040731, 55.549394],
-                    [-2.041145, 55.549299],
-                    [-2.041392, 55.549258],
-                    [-2.041429, 55.549279],
-                    [-2.041756, 55.549131],
-                    [-2.041809, 55.549169],
-                    [-2.042065, 55.548963],
-                    [-2.042145, 55.548844],
-                    [-2.042213, 55.548832],
-                    [-2.042283, 55.54877],
-                    [-2.043001, 55.547747],
-                    [-2.043341, 55.546981],
-                    [-2.043113, 55.546424],
-                    [-2.0429, 55.54607],
-                    [-2.042823, 55.545778],
-                    [-2.043116, 55.545583],
-                    [-2.044506, 55.5457],
-                    [-2.045788, 55.546101],
-                    [-2.046732, 55.546324],
-                    [-2.047216, 55.546403],
-                    [-2.047727, 55.546587],
-                    [-2.048636, 55.546763],
-                    [-2.048777, 55.546861],
-                    [-2.049124, 55.546962],
-                    [-2.049743, 55.547055],
-                    [-2.050157, 55.547047],
-                    [-2.051354, 55.547369],
-                    [-2.053046, 55.547742],
-                    [-2.053834, 55.548065],
-                    [-2.054048, 55.548073],
-                    [-2.054846, 55.547861],
-                    [-2.055363, 55.547783],
-                    [-2.056403, 55.547959],
-                    [-2.057123, 55.547887],
-                    [-2.059623, 55.548099],
-                    [-2.059896, 55.548168],
-                    [-2.060393, 55.548449],
-                    [-2.064364, 55.549873],
-                    [-2.065824, 55.550168],
-                    [-2.065808, 55.550606],
-                    [-2.067769, 55.550835],
-                    [-2.070737, 55.550968],
-                    [-2.069775, 55.555334],
-                    [-2.07051, 55.555938],
-                    [-2.071293, 55.55627],
-                    [-2.070749, 55.556615],
-                    [-2.07064, 55.556722],
-                    [-2.070305, 55.557246],
-                    [-2.070084, 55.557404],
-                    [-2.068985, 55.557713],
-                    [-2.068734, 55.557827],
-                    [-2.068675, 55.557901],
-                    [-2.068537, 55.558257],
-                    [-2.068476, 55.558307],
-                    [-2.068403, 55.558745],
-                    [-2.068321, 55.558745],
-                    [-2.068062, 55.558854],
-                    [-2.067631, 55.558852],
-                    [-2.067581, 55.558872],
-                    [-2.06747, 55.559067],
-                    [-2.067287, 55.559137],
-                    [-2.067233, 55.559481],
-                    [-2.067256, 55.55949],
-                    [-2.067149, 55.560636],
-                    [-2.066046, 55.560538],
-                    [-2.06597, 55.562816],
-                    [-2.06212, 55.562262],
-                    [-2.058207, 55.561757],
-                    [-2.056616, 55.561596],
-                    [-2.048654, 55.560657]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400014,
-              "name": "Battle of Homildon Hill 1402",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000014",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000014"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000014"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400015,
-          "name": "Battle of Hopton Heath 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000015",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.06992, 52.838978],
-                    [-2.069929, 52.838923],
-                    [-2.067906, 52.839034],
-                    [-2.067572, 52.83844],
-                    [-2.066884, 52.837654],
-                    [-2.066829, 52.837552],
-                    [-2.065033, 52.837752],
-                    [-2.065014, 52.837709],
-                    [-2.064369, 52.837783],
-                    [-2.064017, 52.837782],
-                    [-2.06344, 52.837715],
-                    [-2.062902, 52.837579],
-                    [-2.062395, 52.837491],
-                    [-2.061719, 52.837522],
-                    [-2.060478, 52.837699],
-                    [-2.060158, 52.837714],
-                    [-2.059831, 52.837705],
-                    [-2.05899, 52.837491],
-                    [-2.059496, 52.836956],
-                    [-2.059864, 52.836365],
-                    [-2.05997, 52.836085],
-                    [-2.060059, 52.835441],
-                    [-2.05999, 52.833774],
-                    [-2.060016, 52.833609],
-                    [-2.06016, 52.833239],
-                    [-2.060711, 52.83232],
-                    [-2.061541, 52.831287],
-                    [-2.062139, 52.830603],
-                    [-2.062678, 52.830194],
-                    [-2.063271, 52.829863],
-                    [-2.063597, 52.829807],
-                    [-2.064543, 52.830058],
-                    [-2.064893, 52.830179],
-                    [-2.070493, 52.832467],
-                    [-2.072755, 52.833037],
-                    [-2.073906, 52.833245],
-                    [-2.074234, 52.833341],
-                    [-2.075205, 52.833883],
-                    [-2.07773, 52.835405],
-                    [-2.075727, 52.83599],
-                    [-2.075771, 52.836129],
-                    [-2.075907, 52.836107],
-                    [-2.075936, 52.836146],
-                    [-2.076468, 52.836337],
-                    [-2.076105, 52.836814],
-                    [-2.076071, 52.836806],
-                    [-2.075954, 52.836921],
-                    [-2.073696, 52.837446],
-                    [-2.074535, 52.837881],
-                    [-2.074777, 52.838089],
-                    [-2.074832, 52.838177],
-                    [-2.074888, 52.838414],
-                    [-2.074829, 52.838681],
-                    [-2.074684, 52.838888],
-                    [-2.074544, 52.838988],
-                    [-2.074311, 52.839057],
-                    [-2.073939, 52.839087],
-                    [-2.073021, 52.838924],
-                    [-2.072551, 52.838766],
-                    [-2.072265, 52.838604],
-                    [-2.071935, 52.838359],
-                    [-2.071611, 52.838278],
-                    [-2.071013, 52.838335],
-                    [-2.070941, 52.838367],
-                    [-2.070892, 52.838441],
-                    [-2.07088, 52.838552],
-                    [-2.070998, 52.838854],
-                    [-2.071, 52.839045],
-                    [-2.070927, 52.839185],
-                    [-2.070645, 52.839341],
-                    [-2.07045, 52.839367],
-                    [-2.070113, 52.83925],
-                    [-2.069924, 52.839087],
-                    [-2.06992, 52.838978]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400015,
-              "name": "Battle of Hopton Heath 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000015",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000015"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000015"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400016,
-          "name": "Battle of Langport 1645",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000016",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.797617, 51.04845],
-                    [-2.796988, 51.048503],
-                    [-2.796299, 51.048444],
-                    [-2.796173, 51.048461],
-                    [-2.79615, 51.048493],
-                    [-2.793455, 51.048766],
-                    [-2.793016, 51.0464],
-                    [-2.79485, 51.046267],
-                    [-2.794658, 51.045367],
-                    [-2.794233, 51.04544],
-                    [-2.794145, 51.04511],
-                    [-2.79384, 51.044438],
-                    [-2.793647, 51.043846],
-                    [-2.793441, 51.042903],
-                    [-2.79335, 51.04267],
-                    [-2.793369, 51.042642],
-                    [-2.793281, 51.042258],
-                    [-2.79321, 51.042072],
-                    [-2.792139, 51.041601],
-                    [-2.792434, 51.04041],
-                    [-2.795832, 51.040466],
-                    [-2.798117, 51.040677],
-                    [-2.800101, 51.040737],
-                    [-2.801172, 51.040763],
-                    [-2.801771, 51.040759],
-                    [-2.801822, 51.040731],
-                    [-2.801891, 51.040755],
-                    [-2.803712, 51.040736],
-                    [-2.804385, 51.041111],
-                    [-2.804379, 51.041161],
-                    [-2.803993, 51.041547],
-                    [-2.803861, 51.041735],
-                    [-2.803452, 51.041689],
-                    [-2.803271, 51.042816],
-                    [-2.803382, 51.04382],
-                    [-2.803362, 51.044387],
-                    [-2.803316, 51.044676],
-                    [-2.803776, 51.044719],
-                    [-2.803812, 51.045081],
-                    [-2.80435, 51.046446],
-                    [-2.804614, 51.047453],
-                    [-2.803108, 51.047627],
-                    [-2.802385, 51.04767],
-                    [-2.802345, 51.047754],
-                    [-2.802298, 51.047777],
-                    [-2.801769, 51.047762],
-                    [-2.801871, 51.048045],
-                    [-2.802132, 51.048547],
-                    [-2.801066, 51.048741],
-                    [-2.800304, 51.048788],
-                    [-2.800272, 51.048802],
-                    [-2.800262, 51.049113],
-                    [-2.79773, 51.049309],
-                    [-2.797756, 51.049108],
-                    [-2.797631, 51.048961],
-                    [-2.797656, 51.048791],
-                    [-2.797554, 51.048668],
-                    [-2.797617, 51.04845]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400016,
-              "name": "Battle of Langport 1645",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000016",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000016"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000016"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400017,
-          "name": "Battle of Lansdown (Hill) 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000017",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.391194, 51.440846],
-                    [-2.390265, 51.440122],
-                    [-2.389845, 51.439841],
-                    [-2.389955, 51.439495],
-                    [-2.390173, 51.439026],
-                    [-2.390427, 51.438688],
-                    [-2.390641, 51.438513],
-                    [-2.3919, 51.438086],
-                    [-2.391964, 51.438047],
-                    [-2.391906, 51.437899],
-                    [-2.392137, 51.437804],
-                    [-2.392223, 51.437708],
-                    [-2.392239, 51.437639],
-                    [-2.392074, 51.43729],
-                    [-2.392009, 51.43728],
-                    [-2.391971, 51.437213],
-                    [-2.391583, 51.436864],
-                    [-2.390975, 51.436455],
-                    [-2.390588, 51.436268],
-                    [-2.390312, 51.435497],
-                    [-2.390838, 51.434843],
-                    [-2.39134, 51.434126],
-                    [-2.392287, 51.432671],
-                    [-2.392613, 51.432112],
-                    [-2.393012, 51.431015],
-                    [-2.393086, 51.42995],
-                    [-2.393838, 51.429855],
-                    [-2.395398, 51.429743],
-                    [-2.395828, 51.429673],
-                    [-2.396678, 51.429769],
-                    [-2.397048, 51.429766],
-                    [-2.397477, 51.42969],
-                    [-2.397775, 51.429578],
-                    [-2.398149, 51.429311],
-                    [-2.398423, 51.429015],
-                    [-2.398662, 51.428551],
-                    [-2.398648, 51.428511],
-                    [-2.398442, 51.42851],
-                    [-2.398206, 51.427927],
-                    [-2.398023, 51.42795],
-                    [-2.397775, 51.427934],
-                    [-2.397667, 51.427715],
-                    [-2.397931, 51.427498],
-                    [-2.398043, 51.427519],
-                    [-2.401209, 51.426784],
-                    [-2.401751, 51.426695],
-                    [-2.4013, 51.425897],
-                    [-2.404495, 51.426148],
-                    [-2.409349, 51.426852],
-                    [-2.409547, 51.426991],
-                    [-2.409523, 51.427008],
-                    [-2.409512, 51.427484],
-                    [-2.412365, 51.430015],
-                    [-2.414031, 51.432104],
-                    [-2.414999, 51.433473],
-                    [-2.41483, 51.433537],
-                    [-2.414132, 51.433679],
-                    [-2.412208, 51.43395],
-                    [-2.411556, 51.434089],
-                    [-2.407243, 51.434745],
-                    [-2.406965, 51.434839],
-                    [-2.407413, 51.435468],
-                    [-2.407556, 51.435782],
-                    [-2.407639, 51.436108],
-                    [-2.407615, 51.436277],
-                    [-2.407521, 51.436508],
-                    [-2.407287, 51.436617],
-                    [-2.407131, 51.436742],
-                    [-2.4069, 51.43701],
-                    [-2.406612, 51.437452],
-                    [-2.406399, 51.43791],
-                    [-2.406329, 51.437985],
-                    [-2.406102, 51.438104],
-                    [-2.406025, 51.438095],
-                    [-2.405406, 51.438224],
-                    [-2.404886, 51.438351],
-                    [-2.404574, 51.438467],
-                    [-2.404626, 51.441721],
-                    [-2.404697, 51.44183],
-                    [-2.404776, 51.442216],
-                    [-2.404669, 51.442306],
-                    [-2.404255, 51.442409],
-                    [-2.402907, 51.442502],
-                    [-2.402337, 51.442507],
-                    [-2.401965, 51.442601],
-                    [-2.401284, 51.442834],
-                    [-2.400939, 51.442893],
-                    [-2.40089, 51.442867],
-                    [-2.400905, 51.442827],
-                    [-2.400805, 51.442813],
-                    [-2.400773, 51.44268],
-                    [-2.400634, 51.442492],
-                    [-2.398997, 51.443304],
-                    [-2.398728, 51.44349],
-                    [-2.399015, 51.443725],
-                    [-2.39925, 51.443825],
-                    [-2.399214, 51.443873],
-                    [-2.399239, 51.444018],
-                    [-2.399548, 51.444282],
-                    [-2.399435, 51.444302],
-                    [-2.399411, 51.444384],
-                    [-2.399316, 51.444431],
-                    [-2.39923, 51.444601],
-                    [-2.399066, 51.444695],
-                    [-2.398652, 51.444773],
-                    [-2.398293, 51.444648],
-                    [-2.397519, 51.445074],
-                    [-2.397297, 51.445332],
-                    [-2.397294, 51.445395],
-                    [-2.397364, 51.445508],
-                    [-2.39736, 51.445558],
-                    [-2.397025, 51.445903],
-                    [-2.396439, 51.446631],
-                    [-2.395992, 51.446665],
-                    [-2.395543, 51.447244],
-                    [-2.395569, 51.447253],
-                    [-2.395422, 51.447509],
-                    [-2.395425, 51.447634],
-                    [-2.395377, 51.447803],
-                    [-2.395332, 51.447841],
-                    [-2.395351, 51.447938],
-                    [-2.395396, 51.447949],
-                    [-2.395419, 51.448148],
-                    [-2.395562, 51.448462],
-                    [-2.396035, 51.449178],
-                    [-2.396027, 51.449236],
-                    [-2.396144, 51.449339],
-                    [-2.396557, 51.450709],
-                    [-2.396517, 51.450966],
-                    [-2.39642, 51.451086],
-                    [-2.396273, 51.451178],
-                    [-2.394605, 51.451686],
-                    [-2.393423, 51.452],
-                    [-2.39347, 51.452151],
-                    [-2.392754, 51.452295],
-                    [-2.391688, 51.452448],
-                    [-2.390234, 51.452592],
-                    [-2.389426, 51.45257],
-                    [-2.388328, 51.452382],
-                    [-2.387156, 51.452271],
-                    [-2.386635, 51.45219],
-                    [-2.384508, 51.45211],
-                    [-2.382913, 51.452013],
-                    [-2.382909, 51.45128],
-                    [-2.382804, 51.450954],
-                    [-2.382509, 51.450335],
-                    [-2.382342, 51.449774],
-                    [-2.382523, 51.449028],
-                    [-2.383046, 51.448142],
-                    [-2.383158, 51.447896],
-                    [-2.385264, 51.447937],
-                    [-2.385783, 51.446439],
-                    [-2.388096, 51.446971],
-                    [-2.388111, 51.447015],
-                    [-2.388198, 51.447057],
-                    [-2.388292, 51.447162],
-                    [-2.389728, 51.447266],
-                    [-2.3904, 51.446851],
-                    [-2.39059, 51.44668],
-                    [-2.390906, 51.446027],
-                    [-2.391158, 51.445675],
-                    [-2.390648, 51.445306],
-                    [-2.389664, 51.44476],
-                    [-2.389828, 51.444591],
-                    [-2.389878, 51.444496],
-                    [-2.389855, 51.444197],
-                    [-2.389874, 51.443979],
-                    [-2.389837, 51.443966],
-                    [-2.389867, 51.443932],
-                    [-2.389924, 51.443565],
-                    [-2.389912, 51.443523],
-                    [-2.38969, 51.443415],
-                    [-2.390898, 51.442172],
-                    [-2.391847, 51.441388],
-                    [-2.391194, 51.440846]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400017,
-              "name": "Battle of Lansdown (Hill) 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000017",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000017"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000017"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400018,
-          "name": "Battle of Lewes 1264",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000018",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.011677, 50.88498],
-                    [-0.011252, 50.884978],
-                    [-0.010922, 50.885083],
-                    [-0.010598, 50.885105],
-                    [-0.009769, 50.884917],
-                    [-0.008633, 50.882625],
-                    [-0.0081, 50.881684],
-                    [-0.005431, 50.880699],
-                    [-0.004185, 50.880319],
-                    [-0.003901, 50.880088],
-                    [-0.003746, 50.879874],
-                    [-0.003573, 50.879321],
-                    [-0.00369, 50.879192],
-                    [-0.00414, 50.879524],
-                    [-0.005269, 50.879942],
-                    [-0.006496, 50.88029],
-                    [-0.006562, 50.880236],
-                    [-0.00693, 50.880357],
-                    [-0.007738, 50.880711],
-                    [-0.008142, 50.880839],
-                    [-0.008552, 50.880922],
-                    [-0.010629, 50.881683],
-                    [-0.010928, 50.881759],
-                    [-0.0112, 50.881795],
-                    [-0.01171, 50.88178],
-                    [-0.013303, 50.881394],
-                    [-0.012698, 50.879581],
-                    [-0.012658, 50.879329],
-                    [-0.012577, 50.87916],
-                    [-0.011403, 50.875445],
-                    [-0.011425, 50.875186],
-                    [-0.010455, 50.874673],
-                    [-0.009738, 50.874435],
-                    [-0.009054, 50.874086],
-                    [-0.008668, 50.874172],
-                    [-0.008443, 50.874328],
-                    [-0.006015, 50.874884],
-                    [-0.005603, 50.874186],
-                    [-0.004813, 50.874355],
-                    [-0.004658, 50.873993],
-                    [-0.004236, 50.874091],
-                    [-0.004093, 50.874054],
-                    [-0.004011, 50.873911],
-                    [-0.003947, 50.873639],
-                    [-0.00394, 50.873442],
-                    [-0.004267, 50.873472],
-                    [-0.006334, 50.873853],
-                    [-0.007465, 50.873591],
-                    [-0.008583, 50.873398],
-                    [-0.009025, 50.873261],
-                    [-0.010055, 50.873079],
-                    [-0.010459, 50.873066],
-                    [-0.011406, 50.873133],
-                    [-0.011852, 50.873197],
-                    [-0.012414, 50.873322],
-                    [-0.012949, 50.873499],
-                    [-0.013369, 50.873694],
-                    [-0.013713, 50.873915],
-                    [-0.014018, 50.87417],
-                    [-0.014585, 50.874801],
-                    [-0.014954, 50.875249],
-                    [-0.01549, 50.87599],
-                    [-0.016997, 50.878148],
-                    [-0.01768, 50.878913],
-                    [-0.018178, 50.879393],
-                    [-0.018836, 50.879935],
-                    [-0.019586, 50.880668],
-                    [-0.023243, 50.883466],
-                    [-0.02367, 50.884352],
-                    [-0.023754, 50.884623],
-                    [-0.023702, 50.885086],
-                    [-0.023429, 50.885519],
-                    [-0.023273, 50.885671],
-                    [-0.023004, 50.885754],
-                    [-0.021634, 50.885947],
-                    [-0.020561, 50.886157],
-                    [-0.019495, 50.886435],
-                    [-0.018412, 50.886901],
-                    [-0.016598, 50.888001],
-                    [-0.015701, 50.888599],
-                    [-0.01543, 50.888715],
-                    [-0.015067, 50.888782],
-                    [-0.014437, 50.888779],
-                    [-0.013887, 50.888709],
-                    [-0.013307, 50.888592],
-                    [-0.012498, 50.888329],
-                    [-0.011274, 50.887833],
-                    [-0.011023, 50.887696],
-                    [-0.010871, 50.887554],
-                    [-0.010811, 50.887462],
-                    [-0.0108, 50.887358],
-                    [-0.010805, 50.887237],
-                    [-0.010879, 50.887133],
-                    [-0.011341, 50.886894],
-                    [-0.011687, 50.88678],
-                    [-0.011901, 50.88663],
-                    [-0.011954, 50.886514],
-                    [-0.011957, 50.886321],
-                    [-0.011932, 50.886275],
-                    [-0.011857, 50.886196],
-                    [-0.011588, 50.886066],
-                    [-0.011745, 50.885966],
-                    [-0.011976, 50.885954],
-                    [-0.011967, 50.885852],
-                    [-0.012119, 50.885775],
-                    [-0.01209, 50.885738],
-                    [-0.012151, 50.885679],
-                    [-0.012186, 50.885493],
-                    [-0.012491, 50.88549],
-                    [-0.012252, 50.885218],
-                    [-0.012062, 50.88509],
-                    [-0.011677, 50.88498]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400018,
-              "name": "Battle of Lewes 1264",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000018",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000018"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000018"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400019,
-          "name": "Battle of Maldon 991",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000019",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [0.701734, 51.721204],
-                    [0.70173, 51.721153],
-                    [0.701857, 51.721176],
-                    [0.701941, 51.721138],
-                    [0.702012, 51.721165],
-                    [0.702148, 51.721087],
-                    [0.702215, 51.721158],
-                    [0.702287, 51.721117],
-                    [0.702428, 51.721126],
-                    [0.70255, 51.721087],
-                    [0.702746, 51.721097],
-                    [0.702758, 51.721034],
-                    [0.702861, 51.720993],
-                    [0.702995, 51.721024],
-                    [0.703215, 51.720962],
-                    [0.703257, 51.720843],
-                    [0.70335, 51.720897],
-                    [0.703574, 51.72085],
-                    [0.703589, 51.720787],
-                    [0.703456, 51.720715],
-                    [0.703703, 51.720736],
-                    [0.703833, 51.720546],
-                    [0.703809, 51.720391],
-                    [0.703575, 51.720288],
-                    [0.703489, 51.720125],
-                    [0.703528, 51.720001],
-                    [0.703478, 51.719957],
-                    [0.702987, 51.719835],
-                    [0.703138, 51.719727],
-                    [0.703283, 51.719783],
-                    [0.703691, 51.719832],
-                    [0.703783, 51.719915],
-                    [0.70387, 51.720129],
-                    [0.704003, 51.720152],
-                    [0.704204, 51.720155],
-                    [0.704418, 51.72005],
-                    [0.704561, 51.720044],
-                    [0.704609, 51.719822],
-                    [0.704736, 51.719796],
-                    [0.704722, 51.71973],
-                    [0.704817, 51.719714],
-                    [0.704878, 51.719754],
-                    [0.704952, 51.719733],
-                    [0.704978, 51.719665],
-                    [0.70494, 51.719615],
-                    [0.704845, 51.719627],
-                    [0.704662, 51.719526],
-                    [0.704714, 51.71947],
-                    [0.704682, 51.719419],
-                    [0.704718, 51.719298],
-                    [0.70477, 51.719265],
-                    [0.704675, 51.719236],
-                    [0.704646, 51.719165],
-                    [0.704694, 51.71914],
-                    [0.704471, 51.718932],
-                    [0.70454, 51.718929],
-                    [0.704621, 51.718997],
-                    [0.704716, 51.719002],
-                    [0.704827, 51.719106],
-                    [0.704846, 51.719154],
-                    [0.704782, 51.719197],
-                    [0.70486, 51.719198],
-                    [0.704903, 51.719257],
-                    [0.704811, 51.719388],
-                    [0.705187, 51.719512],
-                    [0.705304, 51.719527],
-                    [0.705367, 51.719504],
-                    [0.705421, 51.719449],
-                    [0.705356, 51.719375],
-                    [0.705363, 51.719339],
-                    [0.705601, 51.719272],
-                    [0.705599, 51.719129],
-                    [0.705754, 51.719118],
-                    [0.705761, 51.71905],
-                    [0.705835, 51.719016],
-                    [0.705788, 51.718988],
-                    [0.705723, 51.718866],
-                    [0.705878, 51.71885],
-                    [0.706043, 51.718728],
-                    [0.706107, 51.718618],
-                    [0.706219, 51.718538],
-                    [0.706407, 51.718531],
-                    [0.706746, 51.718396],
-                    [0.706455, 51.718217],
-                    [0.706435, 51.718135],
-                    [0.706678, 51.718144],
-                    [0.70692, 51.718224],
-                    [0.707301, 51.718035],
-                    [0.707414, 51.718023],
-                    [0.707419, 51.717986],
-                    [0.707341, 51.71792],
-                    [0.707273, 51.717909],
-                    [0.707123, 51.717978],
-                    [0.707039, 51.717967],
-                    [0.706941, 51.717855],
-                    [0.707158, 51.717676],
-                    [0.706906, 51.717648],
-                    [0.706859, 51.717621],
-                    [0.70688, 51.717597],
-                    [0.70717, 51.717619],
-                    [0.70735, 51.717586],
-                    [0.707389, 51.717609],
-                    [0.707319, 51.717714],
-                    [0.70737, 51.717798],
-                    [0.707616, 51.717863],
-                    [0.707705, 51.717941],
-                    [0.7078, 51.717896],
-                    [0.707983, 51.717901],
-                    [0.708161, 51.717831],
-                    [0.708212, 51.717807],
-                    [0.708193, 51.717732],
-                    [0.708368, 51.717629],
-                    [0.708404, 51.717468],
-                    [0.708464, 51.717443],
-                    [0.708649, 51.717495],
-                    [0.708769, 51.717478],
-                    [0.708967, 51.717361],
-                    [0.708992, 51.717407],
-                    [0.709052, 51.717396],
-                    [0.709244, 51.717175],
-                    [0.709333, 51.717145],
-                    [0.709412, 51.717063],
-                    [0.70943, 51.716987],
-                    [0.70919, 51.71687],
-                    [0.709216, 51.716938],
-                    [0.709201, 51.717041],
-                    [0.709086, 51.717092],
-                    [0.709061, 51.716916],
-                    [0.70898, 51.716814],
-                    [0.709213, 51.71678],
-                    [0.709415, 51.71686],
-                    [0.709448, 51.716836],
-                    [0.709319, 51.716783],
-                    [0.70926, 51.71669],
-                    [0.709267, 51.716543],
-                    [0.709315, 51.716472],
-                    [0.706758, 51.715444],
-                    [0.705957, 51.715205],
-                    [0.705122, 51.715024],
-                    [0.704659, 51.714883],
-                    [0.704445, 51.71485],
-                    [0.70373, 51.714849],
-                    [0.703146, 51.714762],
-                    [0.701589, 51.71465],
-                    [0.701239, 51.714554],
-                    [0.700858, 51.714497],
-                    [0.697517, 51.71681],
-                    [0.697166, 51.717191],
-                    [0.697443, 51.717271],
-                    [0.697322, 51.717455],
-                    [0.697016, 51.717388],
-                    [0.696944, 51.717516],
-                    [0.696681, 51.717465],
-                    [0.6964, 51.718108],
-                    [0.69642, 51.718179],
-                    [0.696189, 51.718554],
-                    [0.69539, 51.719452],
-                    [0.695012, 51.719822],
-                    [0.696675, 51.720537],
-                    [0.697472, 51.720206],
-                    [0.697721, 51.720022],
-                    [0.697746, 51.719925],
-                    [0.697843, 51.719824],
-                    [0.698312, 51.719538],
-                    [0.698271, 51.719426],
-                    [0.698567, 51.719339],
-                    [0.69855, 51.719514],
-                    [0.698822, 51.719718],
-                    [0.700349, 51.720505],
-                    [0.700473, 51.720618],
-                    [0.700931, 51.720784],
-                    [0.700986, 51.72074],
-                    [0.701158, 51.720888],
-                    [0.701269, 51.720919],
-                    [0.701302, 51.720906],
-                    [0.701257, 51.72077],
-                    [0.701105, 51.720682],
-                    [0.701081, 51.720597],
-                    [0.701236, 51.720411],
-                    [0.701142, 51.720607],
-                    [0.701158, 51.720658],
-                    [0.701262, 51.720701],
-                    [0.701346, 51.720791],
-                    [0.701378, 51.720876],
-                    [0.701485, 51.720976],
-                    [0.701519, 51.721069],
-                    [0.701734, 51.721204]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400019,
-              "name": "Battle of Maldon 991",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000019",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000019"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000019"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400020,
-          "name": "Battle of Marston Moor 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000020",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.243221, 53.975839],
-                    [-1.241368, 53.97403],
-                    [-1.238709, 53.971546],
-                    [-1.23914, 53.971224],
-                    [-1.234898, 53.968143],
-                    [-1.23571, 53.967863],
-                    [-1.23566, 53.967806],
-                    [-1.235436, 53.967683],
-                    [-1.23523, 53.967512],
-                    [-1.234877, 53.967293],
-                    [-1.232376, 53.966074],
-                    [-1.232978, 53.965851],
-                    [-1.231043, 53.963321],
-                    [-1.231812, 53.963082],
-                    [-1.232023, 53.962983],
-                    [-1.238072, 53.962059],
-                    [-1.23857, 53.961547],
-                    [-1.239544, 53.960868],
-                    [-1.239891, 53.960543],
-                    [-1.239985, 53.960535],
-                    [-1.240288, 53.96035],
-                    [-1.242982, 53.958504],
-                    [-1.242964, 53.95846],
-                    [-1.242267, 53.958036],
-                    [-1.241633, 53.957752],
-                    [-1.241964, 53.957455],
-                    [-1.24226, 53.957305],
-                    [-1.243835, 53.95579],
-                    [-1.245408, 53.954582],
-                    [-1.246075, 53.955013],
-                    [-1.246573, 53.955273],
-                    [-1.248359, 53.956091],
-                    [-1.253644, 53.954102],
-                    [-1.257097, 53.956387],
-                    [-1.261063, 53.95487],
-                    [-1.261564, 53.954915],
-                    [-1.26482, 53.955393],
-                    [-1.270496, 53.955766],
-                    [-1.274437, 53.956061],
-                    [-1.275231, 53.956228],
-                    [-1.276128, 53.956608],
-                    [-1.277237, 53.957026],
-                    [-1.27857, 53.957332],
-                    [-1.278227, 53.957661],
-                    [-1.27812, 53.957823],
-                    [-1.280442, 53.959458],
-                    [-1.281476, 53.95997],
-                    [-1.28219, 53.96024],
-                    [-1.282415, 53.960465],
-                    [-1.282408, 53.960508],
-                    [-1.282544, 53.960742],
-                    [-1.282446, 53.961003],
-                    [-1.282454, 53.961202],
-                    [-1.282692, 53.961651],
-                    [-1.282821, 53.961791],
-                    [-1.283399, 53.96219],
-                    [-1.283445, 53.962263],
-                    [-1.283467, 53.962502],
-                    [-1.283741, 53.962761],
-                    [-1.283751, 53.963074],
-                    [-1.283471, 53.96384],
-                    [-1.283474, 53.963942],
-                    [-1.283239, 53.963892],
-                    [-1.283043, 53.963903],
-                    [-1.282938, 53.963948],
-                    [-1.282851, 53.964098],
-                    [-1.282838, 53.964194],
-                    [-1.282689, 53.964347],
-                    [-1.282614, 53.964523],
-                    [-1.282603, 53.9647],
-                    [-1.282373, 53.965213],
-                    [-1.282127, 53.965462],
-                    [-1.282139, 53.965539],
-                    [-1.282063, 53.965681],
-                    [-1.282075, 53.965778],
-                    [-1.281963, 53.966813],
-                    [-1.281964, 53.967142],
-                    [-1.282005, 53.967381],
-                    [-1.282085, 53.967632],
-                    [-1.28218, 53.96779],
-                    [-1.282783, 53.968546],
-                    [-1.282843, 53.968805],
-                    [-1.283182, 53.968886],
-                    [-1.283591, 53.968769],
-                    [-1.283814, 53.968915],
-                    [-1.284103, 53.969171],
-                    [-1.284574, 53.969718],
-                    [-1.285559, 53.970728],
-                    [-1.285515, 53.970786],
-                    [-1.285555, 53.970791],
-                    [-1.285537, 53.970817],
-                    [-1.285401, 53.970891],
-                    [-1.285081, 53.971176],
-                    [-1.284269, 53.971741],
-                    [-1.284496, 53.971945],
-                    [-1.284797, 53.972112],
-                    [-1.286199, 53.972325],
-                    [-1.286389, 53.972411],
-                    [-1.285956, 53.972729],
-                    [-1.285732, 53.972939],
-                    [-1.284774, 53.974667],
-                    [-1.284697, 53.974734],
-                    [-1.284692, 53.974877],
-                    [-1.284537, 53.975269],
-                    [-1.274316, 53.97643],
-                    [-1.274147, 53.97532],
-                    [-1.273565, 53.975378],
-                    [-1.273489, 53.975071],
-                    [-1.268564, 53.975701],
-                    [-1.266714, 53.976065],
-                    [-1.265918, 53.976302],
-                    [-1.265202, 53.97489],
-                    [-1.257475, 53.976088],
-                    [-1.255044, 53.975694],
-                    [-1.254923, 53.975746],
-                    [-1.254043, 53.975936],
-                    [-1.245273, 53.976262],
-                    [-1.245147, 53.976296],
-                    [-1.244396, 53.976955],
-                    [-1.243221, 53.975839]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400020,
-              "name": "Battle of Marston Moor 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000020",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000020"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000020"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400021,
-          "name": "Battle of Myton 1319",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000021",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.337371, 54.106141],
-                    [-1.336842, 54.106024],
-                    [-1.336607, 54.106004],
-                    [-1.336073, 54.106013],
-                    [-1.33584, 54.106049],
-                    [-1.331048, 54.103656],
-                    [-1.331031, 54.103429],
-                    [-1.330968, 54.103313],
-                    [-1.330818, 54.103205],
-                    [-1.330561, 54.103111],
-                    [-1.330135, 54.103033],
-                    [-1.329788, 54.103016],
-                    [-1.329412, 54.103069],
-                    [-1.329206, 54.103064],
-                    [-1.32865, 54.102964],
-                    [-1.328217, 54.102787],
-                    [-1.32795, 54.102596],
-                    [-1.327795, 54.102451],
-                    [-1.327579, 54.102161],
-                    [-1.327496, 54.101909],
-                    [-1.327514, 54.101708],
-                    [-1.327625, 54.101538],
-                    [-1.327794, 54.101409],
-                    [-1.32802, 54.101302],
-                    [-1.328287, 54.101228],
-                    [-1.328978, 54.101164],
-                    [-1.329589, 54.101054],
-                    [-1.329898, 54.100972],
-                    [-1.330367, 54.100775],
-                    [-1.330704, 54.100628],
-                    [-1.330932, 54.100385],
-                    [-1.331007, 54.100344],
-                    [-1.331114, 54.100198],
-                    [-1.331095, 54.100167],
-                    [-1.331183, 54.099958],
-                    [-1.331207, 54.099595],
-                    [-1.331172, 54.099432],
-                    [-1.331031, 54.099166],
-                    [-1.33098, 54.098996],
-                    [-1.330856, 54.098796],
-                    [-1.330864, 54.09843],
-                    [-1.331054, 54.09787],
-                    [-1.331014, 54.097436],
-                    [-1.330957, 54.097266],
-                    [-1.330759, 54.096936],
-                    [-1.330595, 54.096785],
-                    [-1.330383, 54.096513],
-                    [-1.330008, 54.096181],
-                    [-1.329927, 54.096054],
-                    [-1.32931, 54.095623],
-                    [-1.329057, 54.095328],
-                    [-1.32906, 54.095176],
-                    [-1.329109, 54.095066],
-                    [-1.329548, 54.094618],
-                    [-1.330113, 54.094355],
-                    [-1.330874, 54.094217],
-                    [-1.331608, 54.094249],
-                    [-1.33219, 54.094401],
-                    [-1.332367, 54.094494],
-                    [-1.332862, 54.09466],
-                    [-1.333284, 54.094759],
-                    [-1.334103, 54.094712],
-                    [-1.334912, 54.094601],
-                    [-1.335461, 54.094489],
-                    [-1.335648, 54.094446],
-                    [-1.335922, 54.094278],
-                    [-1.336117, 54.0941],
-                    [-1.336979, 54.09313],
-                    [-1.33741, 54.092922],
-                    [-1.338007, 54.092768],
-                    [-1.338443, 54.092781],
-                    [-1.338613, 54.092821],
-                    [-1.339301, 54.092869],
-                    [-1.33986, 54.092845],
-                    [-1.340623, 54.092714],
-                    [-1.342429, 54.092291],
-                    [-1.343013, 54.092115],
-                    [-1.34341, 54.091952],
-                    [-1.344252, 54.091485],
-                    [-1.344204, 54.093099],
-                    [-1.345619, 54.093156],
-                    [-1.349736, 54.09345],
-                    [-1.362482, 54.095636],
-                    [-1.362624, 54.095618],
-                    [-1.36278, 54.095654],
-                    [-1.361556, 54.097204],
-                    [-1.361301, 54.098214],
-                    [-1.361193, 54.099602],
-                    [-1.360927, 54.100432],
-                    [-1.36045, 54.101732],
-                    [-1.360068, 54.103015],
-                    [-1.360021, 54.103071],
-                    [-1.359648, 54.104039],
-                    [-1.349423, 54.106218],
-                    [-1.349378, 54.106197],
-                    [-1.349248, 54.106268],
-                    [-1.348451, 54.106437],
-                    [-1.340901, 54.107822],
-                    [-1.340846, 54.107577],
-                    [-1.340617, 54.107164],
-                    [-1.340453, 54.107017],
-                    [-1.340181, 54.10687],
-                    [-1.339417, 54.106622],
-                    [-1.338997, 54.106541],
-                    [-1.338384, 54.106351],
-                    [-1.338014, 54.106304],
-                    [-1.337371, 54.106141]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400021,
-              "name": "Battle of Myton 1319",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000021",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000021"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000021"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400022,
-          "name": "Battle of Nantwich 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000022",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.548016, 53.085611],
-                    [-2.542191, 53.082721],
-                    [-2.540038, 53.082638],
-                    [-2.539318, 53.08258],
-                    [-2.537567, 53.082346],
-                    [-2.536368, 53.082113],
-                    [-2.533601, 53.081116],
-                    [-2.533994, 53.080666],
-                    [-2.534234, 53.080308],
-                    [-2.534646, 53.079833],
-                    [-2.535076, 53.079431],
-                    [-2.535863, 53.078608],
-                    [-2.536312, 53.078231],
-                    [-2.53719, 53.077649],
-                    [-2.537445, 53.077409],
-                    [-2.537453, 53.077341],
-                    [-2.537333, 53.076892],
-                    [-2.537288, 53.075984],
-                    [-2.537098, 53.074815],
-                    [-2.536931, 53.074602],
-                    [-2.536683, 53.074447],
-                    [-2.535742, 53.073978],
-                    [-2.535574, 53.073809],
-                    [-2.535455, 53.073563],
-                    [-2.53544, 53.073454],
-                    [-2.535487, 53.073252],
-                    [-2.53559, 53.073077],
-                    [-2.536098, 53.072669],
-                    [-2.53616, 53.072578],
-                    [-2.536291, 53.072204],
-                    [-2.536211, 53.071927],
-                    [-2.536023, 53.071547],
-                    [-2.536025, 53.071369],
-                    [-2.536161, 53.07095],
-                    [-2.536127, 53.07085],
-                    [-2.535102, 53.070165],
-                    [-2.534943, 53.069955],
-                    [-2.534886, 53.069818],
-                    [-2.534893, 53.069639],
-                    [-2.534948, 53.069465],
-                    [-2.53501, 53.069343],
-                    [-2.535155, 53.069185],
-                    [-2.536894, 53.06964],
-                    [-2.538661, 53.070295],
-                    [-2.539625, 53.070556],
-                    [-2.54027, 53.070687],
-                    [-2.54163, 53.070827],
-                    [-2.544827, 53.070887],
-                    [-2.546624, 53.071017],
-                    [-2.548778, 53.071281],
-                    [-2.549394, 53.071475],
-                    [-2.549571, 53.071584],
-                    [-2.549727, 53.071751],
-                    [-2.549862, 53.071974],
-                    [-2.550106, 53.073125],
-                    [-2.548733, 53.073267],
-                    [-2.548712, 53.07321],
-                    [-2.548637, 53.073217],
-                    [-2.548303, 53.073265],
-                    [-2.548202, 53.073359],
-                    [-2.548008, 53.073399],
-                    [-2.547744, 53.073383],
-                    [-2.547511, 53.073453],
-                    [-2.547428, 53.073415],
-                    [-2.547007, 53.073474],
-                    [-2.547344, 53.074449],
-                    [-2.547056, 53.074482],
-                    [-2.547487, 53.075198],
-                    [-2.547704, 53.07548],
-                    [-2.547794, 53.075508],
-                    [-2.548104, 53.075578],
-                    [-2.548065, 53.075484],
-                    [-2.548089, 53.075432],
-                    [-2.550031, 53.074893],
-                    [-2.550511, 53.074824],
-                    [-2.551134, 53.074769],
-                    [-2.551171, 53.074917],
-                    [-2.553748, 53.074679],
-                    [-2.553945, 53.075342],
-                    [-2.554155, 53.075751],
-                    [-2.554306, 53.076252],
-                    [-2.554363, 53.076322],
-                    [-2.554411, 53.076539],
-                    [-2.554518, 53.076667],
-                    [-2.554641, 53.076718],
-                    [-2.554861, 53.076567],
-                    [-2.554935, 53.076716],
-                    [-2.55508, 53.077124],
-                    [-2.555378, 53.078238],
-                    [-2.5554, 53.078269],
-                    [-2.555579, 53.078299],
-                    [-2.555564, 53.078349],
-                    [-2.555836, 53.078587],
-                    [-2.55618, 53.079046],
-                    [-2.556428, 53.079173],
-                    [-2.5568, 53.079541],
-                    [-2.556854, 53.079556],
-                    [-2.55689, 53.080024],
-                    [-2.556985, 53.080275],
-                    [-2.556983, 53.080488],
-                    [-2.557006, 53.080574],
-                    [-2.55712, 53.080735],
-                    [-2.556914, 53.081182],
-                    [-2.556935, 53.081424],
-                    [-2.556614, 53.081883],
-                    [-2.556557, 53.082035],
-                    [-2.556264, 53.082323],
-                    [-2.556244, 53.082478],
-                    [-2.556211, 53.082501],
-                    [-2.555829, 53.082593],
-                    [-2.55564, 53.082748],
-                    [-2.55529, 53.082657],
-                    [-2.554702, 53.082627],
-                    [-2.55472, 53.083122],
-                    [-2.554067, 53.084062],
-                    [-2.554027, 53.084171],
-                    [-2.553604, 53.084311],
-                    [-2.553119, 53.084386],
-                    [-2.552991, 53.084676],
-                    [-2.552835, 53.084709],
-                    [-2.552364, 53.084918],
-                    [-2.552377, 53.085054],
-                    [-2.552111, 53.085248],
-                    [-2.55212, 53.085386],
-                    [-2.552063, 53.085526],
-                    [-2.551043, 53.085456],
-                    [-2.550726, 53.085394],
-                    [-2.550263, 53.085588],
-                    [-2.549601, 53.085709],
-                    [-2.549226, 53.08571],
-                    [-2.548016, 53.085611]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400022,
-              "name": "Battle of Nantwich 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000022",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000022"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000022"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400023,
-          "name": "Battle of Naseby 1645",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000023",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.007403, 52.42332],
-                    [-1.005355, 52.422805],
-                    [-1.005145, 52.422966],
-                    [-1.004834, 52.423106],
-                    [-1.003816, 52.422769],
-                    [-1.003253, 52.422519],
-                    [-1.002826, 52.422372],
-                    [-1.001483, 52.422092],
-                    [-0.999195, 52.421909],
-                    [-0.998439, 52.421899],
-                    [-0.998165, 52.42184],
-                    [-0.99789, 52.421823],
-                    [-0.997405, 52.421824],
-                    [-0.996315, 52.421895],
-                    [-0.996122, 52.421871],
-                    [-0.99545, 52.421651],
-                    [-0.994476, 52.421607],
-                    [-0.992736, 52.421301],
-                    [-0.992695, 52.421433],
-                    [-0.992679, 52.42228],
-                    [-0.98949, 52.421966],
-                    [-0.987606, 52.421664],
-                    [-0.98547, 52.417955],
-                    [-0.98515, 52.417848],
-                    [-0.984679, 52.417582],
-                    [-0.983759, 52.417134],
-                    [-0.983491, 52.417108],
-                    [-0.983339, 52.417176],
-                    [-0.982026, 52.416469],
-                    [-0.98125, 52.415947],
-                    [-0.981166, 52.41594],
-                    [-0.980574, 52.415513],
-                    [-0.980344, 52.415385],
-                    [-0.980189, 52.415369],
-                    [-0.980016, 52.415155],
-                    [-0.980003, 52.415057],
-                    [-0.979856, 52.414774],
-                    [-0.979598, 52.414061],
-                    [-0.979423, 52.413792],
-                    [-0.979307, 52.41379],
-                    [-0.979154, 52.413733],
-                    [-0.978868, 52.41374],
-                    [-0.978963, 52.413555],
-                    [-0.978993, 52.41326],
-                    [-0.979077, 52.413139],
-                    [-0.979058, 52.412922],
-                    [-0.979225, 52.412859],
-                    [-0.979185, 52.412707],
-                    [-0.97919, 52.412516],
-                    [-0.979298, 52.412433],
-                    [-0.979346, 52.412227],
-                    [-0.979289, 52.412124],
-                    [-0.979271, 52.411929],
-                    [-0.979187, 52.411762],
-                    [-0.985779, 52.409477],
-                    [-0.985892, 52.409317],
-                    [-0.985785, 52.409291],
-                    [-0.987616, 52.406457],
-                    [-0.988195, 52.405333],
-                    [-0.992713, 52.405047],
-                    [-0.992343, 52.399915],
-                    [-0.992403, 52.399887],
-                    [-0.993778, 52.400147],
-                    [-0.995059, 52.399055],
-                    [-0.993988, 52.398815],
-                    [-0.994122, 52.39808],
-                    [-0.994127, 52.397606],
-                    [-0.994278, 52.396954],
-                    [-0.994281, 52.396423],
-                    [-0.993541, 52.39635],
-                    [-0.993791, 52.396087],
-                    [-0.993878, 52.395941],
-                    [-0.994351, 52.395516],
-                    [-0.994608, 52.395506],
-                    [-0.994713, 52.395458],
-                    [-0.995413, 52.394826],
-                    [-0.995977, 52.394408],
-                    [-0.995633, 52.394043],
-                    [-0.995624, 52.393974],
-                    [-0.995242, 52.393788],
-                    [-0.995768, 52.393588],
-                    [-1.000656, 52.392567],
-                    [-1.001857, 52.393626],
-                    [-1.003096, 52.394455],
-                    [-1.003653, 52.394887],
-                    [-1.004669, 52.394733],
-                    [-1.005184, 52.397065],
-                    [-1.004288, 52.397028],
-                    [-1.004237, 52.399209],
-                    [-1.005113, 52.399815],
-                    [-1.007083, 52.401269],
-                    [-1.008279, 52.402087],
-                    [-1.007836, 52.403923],
-                    [-1.007165, 52.404799],
-                    [-1.008128, 52.405098],
-                    [-1.007998, 52.407085],
-                    [-1.007737, 52.408924],
-                    [-1.007989, 52.409053],
-                    [-1.008366, 52.40943],
-                    [-1.009063, 52.409761],
-                    [-1.012041, 52.411971],
-                    [-1.013524, 52.418662],
-                    [-1.009514, 52.420523],
-                    [-1.00946, 52.421861],
-                    [-1.009286, 52.422484],
-                    [-1.009133, 52.422773],
-                    [-1.008921, 52.423013],
-                    [-1.008557, 52.423307],
-                    [-1.007403, 52.42332]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400023,
-              "name": "Battle of Naseby 1645",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000023",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000023"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000023"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400024,
-          "name": "Battle of Neville's Cross 1346",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000024",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.597759, 54.780684],
-                    [-1.597818, 54.780622],
-                    [-1.598013, 54.780682],
-                    [-1.599416, 54.78023],
-                    [-1.600063, 54.779938],
-                    [-1.600292, 54.779664],
-                    [-1.599281, 54.779486],
-                    [-1.599573, 54.778855],
-                    [-1.599443, 54.778781],
-                    [-1.599485, 54.778663],
-                    [-1.599469, 54.778615],
-                    [-1.599537, 54.778492],
-                    [-1.599263, 54.778446],
-                    [-1.599286, 54.778391],
-                    [-1.598548, 54.778385],
-                    [-1.598403, 54.777994],
-                    [-1.598063, 54.777283],
-                    [-1.598156, 54.777263],
-                    [-1.598403, 54.777347],
-                    [-1.598657, 54.777265],
-                    [-1.599093, 54.777056],
-                    [-1.598815, 54.776863],
-                    [-1.59871, 54.776668],
-                    [-1.599604, 54.776576],
-                    [-1.599822, 54.776027],
-                    [-1.599134, 54.775688],
-                    [-1.598156, 54.775292],
-                    [-1.598072, 54.775227],
-                    [-1.598043, 54.775106],
-                    [-1.598034, 54.774653],
-                    [-1.598057, 54.774607],
-                    [-1.597868, 54.774481],
-                    [-1.596741, 54.774731],
-                    [-1.596666, 54.7746],
-                    [-1.59494, 54.77247],
-                    [-1.594846, 54.772172],
-                    [-1.594919, 54.771983],
-                    [-1.596415, 54.771996],
-                    [-1.596576, 54.771839],
-                    [-1.596584, 54.7716],
-                    [-1.596748, 54.77146],
-                    [-1.59787, 54.772007],
-                    [-1.598877, 54.771999],
-                    [-1.598987, 54.772187],
-                    [-1.599056, 54.772237],
-                    [-1.599105, 54.772359],
-                    [-1.599219, 54.772471],
-                    [-1.599505, 54.772682],
-                    [-1.599681, 54.772749],
-                    [-1.5997, 54.772786],
-                    [-1.600318, 54.773066],
-                    [-1.600373, 54.773118],
-                    [-1.600636, 54.773164],
-                    [-1.601776, 54.773543],
-                    [-1.602564, 54.773841],
-                    [-1.603002, 54.774052],
-                    [-1.603236, 54.774092],
-                    [-1.60367, 54.774086],
-                    [-1.603919, 54.774038],
-                    [-1.604177, 54.774072],
-                    [-1.604363, 54.774211],
-                    [-1.604382, 54.774291],
-                    [-1.604475, 54.774412],
-                    [-1.6046, 54.774513],
-                    [-1.60469, 54.774798],
-                    [-1.604683, 54.774909],
-                    [-1.604491, 54.775151],
-                    [-1.603883, 54.775741],
-                    [-1.603807, 54.775869],
-                    [-1.603831, 54.775928],
-                    [-1.603808, 54.776042],
-                    [-1.603648, 54.776389],
-                    [-1.603705, 54.776562],
-                    [-1.603873, 54.776805],
-                    [-1.604843, 54.777512],
-                    [-1.605724, 54.777726],
-                    [-1.607328, 54.777704],
-                    [-1.607459, 54.777655],
-                    [-1.607981, 54.777577],
-                    [-1.608003, 54.777868],
-                    [-1.608056, 54.777913],
-                    [-1.607988, 54.778075],
-                    [-1.608002, 54.778107],
-                    [-1.607769, 54.778642],
-                    [-1.606978, 54.779382],
-                    [-1.605872, 54.779887],
-                    [-1.605573, 54.779981],
-                    [-1.605617, 54.780024],
-                    [-1.605159, 54.780188],
-                    [-1.604748, 54.780151],
-                    [-1.603289, 54.780646],
-                    [-1.603146, 54.780564],
-                    [-1.602552, 54.780882],
-                    [-1.602433, 54.780914],
-                    [-1.601791, 54.780831],
-                    [-1.601597, 54.780681],
-                    [-1.597569, 54.780939],
-                    [-1.597759, 54.780684]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400024,
-              "name": "Battle of Neville's Cross 1346",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000024",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000024"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000024"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400025,
-          "name": "Battle of Newburn Ford 1640",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000025",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.751486, 54.989718],
-                    [-1.751067, 54.989589],
-                    [-1.751, 54.989616],
-                    [-1.750292, 54.989355],
-                    [-1.74992, 54.989409],
-                    [-1.749438, 54.989201],
-                    [-1.749184, 54.988991],
-                    [-1.749015, 54.988742],
-                    [-1.74871, 54.98854],
-                    [-1.747915, 54.988358],
-                    [-1.747923, 54.988487],
-                    [-1.747565, 54.9888],
-                    [-1.746484, 54.989229],
-                    [-1.744463, 54.987816],
-                    [-1.743571, 54.986897],
-                    [-1.743497, 54.986696],
-                    [-1.743491, 54.986148],
-                    [-1.745663, 54.985259],
-                    [-1.745632, 54.985228],
-                    [-1.747317, 54.984507],
-                    [-1.747002, 54.984376],
-                    [-1.746618, 54.984135],
-                    [-1.744767, 54.982621],
-                    [-1.744155, 54.982203],
-                    [-1.743618, 54.981948],
-                    [-1.742496, 54.981329],
-                    [-1.74177, 54.981045],
-                    [-1.740951, 54.980895],
-                    [-1.739992, 54.98077],
-                    [-1.737642, 54.980489],
-                    [-1.73632, 54.98042],
-                    [-1.734975, 54.98041],
-                    [-1.733742, 54.980263],
-                    [-1.729182, 54.979369],
-                    [-1.72614, 54.978856],
-                    [-1.726383, 54.978438],
-                    [-1.726398, 54.978332],
-                    [-1.72633, 54.978271],
-                    [-1.72606, 54.978194],
-                    [-1.724744, 54.977978],
-                    [-1.724786, 54.977881],
-                    [-1.724869, 54.977866],
-                    [-1.725184, 54.976845],
-                    [-1.726346, 54.975724],
-                    [-1.725863, 54.9754],
-                    [-1.725715, 54.974503],
-                    [-1.72578, 54.973705],
-                    [-1.725851, 54.973457],
-                    [-1.726222, 54.972958],
-                    [-1.726661, 54.97258],
-                    [-1.72799, 54.971746],
-                    [-1.730474, 54.970084],
-                    [-1.730523, 54.970037],
-                    [-1.7305, 54.970026],
-                    [-1.730827, 54.969912],
-                    [-1.731014, 54.969719],
-                    [-1.731197, 54.969638],
-                    [-1.731137, 54.969575],
-                    [-1.731576, 54.969328],
-                    [-1.731679, 54.969238],
-                    [-1.732376, 54.968532],
-                    [-1.731177, 54.968219],
-                    [-1.731612, 54.967626],
-                    [-1.731415, 54.967046],
-                    [-1.731454, 54.966844],
-                    [-1.73155, 54.966598],
-                    [-1.731836, 54.966279],
-                    [-1.732513, 54.965293],
-                    [-1.733295, 54.965215],
-                    [-1.733533, 54.965534],
-                    [-1.73407, 54.966006],
-                    [-1.734573, 54.966181],
-                    [-1.736117, 54.966351],
-                    [-1.736587, 54.966446],
-                    [-1.73859, 54.967111],
-                    [-1.74027, 54.96739],
-                    [-1.740915, 54.967682],
-                    [-1.741214, 54.967786],
-                    [-1.74282, 54.968156],
-                    [-1.744279, 54.968826],
-                    [-1.745012, 54.969202],
-                    [-1.74583, 54.969747],
-                    [-1.746156, 54.969902],
-                    [-1.745691, 54.970187],
-                    [-1.745594, 54.970277],
-                    [-1.744614, 54.971426],
-                    [-1.743984, 54.972047],
-                    [-1.743874, 54.972294],
-                    [-1.743896, 54.972438],
-                    [-1.744271, 54.972448],
-                    [-1.74394, 54.972724],
-                    [-1.745796, 54.973051],
-                    [-1.745969, 54.973213],
-                    [-1.745286, 54.973678],
-                    [-1.745303, 54.973837],
-                    [-1.745399, 54.973952],
-                    [-1.745572, 54.974081],
-                    [-1.745735, 54.974139],
-                    [-1.746027, 54.974325],
-                    [-1.746415, 54.974484],
-                    [-1.746589, 54.974597],
-                    [-1.747038, 54.974679],
-                    [-1.747365, 54.974678],
-                    [-1.747671, 54.974825],
-                    [-1.748152, 54.97512],
-                    [-1.748398, 54.975209],
-                    [-1.750009, 54.975492],
-                    [-1.751243, 54.975672],
-                    [-1.751182, 54.975813],
-                    [-1.75118, 54.976078],
-                    [-1.751343, 54.97612],
-                    [-1.751682, 54.976088],
-                    [-1.752258, 54.975862],
-                    [-1.75257, 54.975704],
-                    [-1.75269, 54.975514],
-                    [-1.752687, 54.975239],
-                    [-1.753322, 54.975287],
-                    [-1.753516, 54.975335],
-                    [-1.753673, 54.975418],
-                    [-1.753975, 54.975296],
-                    [-1.754515, 54.975191],
-                    [-1.754438, 54.975055],
-                    [-1.754401, 54.974905],
-                    [-1.755375, 54.975821],
-                    [-1.75563, 54.976214],
-                    [-1.755691, 54.976413],
-                    [-1.756765, 54.976435],
-                    [-1.757511, 54.976555],
-                    [-1.758643, 54.976496],
-                    [-1.759178, 54.976449],
-                    [-1.759316, 54.976411],
-                    [-1.759854, 54.976382],
-                    [-1.760984, 54.976397],
-                    [-1.761357, 54.976432],
-                    [-1.761719, 54.976559],
-                    [-1.762234, 54.976653],
-                    [-1.763373, 54.976642],
-                    [-1.764258, 54.977096],
-                    [-1.764179, 54.977352],
-                    [-1.764246, 54.977358],
-                    [-1.764254, 54.977876],
-                    [-1.764619, 54.978605],
-                    [-1.764703, 54.97883],
-                    [-1.764783, 54.97959],
-                    [-1.766101, 54.980128],
-                    [-1.766206, 54.981914],
-                    [-1.766359, 54.983456],
-                    [-1.766277, 54.983465],
-                    [-1.766267, 54.983897],
-                    [-1.766236, 54.983942],
-                    [-1.765941, 54.983982],
-                    [-1.76443, 54.984044],
-                    [-1.763444, 54.984035],
-                    [-1.762984, 54.98395],
-                    [-1.76242, 54.983906],
-                    [-1.762244, 54.983844],
-                    [-1.761973, 54.983963],
-                    [-1.761883, 54.984134],
-                    [-1.761619, 54.984472],
-                    [-1.761472, 54.984632],
-                    [-1.761296, 54.98476],
-                    [-1.759983, 54.985297],
-                    [-1.759846, 54.985447],
-                    [-1.759708, 54.985929],
-                    [-1.759705, 54.986037],
-                    [-1.759786, 54.986073],
-                    [-1.759793, 54.986659],
-                    [-1.759575, 54.986666],
-                    [-1.759563, 54.986815],
-                    [-1.759511, 54.986927],
-                    [-1.759423, 54.987056],
-                    [-1.759329, 54.987106],
-                    [-1.759293, 54.987241],
-                    [-1.759326, 54.98729],
-                    [-1.759314, 54.98761],
-                    [-1.759276, 54.987645],
-                    [-1.759278, 54.987678],
-                    [-1.759338, 54.987705],
-                    [-1.759198, 54.98791],
-                    [-1.759091, 54.988008],
-                    [-1.759082, 54.98821],
-                    [-1.758915, 54.988373],
-                    [-1.758918, 54.988537],
-                    [-1.758864, 54.988573],
-                    [-1.758836, 54.988741],
-                    [-1.758748, 54.988799],
-                    [-1.758748, 54.988861],
-                    [-1.758641, 54.988993],
-                    [-1.758628, 54.989062],
-                    [-1.75857, 54.989061],
-                    [-1.758481, 54.989117],
-                    [-1.758431, 54.989182],
-                    [-1.758447, 54.989207],
-                    [-1.758301, 54.989343],
-                    [-1.758315, 54.989382],
-                    [-1.758192, 54.989442],
-                    [-1.758163, 54.989546],
-                    [-1.758218, 54.98966],
-                    [-1.75808, 54.98989],
-                    [-1.757992, 54.990017],
-                    [-1.757763, 54.990084],
-                    [-1.756849, 54.990152],
-                    [-1.75678, 54.990179],
-                    [-1.756624, 54.990272],
-                    [-1.75644, 54.990595],
-                    [-1.756364, 54.990684],
-                    [-1.756273, 54.990729],
-                    [-1.755574, 54.990554],
-                    [-1.753436, 54.990263],
-                    [-1.753029, 54.990174],
-                    [-1.751486, 54.989718]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400025,
-              "name": "Battle of Newburn Ford 1640",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000025",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000025"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000025"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400026,
-          "name": "Battle of Newbury 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000026",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.342433, 51.402299],
-                    [-1.342086, 51.402182],
-                    [-1.341929, 51.402044],
-                    [-1.341394, 51.401751],
-                    [-1.340749, 51.401343],
-                    [-1.34078, 51.401191],
-                    [-1.340381, 51.40018],
-                    [-1.34014, 51.399671],
-                    [-1.339935, 51.399318],
-                    [-1.339886, 51.399126],
-                    [-1.338954, 51.3983],
-                    [-1.338383, 51.397907],
-                    [-1.337663, 51.397563],
-                    [-1.33728, 51.397427],
-                    [-1.337212, 51.397345],
-                    [-1.337127, 51.397304],
-                    [-1.336886, 51.397294],
-                    [-1.336257, 51.397158],
-                    [-1.339338, 51.397104],
-                    [-1.338942, 51.394994],
-                    [-1.340656, 51.394939],
-                    [-1.340398, 51.393417],
-                    [-1.340423, 51.3934],
-                    [-1.340285, 51.391756],
-                    [-1.34123, 51.391473],
-                    [-1.342883, 51.391436],
-                    [-1.342886, 51.391408],
-                    [-1.342979, 51.391406],
-                    [-1.342926, 51.391273],
-                    [-1.342847, 51.390706],
-                    [-1.342827, 51.390312],
-                    [-1.342901, 51.390179],
-                    [-1.342396, 51.390114],
-                    [-1.341161, 51.390181],
-                    [-1.340592, 51.390085],
-                    [-1.340759, 51.389881],
-                    [-1.340884, 51.389634],
-                    [-1.34105, 51.389438],
-                    [-1.341588, 51.388285],
-                    [-1.341656, 51.388296],
-                    [-1.341673, 51.388221],
-                    [-1.342275, 51.3878],
-                    [-1.342262, 51.387781],
-                    [-1.342723, 51.387417],
-                    [-1.342813, 51.38732],
-                    [-1.342839, 51.387194],
-                    [-1.343388, 51.386543],
-                    [-1.343472, 51.386488],
-                    [-1.343661, 51.386252],
-                    [-1.34393, 51.386046],
-                    [-1.344322, 51.385631],
-                    [-1.344716, 51.385356],
-                    [-1.345342, 51.384815],
-                    [-1.345506, 51.384737],
-                    [-1.345963, 51.384402],
-                    [-1.34605, 51.384286],
-                    [-1.346175, 51.383994],
-                    [-1.346695, 51.383393],
-                    [-1.346854, 51.383128],
-                    [-1.347062, 51.383047],
-                    [-1.347321, 51.382994],
-                    [-1.348179, 51.382858],
-                    [-1.348437, 51.382846],
-                    [-1.348465, 51.382729],
-                    [-1.348521, 51.382662],
-                    [-1.349077, 51.382546],
-                    [-1.350038, 51.382435],
-                    [-1.351062, 51.382433],
-                    [-1.350876, 51.38212],
-                    [-1.350897, 51.381319],
-                    [-1.350931, 51.381177],
-                    [-1.351115, 51.38079],
-                    [-1.351256, 51.380602],
-                    [-1.351478, 51.380448],
-                    [-1.351628, 51.380288],
-                    [-1.351782, 51.380045],
-                    [-1.352115, 51.37933],
-                    [-1.352486, 51.378242],
-                    [-1.352841, 51.37671],
-                    [-1.353012, 51.376307],
-                    [-1.353253, 51.375886],
-                    [-1.35338, 51.375478],
-                    [-1.353426, 51.375102],
-                    [-1.353502, 51.374913],
-                    [-1.353859, 51.374372],
-                    [-1.35456, 51.373609],
-                    [-1.354908, 51.37328],
-                    [-1.355304, 51.373107],
-                    [-1.356665, 51.372819],
-                    [-1.357279, 51.372648],
-                    [-1.358901, 51.372504],
-                    [-1.357961, 51.375211],
-                    [-1.356699, 51.376697],
-                    [-1.356708, 51.376852],
-                    [-1.356118, 51.376997],
-                    [-1.356263, 51.377198],
-                    [-1.35628, 51.377277],
-                    [-1.356182, 51.377852],
-                    [-1.356116, 51.377956],
-                    [-1.355653, 51.378174],
-                    [-1.355366, 51.378167],
-                    [-1.355279, 51.378916],
-                    [-1.355137, 51.379119],
-                    [-1.355277, 51.379239],
-                    [-1.355344, 51.379447],
-                    [-1.355354, 51.379652],
-                    [-1.354454, 51.380477],
-                    [-1.354555, 51.381267],
-                    [-1.354637, 51.38127],
-                    [-1.354748, 51.382088],
-                    [-1.354391, 51.382039],
-                    [-1.35394, 51.382026],
-                    [-1.353914, 51.38217],
-                    [-1.354049, 51.382301],
-                    [-1.354195, 51.382579],
-                    [-1.354215, 51.382641],
-                    [-1.354196, 51.382844],
-                    [-1.35424, 51.383008],
-                    [-1.354202, 51.383076],
-                    [-1.354133, 51.383104],
-                    [-1.354408, 51.383635],
-                    [-1.354894, 51.384144],
-                    [-1.355646, 51.384315],
-                    [-1.355644, 51.384348],
-                    [-1.355774, 51.38447],
-                    [-1.355867, 51.384795],
-                    [-1.355501, 51.38498],
-                    [-1.355349, 51.385131],
-                    [-1.355283, 51.385279],
-                    [-1.356375, 51.385745],
-                    [-1.356853, 51.386016],
-                    [-1.356918, 51.386083],
-                    [-1.356895, 51.386156],
-                    [-1.356629, 51.386396],
-                    [-1.356581, 51.386489],
-                    [-1.3566, 51.386664],
-                    [-1.357852, 51.389092],
-                    [-1.35824, 51.393894],
-                    [-1.358201, 51.394382],
-                    [-1.357679, 51.398047],
-                    [-1.355837, 51.399566],
-                    [-1.355586, 51.399586],
-                    [-1.355269, 51.399529],
-                    [-1.354807, 51.399376],
-                    [-1.354567, 51.399263],
-                    [-1.354224, 51.399181],
-                    [-1.353993, 51.399059],
-                    [-1.353721, 51.39901],
-                    [-1.353265, 51.399014],
-                    [-1.352001, 51.399242],
-                    [-1.351843, 51.399278],
-                    [-1.35146, 51.399503],
-                    [-1.351107, 51.399627],
-                    [-1.350043, 51.399915],
-                    [-1.349344, 51.399982],
-                    [-1.348957, 51.400166],
-                    [-1.348714, 51.400196],
-                    [-1.346634, 51.401064],
-                    [-1.346069, 51.401021],
-                    [-1.344829, 51.401212],
-                    [-1.344579, 51.401322],
-                    [-1.344411, 51.401473],
-                    [-1.344382, 51.401627],
-                    [-1.344255, 51.401937],
-                    [-1.344177, 51.402029],
-                    [-1.343686, 51.402268],
-                    [-1.343352, 51.402308],
-                    [-1.342433, 51.402299]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400026,
-              "name": "Battle of Newbury 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000026",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000026"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000026"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400027,
-          "name": "Battle of Northallerton 1138",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000027",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.439766, 54.382025],
-                    [-1.439742, 54.381827],
-                    [-1.439681, 54.381648],
-                    [-1.439695, 54.381214],
-                    [-1.439619, 54.380967],
-                    [-1.4396, 54.38055],
-                    [-1.43936, 54.378715],
-                    [-1.438086, 54.375],
-                    [-1.436992, 54.373456],
-                    [-1.437261, 54.373444],
-                    [-1.437707, 54.373317],
-                    [-1.438721, 54.372909],
-                    [-1.439625, 54.372616],
-                    [-1.44045, 54.372393],
-                    [-1.442558, 54.371705],
-                    [-1.44444, 54.371273],
-                    [-1.445467, 54.371099],
-                    [-1.44559, 54.371028],
-                    [-1.445722, 54.37088],
-                    [-1.445983, 54.370363],
-                    [-1.445969, 54.370281],
-                    [-1.448485, 54.370558],
-                    [-1.448137, 54.371311],
-                    [-1.447824, 54.372126],
-                    [-1.447576, 54.373311],
-                    [-1.4493, 54.37341],
-                    [-1.449256, 54.373624],
-                    [-1.449417, 54.375113],
-                    [-1.44987, 54.375319],
-                    [-1.450334, 54.375621],
-                    [-1.450728, 54.376026],
-                    [-1.449015, 54.376296],
-                    [-1.449488, 54.377109],
-                    [-1.44967, 54.377464],
-                    [-1.44965, 54.377494],
-                    [-1.449693, 54.37752],
-                    [-1.450031, 54.378177],
-                    [-1.450605, 54.378065],
-                    [-1.450572, 54.37827],
-                    [-1.450687, 54.378442],
-                    [-1.450678, 54.378645],
-                    [-1.450723, 54.378789],
-                    [-1.450735, 54.379029],
-                    [-1.450918, 54.379184],
-                    [-1.451169, 54.379193],
-                    [-1.451398, 54.379385],
-                    [-1.44726, 54.380653],
-                    [-1.446772, 54.38083],
-                    [-1.446697, 54.381744],
-                    [-1.44364, 54.382173],
-                    [-1.443509, 54.381948],
-                    [-1.44145, 54.382104],
-                    [-1.439825, 54.382275],
-                    [-1.439766, 54.382025]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400027,
-              "name": "Battle of Northallerton 1138",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000027",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000027"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000027"
-          }
-        },
-        {
-          "entry-date": "2015-04-23",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400028,
-          "name": "Battle of Northampton 1460",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000028",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.875484, 52.23026],
-                    [-0.875292, 52.230207],
-                    [-0.874779, 52.230229],
-                    [-0.874433, 52.23038],
-                    [-0.874005, 52.230644],
-                    [-0.87379, 52.230664],
-                    [-0.87354, 52.230617],
-                    [-0.872311, 52.230106],
-                    [-0.871493, 52.229726],
-                    [-0.871289, 52.229578],
-                    [-0.871122, 52.229186],
-                    [-0.870936, 52.228939],
-                    [-0.870792, 52.228595],
-                    [-0.870906, 52.22834],
-                    [-0.871013, 52.227896],
-                    [-0.871858, 52.227047],
-                    [-0.872552, 52.225963],
-                    [-0.872819, 52.225464],
-                    [-0.873986, 52.224019],
-                    [-0.874233, 52.223761],
-                    [-0.875384, 52.222761],
-                    [-0.876005, 52.22234],
-                    [-0.876698, 52.222106],
-                    [-0.877003, 52.221947],
-                    [-0.877103, 52.221818],
-                    [-0.877237, 52.221347],
-                    [-0.878602, 52.219748],
-                    [-0.878712, 52.219628],
-                    [-0.878961, 52.21945],
-                    [-0.879593, 52.219078],
-                    [-0.880123, 52.218819],
-                    [-0.881396, 52.218285],
-                    [-0.881769, 52.21815],
-                    [-0.883132, 52.217765],
-                    [-0.884069, 52.217548],
-                    [-0.887814, 52.217162],
-                    [-0.888927, 52.217087],
-                    [-0.889694, 52.216987],
-                    [-0.889699, 52.216965],
-                    [-0.889891, 52.216964],
-                    [-0.890465, 52.216871],
-                    [-0.89153, 52.2166],
-                    [-0.89233, 52.216527],
-                    [-0.894043, 52.215635],
-                    [-0.894286, 52.215529],
-                    [-0.8946, 52.215443],
-                    [-0.894937, 52.215439],
-                    [-0.895602, 52.215644],
-                    [-0.896175, 52.215686],
-                    [-0.896468, 52.215754],
-                    [-0.896729, 52.215866],
-                    [-0.897197, 52.216343],
-                    [-0.897603, 52.217223],
-                    [-0.897643, 52.217478],
-                    [-0.897664, 52.218572],
-                    [-0.897553, 52.219049],
-                    [-0.897299, 52.219635],
-                    [-0.89722, 52.220151],
-                    [-0.897134, 52.222114],
-                    [-0.897085, 52.222441],
-                    [-0.897056, 52.22355],
-                    [-0.896643, 52.22564],
-                    [-0.896607, 52.225961],
-                    [-0.896572, 52.225964],
-                    [-0.89657, 52.226431],
-                    [-0.896175, 52.22648],
-                    [-0.89517, 52.226715],
-                    [-0.893366, 52.226882],
-                    [-0.891853, 52.22683],
-                    [-0.89092, 52.226539],
-                    [-0.890504, 52.226497],
-                    [-0.89026, 52.226557],
-                    [-0.890034, 52.226688],
-                    [-0.889457, 52.226577],
-                    [-0.888879, 52.22657],
-                    [-0.887758, 52.226678],
-                    [-0.885074, 52.227022],
-                    [-0.885322, 52.227715],
-                    [-0.885608, 52.228763],
-                    [-0.883093, 52.228793],
-                    [-0.880643, 52.228572],
-                    [-0.880571, 52.230369],
-                    [-0.880198, 52.230417],
-                    [-0.879884, 52.230541],
-                    [-0.879489, 52.230984],
-                    [-0.879384, 52.231044],
-                    [-0.879146, 52.231097],
-                    [-0.878732, 52.231084],
-                    [-0.878423, 52.231035],
-                    [-0.877204, 52.230733],
-                    [-0.876304, 52.230625],
-                    [-0.876069, 52.230568],
-                    [-0.875824, 52.230413],
-                    [-0.875484, 52.23026]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2015-04-23",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400028,
-              "name": "Battle of Northampton 1460",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000028",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000028"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000028"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400029,
-          "name": "Battle of Otterburn 1388",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000029",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.1844, 55.245363],
-                    [-2.184145, 55.245346],
-                    [-2.184109, 55.245358],
-                    [-2.184126, 55.245432],
-                    [-2.183997, 55.24546],
-                    [-2.183877, 55.245401],
-                    [-2.183835, 55.245316],
-                    [-2.183762, 55.245288],
-                    [-2.183331, 55.245278],
-                    [-2.183189, 55.245242],
-                    [-2.182062, 55.245385],
-                    [-2.18171, 55.245184],
-                    [-2.181651, 55.245188],
-                    [-2.181436, 55.244974],
-                    [-2.181428, 55.244847],
-                    [-2.181257, 55.244624],
-                    [-2.180687, 55.244298],
-                    [-2.180426, 55.244063],
-                    [-2.180332, 55.243886],
-                    [-2.180368, 55.243529],
-                    [-2.18062, 55.242976],
-                    [-2.180891, 55.242632],
-                    [-2.181337, 55.242227],
-                    [-2.181404, 55.2421],
-                    [-2.181395, 55.241991],
-                    [-2.181145, 55.241575],
-                    [-2.181114, 55.241449],
-                    [-2.181183, 55.24123],
-                    [-2.181147, 55.24117],
-                    [-2.181022, 55.241039],
-                    [-2.181034, 55.241016],
-                    [-2.180958, 55.240973],
-                    [-2.18066, 55.240917],
-                    [-2.180475, 55.240999],
-                    [-2.180326, 55.24118],
-                    [-2.179812, 55.241438],
-                    [-2.179768, 55.241434],
-                    [-2.179525, 55.241339],
-                    [-2.179467, 55.241244],
-                    [-2.179465, 55.241157],
-                    [-2.179631, 55.2409],
-                    [-2.17951, 55.240748],
-                    [-2.179273, 55.240711],
-                    [-2.179072, 55.240763],
-                    [-2.178948, 55.240735],
-                    [-2.179014, 55.240622],
-                    [-2.179041, 55.24047],
-                    [-2.179106, 55.240403],
-                    [-2.179206, 55.239888],
-                    [-2.17924, 55.239488],
-                    [-2.179201, 55.239207],
-                    [-2.178996, 55.238856],
-                    [-2.178998, 55.238749],
-                    [-2.179041, 55.2387],
-                    [-2.179467, 55.238487],
-                    [-2.179806, 55.238381],
-                    [-2.180145, 55.238378],
-                    [-2.180572, 55.238461],
-                    [-2.180732, 55.238464],
-                    [-2.181061, 55.238391],
-                    [-2.181439, 55.238141],
-                    [-2.18148, 55.237997],
-                    [-2.18125, 55.237433],
-                    [-2.181161, 55.237323],
-                    [-2.18082, 55.2371],
-                    [-2.180482, 55.236997],
-                    [-2.180156, 55.23683],
-                    [-2.180053, 55.236743],
-                    [-2.180059, 55.23669],
-                    [-2.1801, 55.23666],
-                    [-2.180614, 55.236508],
-                    [-2.180127, 55.236187],
-                    [-2.179772, 55.235725],
-                    [-2.179269, 55.235378],
-                    [-2.179239, 55.235285],
-                    [-2.179246, 55.235104],
-                    [-2.179305, 55.23487],
-                    [-2.18058, 55.234252],
-                    [-2.181638, 55.234142],
-                    [-2.181642, 55.234113],
-                    [-2.182749, 55.234253],
-                    [-2.182977, 55.234458],
-                    [-2.182931, 55.234616],
-                    [-2.182949, 55.234676],
-                    [-2.184275, 55.234372],
-                    [-2.184134, 55.234261],
-                    [-2.183584, 55.23367],
-                    [-2.183527, 55.233365],
-                    [-2.183529, 55.232942],
-                    [-2.183659, 55.232584],
-                    [-2.183657, 55.232145],
-                    [-2.182896, 55.232063],
-                    [-2.182951, 55.231869],
-                    [-2.181949, 55.231762],
-                    [-2.181993, 55.231642],
-                    [-2.180771, 55.231522],
-                    [-2.180887, 55.231327],
-                    [-2.179447, 55.231053],
-                    [-2.179454, 55.230918],
-                    [-2.179768, 55.230481],
-                    [-2.179773, 55.230373],
-                    [-2.179872, 55.23025],
-                    [-2.180015, 55.230201],
-                    [-2.180274, 55.230178],
-                    [-2.180387, 55.230236],
-                    [-2.180225, 55.230317],
-                    [-2.180064, 55.230454],
-                    [-2.180052, 55.23052],
-                    [-2.180124, 55.23059],
-                    [-2.180434, 55.230643],
-                    [-2.180818, 55.230623],
-                    [-2.18095, 55.230424],
-                    [-2.181054, 55.230406],
-                    [-2.181201, 55.23046],
-                    [-2.181306, 55.230561],
-                    [-2.18149, 55.230668],
-                    [-2.181634, 55.23072],
-                    [-2.182228, 55.230821],
-                    [-2.185328, 55.230714],
-                    [-2.185391, 55.230648],
-                    [-2.185426, 55.230546],
-                    [-2.185664, 55.230425],
-                    [-2.185674, 55.230372],
-                    [-2.185784, 55.230322],
-                    [-2.186064, 55.230287],
-                    [-2.186221, 55.230526],
-                    [-2.186232, 55.230624],
-                    [-2.186387, 55.230917],
-                    [-2.186569, 55.231096],
-                    [-2.187009, 55.231353],
-                    [-2.187258, 55.231404],
-                    [-2.187689, 55.231584],
-                    [-2.188389, 55.231976],
-                    [-2.189247, 55.232276],
-                    [-2.19048, 55.232633],
-                    [-2.190989, 55.232821],
-                    [-2.191718, 55.23296],
-                    [-2.193384, 55.233078],
-                    [-2.194122, 55.233019],
-                    [-2.194732, 55.233126],
-                    [-2.194901, 55.233311],
-                    [-2.195247, 55.233573],
-                    [-2.195169, 55.233651],
-                    [-2.195167, 55.233729],
-                    [-2.195275, 55.233894],
-                    [-2.195395, 55.233985],
-                    [-2.195581, 55.234065],
-                    [-2.197089, 55.234464],
-                    [-2.197765, 55.234544],
-                    [-2.198181, 55.234551],
-                    [-2.198805, 55.234663],
-                    [-2.199139, 55.234685],
-                    [-2.200498, 55.234531],
-                    [-2.200973, 55.234409],
-                    [-2.201669, 55.234297],
-                    [-2.201989, 55.234291],
-                    [-2.202724, 55.234364],
-                    [-2.203103, 55.234435],
-                    [-2.203238, 55.234533],
-                    [-2.203343, 55.234679],
-                    [-2.203235, 55.235517],
-                    [-2.203245, 55.235609],
-                    [-2.203314, 55.235676],
-                    [-2.203209, 55.235881],
-                    [-2.203208, 55.235993],
-                    [-2.203243, 55.236036],
-                    [-2.203218, 55.236193],
-                    [-2.203752, 55.236715],
-                    [-2.20386, 55.236948],
-                    [-2.203953, 55.237031],
-                    [-2.203995, 55.237152],
-                    [-2.20399, 55.237399],
-                    [-2.204058, 55.237463],
-                    [-2.204049, 55.237568],
-                    [-2.204121, 55.237774],
-                    [-2.205106, 55.238446],
-                    [-2.205365, 55.238578],
-                    [-2.205394, 55.238583],
-                    [-2.205368, 55.238529],
-                    [-2.205425, 55.238528],
-                    [-2.205611, 55.23864],
-                    [-2.206108, 55.238773],
-                    [-2.206276, 55.238771],
-                    [-2.20622, 55.239208],
-                    [-2.206237, 55.239738],
-                    [-2.206068, 55.239852],
-                    [-2.206026, 55.240587],
-                    [-2.205885, 55.240612],
-                    [-2.206077, 55.240928],
-                    [-2.205675, 55.241004],
-                    [-2.205442, 55.241678],
-                    [-2.20472, 55.242172],
-                    [-2.204714, 55.242241],
-                    [-2.204492, 55.242502],
-                    [-2.204514, 55.242528],
-                    [-2.203628, 55.243768],
-                    [-2.202387, 55.243539],
-                    [-2.201787, 55.245],
-                    [-2.201582, 55.245857],
-                    [-2.201554, 55.246292],
-                    [-2.201504, 55.246449],
-                    [-2.201433, 55.246563],
-                    [-2.20116, 55.246823],
-                    [-2.200331, 55.247227],
-                    [-2.199542, 55.247343],
-                    [-2.197878, 55.247377],
-                    [-2.194788, 55.247269],
-                    [-2.191347, 55.247112],
-                    [-2.187481, 55.246778],
-                    [-2.186341, 55.246243],
-                    [-2.185708, 55.246039],
-                    [-2.184929, 55.245906],
-                    [-2.184901, 55.245787],
-                    [-2.184857, 55.245737],
-                    [-2.184669, 55.245578],
-                    [-2.184515, 55.24554],
-                    [-2.18452, 55.2454],
-                    [-2.1844, 55.245363]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400029,
-              "name": "Battle of Otterburn 1388",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000029",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000029"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000029"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400030,
-          "name": "Battle of Roundway Down 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000030",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.932544, 51.392798],
-                    [-1.933325, 51.391892],
-                    [-1.932429, 51.391814],
-                    [-1.931921, 51.391731],
-                    [-1.930757, 51.391462],
-                    [-1.93163, 51.38817],
-                    [-1.931773, 51.387776],
-                    [-1.933562, 51.385872],
-                    [-1.933924, 51.385387],
-                    [-1.934083, 51.385276],
-                    [-1.934168, 51.385162],
-                    [-1.934673, 51.385129],
-                    [-1.934937, 51.38508],
-                    [-1.935281, 51.384954],
-                    [-1.936643, 51.384281],
-                    [-1.936928, 51.384195],
-                    [-1.938241, 51.383937],
-                    [-1.938337, 51.383872],
-                    [-1.938996, 51.383901],
-                    [-1.939868, 51.384023],
-                    [-1.941379, 51.384097],
-                    [-1.942156, 51.384191],
-                    [-1.943517, 51.384231],
-                    [-1.943894, 51.384315],
-                    [-1.944248, 51.384336],
-                    [-1.94533, 51.384601],
-                    [-1.947717, 51.385408],
-                    [-1.957243, 51.376623],
-                    [-1.960489, 51.372926],
-                    [-1.980744, 51.37113],
-                    [-2.000617, 51.371445],
-                    [-2.003101, 51.372985],
-                    [-2.002556, 51.373376],
-                    [-2.002237, 51.373875],
-                    [-2.002589, 51.374591],
-                    [-2.002702, 51.374743],
-                    [-2.00275, 51.375011],
-                    [-2.002693, 51.375297],
-                    [-2.002718, 51.375472],
-                    [-2.002798, 51.375667],
-                    [-2.002984, 51.375945],
-                    [-2.002963, 51.376028],
-                    [-2.002858, 51.376136],
-                    [-2.0027, 51.376477],
-                    [-2.002838, 51.376886],
-                    [-2.002466, 51.377356],
-                    [-2.003065, 51.377537],
-                    [-2.003875, 51.377836],
-                    [-2.005202, 51.378463],
-                    [-2.005712, 51.378762],
-                    [-2.006018, 51.379009],
-                    [-2.006307, 51.379287],
-                    [-2.006554, 51.379626],
-                    [-2.006587, 51.379712],
-                    [-2.006617, 51.380336],
-                    [-2.006724, 51.380684],
-                    [-2.007388, 51.381893],
-                    [-2.007773, 51.382347],
-                    [-2.007809, 51.382437],
-                    [-2.007822, 51.383103],
-                    [-2.007417, 51.383532],
-                    [-2.007241, 51.383911],
-                    [-2.007239, 51.384047],
-                    [-2.00718, 51.38428],
-                    [-2.007765, 51.384476],
-                    [-2.009351, 51.384761],
-                    [-2.009882, 51.384955],
-                    [-2.0101, 51.385061],
-                    [-2.010641, 51.385418],
-                    [-2.011638, 51.386191],
-                    [-2.011781, 51.386189],
-                    [-2.011923, 51.386279],
-                    [-2.012448, 51.386506],
-                    [-2.012646, 51.386706],
-                    [-2.012985, 51.387279],
-                    [-2.013014, 51.387588],
-                    [-2.012836, 51.388304],
-                    [-2.012867, 51.38836],
-                    [-2.013051, 51.388478],
-                    [-2.013309, 51.388553],
-                    [-2.013273, 51.388645],
-                    [-2.013305, 51.388873],
-                    [-2.013275, 51.388968],
-                    [-2.013034, 51.389419],
-                    [-2.012822, 51.389724],
-                    [-2.012783, 51.389841],
-                    [-2.012765, 51.390983],
-                    [-2.012693, 51.391332],
-                    [-2.012375, 51.391417],
-                    [-2.011452, 51.391291],
-                    [-2.011001, 51.391158],
-                    [-2.010281, 51.3908],
-                    [-2.009578, 51.39052],
-                    [-2.009593, 51.390505],
-                    [-2.009258, 51.390345],
-                    [-2.008669, 51.389969],
-                    [-2.008117, 51.389519],
-                    [-2.007338, 51.389004],
-                    [-2.006825, 51.388794],
-                    [-2.005894, 51.388548],
-                    [-2.00233, 51.38779],
-                    [-1.999709, 51.387403],
-                    [-1.996295, 51.387105],
-                    [-1.995271, 51.387145],
-                    [-1.993875, 51.387109],
-                    [-1.99359, 51.387205],
-                    [-1.99202, 51.387968],
-                    [-1.99062, 51.388551],
-                    [-1.988785, 51.389523],
-                    [-1.988241, 51.389187],
-                    [-1.987006, 51.389319],
-                    [-1.986743, 51.389485],
-                    [-1.986846, 51.389965],
-                    [-1.986121, 51.390124],
-                    [-1.984142, 51.390683],
-                    [-1.981251, 51.391924],
-                    [-1.980343, 51.39224],
-                    [-1.979623, 51.39231],
-                    [-1.978581, 51.392296],
-                    [-1.974778, 51.392117],
-                    [-1.971517, 51.392248],
-                    [-1.971103, 51.392268],
-                    [-1.971099, 51.392311],
-                    [-1.970449, 51.392333],
-                    [-1.966778, 51.392897],
-                    [-1.961426, 51.393159],
-                    [-1.961294, 51.3935],
-                    [-1.961076, 51.393728],
-                    [-1.959836, 51.393895],
-                    [-1.958566, 51.394148],
-                    [-1.952797, 51.395626],
-                    [-1.951049, 51.396157],
-                    [-1.94868, 51.396783],
-                    [-1.947477, 51.397212],
-                    [-1.945092, 51.397872],
-                    [-1.943863, 51.398328],
-                    [-1.937411, 51.395419],
-                    [-1.937127, 51.395262],
-                    [-1.93738, 51.395055],
-                    [-1.932544, 51.392798]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400030,
-              "name": "Battle of Roundway Down 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000030",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000030"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000030"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400031,
-          "name": "Battle of Rowton Heath 1645",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000031",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.813265, 53.164022],
-                    [-2.81325, 53.163959],
-                    [-2.813016, 53.163927],
-                    [-2.812578, 53.163796],
-                    [-2.811555, 53.163339],
-                    [-2.810848, 53.163066],
-                    [-2.810935, 53.162911],
-                    [-2.810711, 53.162801],
-                    [-2.810534, 53.162609],
-                    [-2.810097, 53.161792],
-                    [-2.809743, 53.159545],
-                    [-2.809502, 53.158827],
-                    [-2.809224, 53.158377],
-                    [-2.809112, 53.158023],
-                    [-2.808449, 53.156821],
-                    [-2.809569, 53.156471],
-                    [-2.810008, 53.156388],
-                    [-2.810078, 53.156411],
-                    [-2.810579, 53.156288],
-                    [-2.811501, 53.156209],
-                    [-2.811897, 53.156229],
-                    [-2.811872, 53.156159],
-                    [-2.811908, 53.15602],
-                    [-2.811967, 53.156021],
-                    [-2.812274, 53.155664],
-                    [-2.812829, 53.155638],
-                    [-2.813392, 53.15567],
-                    [-2.813609, 53.155768],
-                    [-2.813671, 53.155848],
-                    [-2.813956, 53.155912],
-                    [-2.815031, 53.156729],
-                    [-2.815946, 53.157504],
-                    [-2.817085, 53.158568],
-                    [-2.817301, 53.158827],
-                    [-2.817917, 53.159356],
-                    [-2.81816, 53.159648],
-                    [-2.818152, 53.159693],
-                    [-2.818067, 53.159746],
-                    [-2.818296, 53.159928],
-                    [-2.818528, 53.160205],
-                    [-2.819839, 53.160986],
-                    [-2.820657, 53.161355],
-                    [-2.820795, 53.161467],
-                    [-2.821203, 53.161596],
-                    [-2.822417, 53.162146],
-                    [-2.821594, 53.163744],
-                    [-2.820563, 53.165491],
-                    [-2.819219, 53.165959],
-                    [-2.818281, 53.166228],
-                    [-2.817207, 53.166614],
-                    [-2.815937, 53.166899],
-                    [-2.815647, 53.166418],
-                    [-2.815353, 53.166102],
-                    [-2.815161, 53.165958],
-                    [-2.814622, 53.16572],
-                    [-2.814007, 53.165613],
-                    [-2.813987, 53.165589],
-                    [-2.814349, 53.165456],
-                    [-2.813265, 53.164022]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400031,
-              "name": "Battle of Rowton Heath 1645",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000031",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000031"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000031"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400032,
-          "name": "Battle of Sedgemoor 1685",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000032",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.921606, 51.12715],
-                    [-2.916001, 51.120142],
-                    [-2.91589, 51.119998],
-                    [-2.915613, 51.119405],
-                    [-2.915221, 51.118944],
-                    [-2.914746, 51.118624],
-                    [-2.914114, 51.118401],
-                    [-2.915246, 51.116857],
-                    [-2.915522, 51.116917],
-                    [-2.917846, 51.11479],
-                    [-2.917583, 51.113817],
-                    [-2.917227, 51.113821],
-                    [-2.917125, 51.113685],
-                    [-2.91725, 51.113099],
-                    [-2.917335, 51.11291],
-                    [-2.917317, 51.112747],
-                    [-2.917504, 51.111933],
-                    [-2.917713, 51.112052],
-                    [-2.918019, 51.112161],
-                    [-2.918547, 51.112159],
-                    [-2.919164, 51.111974],
-                    [-2.920154, 51.111853],
-                    [-2.920242, 51.111787],
-                    [-2.921265, 51.112002],
-                    [-2.92127, 51.112091],
-                    [-2.921471, 51.112169],
-                    [-2.923596, 51.112336],
-                    [-2.924007, 51.112431],
-                    [-2.924138, 51.112057],
-                    [-2.924993, 51.11214],
-                    [-2.925056, 51.112037],
-                    [-2.92527, 51.112083],
-                    [-2.925249, 51.112195],
-                    [-2.926435, 51.112466],
-                    [-2.927245, 51.112114],
-                    [-2.930241, 51.112575],
-                    [-2.93038, 51.112076],
-                    [-2.930967, 51.111966],
-                    [-2.933081, 51.112613],
-                    [-2.933347, 51.11169],
-                    [-2.933445, 51.111661],
-                    [-2.933659, 51.111641],
-                    [-2.935559, 51.11189],
-                    [-2.935539, 51.112037],
-                    [-2.935209, 51.113114],
-                    [-2.94147, 51.113373],
-                    [-2.941055, 51.114731],
-                    [-2.943267, 51.114805],
-                    [-2.942632, 51.117043],
-                    [-2.942591, 51.117055],
-                    [-2.942668, 51.117089],
-                    [-2.942659, 51.117142],
-                    [-2.944848, 51.11719],
-                    [-2.94401, 51.119471],
-                    [-2.943915, 51.120756],
-                    [-2.942527, 51.120522],
-                    [-2.941712, 51.120312],
-                    [-2.941094, 51.120262],
-                    [-2.940699, 51.120178],
-                    [-2.939107, 51.11997],
-                    [-2.938954, 51.11994],
-                    [-2.938851, 51.119871],
-                    [-2.938738, 51.11987],
-                    [-2.937959, 51.12012],
-                    [-2.937266, 51.120214],
-                    [-2.936912, 51.120313],
-                    [-2.936649, 51.120461],
-                    [-2.936188, 51.120546],
-                    [-2.935549, 51.120748],
-                    [-2.934703, 51.121137],
-                    [-2.933712, 51.121831],
-                    [-2.933451, 51.122063],
-                    [-2.932551, 51.122589],
-                    [-2.932164, 51.123044],
-                    [-2.931887, 51.123254],
-                    [-2.931515, 51.123454],
-                    [-2.93133, 51.123798],
-                    [-2.931218, 51.123867],
-                    [-2.930619, 51.124427],
-                    [-2.93009, 51.124861],
-                    [-2.929369, 51.125199],
-                    [-2.928411, 51.125947],
-                    [-2.927772, 51.126333],
-                    [-2.927415, 51.126619],
-                    [-2.927069, 51.126778],
-                    [-2.92694, 51.126817],
-                    [-2.926533, 51.126841],
-                    [-2.924287, 51.127136],
-                    [-2.924242, 51.127068],
-                    [-2.921789, 51.12738],
-                    [-2.921606, 51.12715]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400032,
-              "name": "Battle of Sedgemoor 1685",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000032",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000032"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000032"
-          }
-        },
-        {
-          "entry-date": "2015-09-09",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400033,
-          "name": "Battle of Shrewsbury 1403",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000033",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.734909, 52.754808],
-                    [-2.734553, 52.754726],
-                    [-2.734067, 52.754719],
-                    [-2.733842, 52.754666],
-                    [-2.733684, 52.754768],
-                    [-2.731896, 52.754633],
-                    [-2.730462, 52.754608],
-                    [-2.729207, 52.754513],
-                    [-2.728206, 52.754596],
-                    [-2.726143, 52.753848],
-                    [-2.72391, 52.753115],
-                    [-2.721716, 52.752701],
-                    [-2.720202, 52.752457],
-                    [-2.720216, 52.752428],
-                    [-2.720096, 52.752357],
-                    [-2.719928, 52.752297],
-                    [-2.719635, 52.752274],
-                    [-2.719479, 52.752285],
-                    [-2.719461, 52.752327],
-                    [-2.719232, 52.752318],
-                    [-2.718574, 52.752206],
-                    [-2.718576, 52.752169],
-                    [-2.718747, 52.751687],
-                    [-2.718966, 52.751247],
-                    [-2.720587, 52.748553],
-                    [-2.722697, 52.745736],
-                    [-2.722687, 52.74571],
-                    [-2.722866, 52.745462],
-                    [-2.723455, 52.744714],
-                    [-2.724413, 52.744958],
-                    [-2.724467, 52.744948],
-                    [-2.724656, 52.745046],
-                    [-2.726686, 52.745503],
-                    [-2.728392, 52.745613],
-                    [-2.729771, 52.745804],
-                    [-2.730234, 52.745919],
-                    [-2.730803, 52.746108],
-                    [-2.731783, 52.746269],
-                    [-2.732233, 52.746433],
-                    [-2.733, 52.746656],
-                    [-2.733152, 52.74642],
-                    [-2.733435, 52.746142],
-                    [-2.733591, 52.746209],
-                    [-2.733678, 52.746294],
-                    [-2.733879, 52.746346],
-                    [-2.73407, 52.746597],
-                    [-2.734208, 52.746685],
-                    [-2.734448, 52.746744],
-                    [-2.734544, 52.746867],
-                    [-2.734667, 52.746943],
-                    [-2.734952, 52.746997],
-                    [-2.735319, 52.746989],
-                    [-2.736408, 52.747522],
-                    [-2.737255, 52.748047],
-                    [-2.737456, 52.748139],
-                    [-2.737673, 52.748321],
-                    [-2.737835, 52.748594],
-                    [-2.738067, 52.749304],
-                    [-2.738127, 52.749412],
-                    [-2.738189, 52.749443],
-                    [-2.738165, 52.749589],
-                    [-2.738202, 52.749707],
-                    [-2.740148, 52.750804],
-                    [-2.740152, 52.750976],
-                    [-2.740106, 52.751009],
-                    [-2.740045, 52.75099],
-                    [-2.739378, 52.751964],
-                    [-2.739164, 52.752729],
-                    [-2.738391, 52.752733],
-                    [-2.737924, 52.752852],
-                    [-2.737766, 52.753265],
-                    [-2.737699, 52.753342],
-                    [-2.737655, 52.753508],
-                    [-2.737598, 52.753596],
-                    [-2.736391, 52.754983],
-                    [-2.735767, 52.75495],
-                    [-2.734909, 52.754808]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2015-09-09",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400033,
-              "name": "Battle of Shrewsbury 1403",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000033",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000033"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000033"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400034,
-          "name": "Battle of Solway Moss 1542",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000034",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.975101, 55.007872],
-                    [-2.974985, 55.007747],
-                    [-2.974618, 55.007867],
-                    [-2.974417, 55.007716],
-                    [-2.974316, 55.007752],
-                    [-2.973948, 55.007483],
-                    [-2.973249, 55.007772],
-                    [-2.973036, 55.00755],
-                    [-2.97252, 55.00766],
-                    [-2.972319, 55.007295],
-                    [-2.972058, 55.00696],
-                    [-2.971711, 55.006224],
-                    [-2.970935, 55.006426],
-                    [-2.970691, 55.006391],
-                    [-2.968034, 55.007194],
-                    [-2.966696, 55.006696],
-                    [-2.966312, 55.007008],
-                    [-2.966419, 55.007057],
-                    [-2.96614, 55.007225],
-                    [-2.96588, 55.007283],
-                    [-2.965583, 55.007576],
-                    [-2.963004, 55.007515],
-                    [-2.961785, 55.007572],
-                    [-2.958356, 55.00781],
-                    [-2.957184, 55.007913],
-                    [-2.953833, 55.008501],
-                    [-2.953638, 55.008486],
-                    [-2.953483, 55.008312],
-                    [-2.952002, 55.006101],
-                    [-2.951833, 55.006005],
-                    [-2.951611, 55.00596],
-                    [-2.951489, 55.005504],
-                    [-2.95087, 55.003684],
-                    [-2.948903, 54.999752],
-                    [-2.947444, 54.996085],
-                    [-2.946784, 54.994822],
-                    [-2.946662, 54.994732],
-                    [-2.946495, 54.994398],
-                    [-2.946302, 54.994174],
-                    [-2.94612, 54.993867],
-                    [-2.945844, 54.993585],
-                    [-2.945404, 54.992915],
-                    [-2.945361, 54.992769],
-                    [-2.945439, 54.99273],
-                    [-2.954268, 54.990703],
-                    [-2.955658, 54.989966],
-                    [-2.957176, 54.98905],
-                    [-2.958948, 54.989014],
-                    [-2.959116, 54.989806],
-                    [-2.959861, 54.990611],
-                    [-2.959959, 54.990758],
-                    [-2.960018, 54.991019],
-                    [-2.960175, 54.9912],
-                    [-2.96179, 54.991925],
-                    [-2.964577, 54.992697],
-                    [-2.965292, 54.992923],
-                    [-2.965757, 54.99243],
-                    [-2.965816, 54.992436],
-                    [-2.965987, 54.992346],
-                    [-2.966105, 54.992341],
-                    [-2.966703, 54.992481],
-                    [-2.966931, 54.992484],
-                    [-2.967755, 54.992231],
-                    [-2.968056, 54.992182],
-                    [-2.968467, 54.992179],
-                    [-2.969334, 54.992017],
-                    [-2.969951, 54.992065],
-                    [-2.97078, 54.992218],
-                    [-2.970964, 54.99222],
-                    [-2.971137, 54.992144],
-                    [-2.971139, 54.992203],
-                    [-2.972087, 54.993597],
-                    [-2.974436, 54.993309],
-                    [-2.974491, 54.993793],
-                    [-2.976525, 54.993588],
-                    [-2.977335, 54.995019],
-                    [-2.978396, 54.998176],
-                    [-2.978448, 54.998265],
-                    [-2.981831, 54.997885],
-                    [-2.982132, 54.998734],
-                    [-2.982119, 54.998963],
-                    [-2.982145, 54.999142],
-                    [-2.982261, 54.999562],
-                    [-2.982228, 54.999867],
-                    [-2.982297, 55.000247],
-                    [-2.98238, 55.000461],
-                    [-2.982808, 55.001271],
-                    [-2.982938, 55.001711],
-                    [-2.982984, 55.002225],
-                    [-2.984499, 55.002574],
-                    [-2.983964, 55.002827],
-                    [-2.983782, 55.002882],
-                    [-2.982504, 55.003501],
-                    [-2.982165, 55.003712],
-                    [-2.982109, 55.003787],
-                    [-2.981202, 55.004241],
-                    [-2.981106, 55.004253],
-                    [-2.980844, 55.004384],
-                    [-2.980811, 55.004459],
-                    [-2.980371, 55.00475],
-                    [-2.98027, 55.004881],
-                    [-2.979981, 55.005049],
-                    [-2.979703, 55.005276],
-                    [-2.979009, 55.006113],
-                    [-2.978265, 55.006787],
-                    [-2.977639, 55.007227],
-                    [-2.977256, 55.007669],
-                    [-2.976411, 55.008321],
-                    [-2.97632, 55.008326],
-                    [-2.975791, 55.008771],
-                    [-2.975406, 55.008349],
-                    [-2.975101, 55.007872]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400034,
-              "name": "Battle of Solway Moss 1542",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000034",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000034"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000034"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400035,
-          "name": "Battle of Stamford Bridge 1066",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000035",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.89318, 53.990546],
-                    [-0.893707, 53.989943],
-                    [-0.89414, 53.989771],
-                    [-0.894308, 53.989665],
-                    [-0.894537, 53.989387],
-                    [-0.895334, 53.988585],
-                    [-0.895623, 53.988169],
-                    [-0.897487, 53.988401],
-                    [-0.897629, 53.987799],
-                    [-0.89796, 53.987029],
-                    [-0.898743, 53.985473],
-                    [-0.898934, 53.984894],
-                    [-0.899232, 53.9842],
-                    [-0.899025, 53.983995],
-                    [-0.898848, 53.983896],
-                    [-0.899042, 53.983788],
-                    [-0.899775, 53.983869],
-                    [-0.901043, 53.983595],
-                    [-0.901108, 53.983555],
-                    [-0.901403, 53.983491],
-                    [-0.903126, 53.982974],
-                    [-0.904068, 53.982855],
-                    [-0.905317, 53.982792],
-                    [-0.905683, 53.982717],
-                    [-0.90651, 53.982468],
-                    [-0.907051, 53.982403],
-                    [-0.907743, 53.982442],
-                    [-0.908094, 53.982429],
-                    [-0.908826, 53.982299],
-                    [-0.910169, 53.981966],
-                    [-0.910598, 53.981921],
-                    [-0.911519, 53.981922],
-                    [-0.912266, 53.981988],
-                    [-0.912335, 53.982253],
-                    [-0.912755, 53.983165],
-                    [-0.913029, 53.983544],
-                    [-0.911782, 53.984195],
-                    [-0.910499, 53.984628],
-                    [-0.910588, 53.984984],
-                    [-0.910689, 53.985182],
-                    [-0.91107, 53.985121],
-                    [-0.911295, 53.985743],
-                    [-0.910893, 53.985823],
-                    [-0.910438, 53.985982],
-                    [-0.910126, 53.986193],
-                    [-0.909819, 53.986457],
-                    [-0.909732, 53.986591],
-                    [-0.910094, 53.986704],
-                    [-0.910161, 53.986924],
-                    [-0.910123, 53.987636],
-                    [-0.909562, 53.987782],
-                    [-0.907471, 53.988188],
-                    [-0.906733, 53.988388],
-                    [-0.907539, 53.989494],
-                    [-0.907453, 53.989709],
-                    [-0.906457, 53.990227],
-                    [-0.906219, 53.990281],
-                    [-0.905631, 53.9901],
-                    [-0.904948, 53.989969],
-                    [-0.903111, 53.991481],
-                    [-0.899968, 53.990247],
-                    [-0.896318, 53.991007],
-                    [-0.893976, 53.992373],
-                    [-0.89344, 53.992195],
-                    [-0.892076, 53.991822],
-                    [-0.89318, 53.990546]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400035,
-              "name": "Battle of Stamford Bridge 1066",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000035",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000035"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000035"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400036,
-          "name": "Battle of Stoke (Field) 1487",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000036",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.887919, 53.042763],
-                    [-0.887747, 53.042498],
-                    [-0.887294, 53.042522],
-                    [-0.886928, 53.042427],
-                    [-0.886902, 53.042369],
-                    [-0.886507, 53.042069],
-                    [-0.886186, 53.042149],
-                    [-0.885778, 53.041886],
-                    [-0.88527, 53.041682],
-                    [-0.884562, 53.041466],
-                    [-0.883388, 53.041281],
-                    [-0.882964, 53.041182],
-                    [-0.882525, 53.040993],
-                    [-0.881064, 53.040197],
-                    [-0.880648, 53.040053],
-                    [-0.878486, 53.039451],
-                    [-0.877512, 53.039128],
-                    [-0.877321, 53.039033],
-                    [-0.877157, 53.038902],
-                    [-0.877574, 53.038348],
-                    [-0.878034, 53.037974],
-                    [-0.8785, 53.037427],
-                    [-0.879743, 53.035486],
-                    [-0.879919, 53.035323],
-                    [-0.880198, 53.035114],
-                    [-0.880937, 53.034714],
-                    [-0.881646, 53.034068],
-                    [-0.881876, 53.033898],
-                    [-0.883746, 53.032688],
-                    [-0.885331, 53.031109],
-                    [-0.888533, 53.028575],
-                    [-0.888999, 53.028127],
-                    [-0.889015, 53.028073],
-                    [-0.889844, 53.027316],
-                    [-0.891265, 53.025829],
-                    [-0.891627, 53.025536],
-                    [-0.893095, 53.024202],
-                    [-0.8938, 53.023644],
-                    [-0.894256, 53.023954],
-                    [-0.897574, 53.025683],
-                    [-0.89808, 53.026006],
-                    [-0.899295, 53.026612],
-                    [-0.89956, 53.026383],
-                    [-0.899886, 53.026173],
-                    [-0.899918, 53.026007],
-                    [-0.900147, 53.025686],
-                    [-0.900349, 53.025485],
-                    [-0.900573, 53.02534],
-                    [-0.901418, 53.02495],
-                    [-0.903236, 53.024398],
-                    [-0.903571, 53.024951],
-                    [-0.903868, 53.02532],
-                    [-0.905885, 53.027271],
-                    [-0.906485, 53.027664],
-                    [-0.90803, 53.028846],
-                    [-0.909624, 53.029834],
-                    [-0.910001, 53.030211],
-                    [-0.910082, 53.030389],
-                    [-0.910137, 53.031077],
-                    [-0.910457, 53.031717],
-                    [-0.910382, 53.032377],
-                    [-0.910165, 53.033052],
-                    [-0.910013, 53.033366],
-                    [-0.909428, 53.033658],
-                    [-0.909339, 53.033757],
-                    [-0.909287, 53.033936],
-                    [-0.909375, 53.034258],
-                    [-0.909823, 53.034744],
-                    [-0.910209, 53.035104],
-                    [-0.910338, 53.035424],
-                    [-0.910446, 53.035569],
-                    [-0.9101, 53.035667],
-                    [-0.909661, 53.035715],
-                    [-0.909133, 53.035885],
-                    [-0.908289, 53.035848],
-                    [-0.9059, 53.035927],
-                    [-0.905048, 53.036149],
-                    [-0.904226, 53.036479],
-                    [-0.904007, 53.036627],
-                    [-0.903789, 53.036836],
-                    [-0.902559, 53.037408],
-                    [-0.902031, 53.03784],
-                    [-0.901781, 53.03797],
-                    [-0.901327, 53.038439],
-                    [-0.901192, 53.038494],
-                    [-0.90103, 53.038634],
-                    [-0.897617, 53.039539],
-                    [-0.89471, 53.040478],
-                    [-0.893926, 53.040866],
-                    [-0.892508, 53.041478],
-                    [-0.892247, 53.041676],
-                    [-0.891552, 53.041981],
-                    [-0.889585, 53.042466],
-                    [-0.887919, 53.042763]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400036,
-              "name": "Battle of Stoke (Field) 1487",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000036",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000036"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000036"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400037,
-          "name": "Battle of Stow (-on-the-Wold) 1646",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000037",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.733174, 51.958391],
-                    [-1.731622, 51.957922],
-                    [-1.731332, 51.957871],
-                    [-1.730975, 51.957707],
-                    [-1.730491, 51.957542],
-                    [-1.730335, 51.957437],
-                    [-1.730027, 51.957397],
-                    [-1.729703, 51.957402],
-                    [-1.729405, 51.957449],
-                    [-1.727942, 51.957851],
-                    [-1.727662, 51.956669],
-                    [-1.72706, 51.955369],
-                    [-1.726827, 51.955364],
-                    [-1.725806, 51.955233],
-                    [-1.724594, 51.955327],
-                    [-1.721559, 51.950951],
-                    [-1.722038, 51.949788],
-                    [-1.722329, 51.949533],
-                    [-1.723124, 51.949148],
-                    [-1.723905, 51.948533],
-                    [-1.724107, 51.948404],
-                    [-1.724441, 51.948253],
-                    [-1.725938, 51.948177],
-                    [-1.729879, 51.947599],
-                    [-1.730076, 51.947543],
-                    [-1.733428, 51.946183],
-                    [-1.733835, 51.945968],
-                    [-1.736022, 51.947359],
-                    [-1.73661, 51.94771],
-                    [-1.736657, 51.947712],
-                    [-1.739536, 51.949598],
-                    [-1.739594, 51.949739],
-                    [-1.739862, 51.949956],
-                    [-1.741104, 51.950782],
-                    [-1.74174, 51.951136],
-                    [-1.741481, 51.951508],
-                    [-1.741063, 51.951973],
-                    [-1.740214, 51.953179],
-                    [-1.739908, 51.953522],
-                    [-1.739264, 51.954379],
-                    [-1.742972, 51.955081],
-                    [-1.743156, 51.955529],
-                    [-1.743143, 51.955568],
-                    [-1.742512, 51.956299],
-                    [-1.742329, 51.956414],
-                    [-1.742226, 51.956828],
-                    [-1.741923, 51.957261],
-                    [-1.741485, 51.957524],
-                    [-1.741193, 51.957634],
-                    [-1.740726, 51.957736],
-                    [-1.739875, 51.958033],
-                    [-1.738078, 51.958481],
-                    [-1.737882, 51.958434],
-                    [-1.735731, 51.958663],
-                    [-1.734739, 51.958374],
-                    [-1.733858, 51.958687],
-                    [-1.733174, 51.958391]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400037,
-              "name": "Battle of Stow (-on-the-Wold) 1646",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000037",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000037"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000037"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400038,
-          "name": "Battle of Stratton 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000038",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-4.515457, 50.839774],
-                    [-4.514902, 50.839319],
-                    [-4.514577, 50.838951],
-                    [-4.514249, 50.839114],
-                    [-4.513699, 50.839309],
-                    [-4.513634, 50.839269],
-                    [-4.513609, 50.83917],
-                    [-4.513512, 50.8391],
-                    [-4.51347, 50.838968],
-                    [-4.51336, 50.838867],
-                    [-4.513503, 50.838764],
-                    [-4.513437, 50.838601],
-                    [-4.513494, 50.838547],
-                    [-4.513516, 50.838447],
-                    [-4.513762, 50.838341],
-                    [-4.51395, 50.83831],
-                    [-4.514381, 50.838038],
-                    [-4.514183, 50.837966],
-                    [-4.514369, 50.83792],
-                    [-4.514322, 50.837746],
-                    [-4.514262, 50.837691],
-                    [-4.514289, 50.837643],
-                    [-4.514418, 50.837619],
-                    [-4.514581, 50.837707],
-                    [-4.514719, 50.837728],
-                    [-4.514834, 50.837657],
-                    [-4.514802, 50.837452],
-                    [-4.514941, 50.837326],
-                    [-4.514944, 50.837159],
-                    [-4.514385, 50.836952],
-                    [-4.51459, 50.836747],
-                    [-4.515713, 50.835383],
-                    [-4.515881, 50.835227],
-                    [-4.516564, 50.834417],
-                    [-4.516785, 50.833999],
-                    [-4.517162, 50.832539],
-                    [-4.517211, 50.832562],
-                    [-4.517666, 50.833161],
-                    [-4.517988, 50.833031],
-                    [-4.518177, 50.833011],
-                    [-4.51993, 50.832545],
-                    [-4.519972, 50.832504],
-                    [-4.519936, 50.832384],
-                    [-4.519629, 50.831976],
-                    [-4.519356, 50.83129],
-                    [-4.519371, 50.831191],
-                    [-4.520096, 50.831325],
-                    [-4.520229, 50.831385],
-                    [-4.520338, 50.831356],
-                    [-4.520632, 50.831403],
-                    [-4.52068, 50.83134],
-                    [-4.521566, 50.831142],
-                    [-4.523056, 50.831143],
-                    [-4.524203, 50.830959],
-                    [-4.524927, 50.833686],
-                    [-4.524894, 50.833878],
-                    [-4.524957, 50.833935],
-                    [-4.525411, 50.834726],
-                    [-4.525718, 50.834638],
-                    [-4.525849, 50.834693],
-                    [-4.525892, 50.834819],
-                    [-4.526047, 50.83501],
-                    [-4.526304, 50.835468],
-                    [-4.526429, 50.83596],
-                    [-4.526425, 50.836141],
-                    [-4.526351, 50.83636],
-                    [-4.526182, 50.83639],
-                    [-4.526229, 50.836995],
-                    [-4.526263, 50.837142],
-                    [-4.526304, 50.83719],
-                    [-4.526067, 50.837755],
-                    [-4.526039, 50.837936],
-                    [-4.5261, 50.83846],
-                    [-4.526346, 50.839232],
-                    [-4.525956, 50.839284],
-                    [-4.525349, 50.839296],
-                    [-4.52496, 50.839204],
-                    [-4.524471, 50.839189],
-                    [-4.523547, 50.839037],
-                    [-4.523221, 50.839531],
-                    [-4.522952, 50.839536],
-                    [-4.52249, 50.839648],
-                    [-4.522277, 50.839484],
-                    [-4.521893, 50.83933],
-                    [-4.521735, 50.83935],
-                    [-4.521527, 50.839458],
-                    [-4.521169, 50.839465],
-                    [-4.52085, 50.839618],
-                    [-4.520795, 50.839774],
-                    [-4.520749, 50.83978],
-                    [-4.520752, 50.84002],
-                    [-4.520673, 50.840126],
-                    [-4.520337, 50.84022],
-                    [-4.516028, 50.840023],
-                    [-4.51557, 50.839857],
-                    [-4.515457, 50.839774]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400038,
-              "name": "Battle of Stratton 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000038",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000038"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000038"
-          }
-        },
-        {
-          "entry-date": "2017-03-30",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400039,
-          "name": "Battle of Tewkesbury 1471",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000039",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.160907, 51.988405],
-                    [-2.16052, 51.988036],
-                    [-2.160209, 51.987655],
-                    [-2.159678, 51.987159],
-                    [-2.159592, 51.987102],
-                    [-2.159305, 51.98707],
-                    [-2.15907, 51.987082],
-                    [-2.158596, 51.986402],
-                    [-2.158633, 51.986256],
-                    [-2.158588, 51.986165],
-                    [-2.158012, 51.98547],
-                    [-2.157853, 51.9851],
-                    [-2.15781, 51.984867],
-                    [-2.159033, 51.984682],
-                    [-2.159336, 51.984892],
-                    [-2.159782, 51.98479],
-                    [-2.160364, 51.985679],
-                    [-2.161439, 51.986363],
-                    [-2.161474, 51.986341],
-                    [-2.161656, 51.986369],
-                    [-2.162423, 51.986122],
-                    [-2.162076, 51.985893],
-                    [-2.163, 51.985368],
-                    [-2.16344, 51.985237],
-                    [-2.164638, 51.984538],
-                    [-2.164315, 51.984392],
-                    [-2.163365, 51.983822],
-                    [-2.162457, 51.983385],
-                    [-2.161818, 51.983032],
-                    [-2.161273, 51.982815],
-                    [-2.160925, 51.983275],
-                    [-2.160719, 51.983643],
-                    [-2.160445, 51.98434],
-                    [-2.160068, 51.984621],
-                    [-2.157368, 51.982955],
-                    [-2.156571, 51.9825],
-                    [-2.155977, 51.982608],
-                    [-2.15588, 51.9824],
-                    [-2.155417, 51.982133],
-                    [-2.154926, 51.981572],
-                    [-2.15333, 51.981549],
-                    [-2.153243, 51.98116],
-                    [-2.153152, 51.98095],
-                    [-2.15325, 51.980599],
-                    [-2.150234, 51.981033],
-                    [-2.150426, 51.981461],
-                    [-2.150414, 51.981505],
-                    [-2.150458, 51.981552],
-                    [-2.150597, 51.982063],
-                    [-2.150868, 51.982032],
-                    [-2.150835, 51.982209],
-                    [-2.151372, 51.982131],
-                    [-2.151536, 51.982272],
-                    [-2.152616, 51.982289],
-                    [-2.152793, 51.982332],
-                    [-2.152636, 51.982447],
-                    [-2.15253, 51.982444],
-                    [-2.152519, 51.982598],
-                    [-2.151636, 51.982606],
-                    [-2.151788, 51.983448],
-                    [-2.151905, 51.983449],
-                    [-2.151808, 51.983709],
-                    [-2.151849, 51.983717],
-                    [-2.151724, 51.984111],
-                    [-2.15145, 51.984522],
-                    [-2.151216, 51.985241],
-                    [-2.149742, 51.985313],
-                    [-2.149378, 51.985255],
-                    [-2.149332, 51.985122],
-                    [-2.149259, 51.985038],
-                    [-2.149279, 51.985029],
-                    [-2.149241, 51.985018],
-                    [-2.149152, 51.984884],
-                    [-2.149028, 51.984806],
-                    [-2.148795, 51.984462],
-                    [-2.148732, 51.984271],
-                    [-2.148729, 51.984144],
-                    [-2.148781, 51.984042],
-                    [-2.14874, 51.983907],
-                    [-2.148668, 51.983849],
-                    [-2.148468, 51.983807],
-                    [-2.148357, 51.983673],
-                    [-2.14817, 51.983285],
-                    [-2.148219, 51.983142],
-                    [-2.148565, 51.98289],
-                    [-2.148551, 51.982732],
-                    [-2.148595, 51.982624],
-                    [-2.148784, 51.982502],
-                    [-2.148979, 51.982434],
-                    [-2.149014, 51.98238],
-                    [-2.148852, 51.982073],
-                    [-2.148667, 51.981885],
-                    [-2.148752, 51.981612],
-                    [-2.148637, 51.981416],
-                    [-2.148624, 51.981161],
-                    [-2.148735, 51.981037],
-                    [-2.148752, 51.980971],
-                    [-2.148623, 51.980816],
-                    [-2.148443, 51.980689],
-                    [-2.148341, 51.980451],
-                    [-2.148654, 51.980039],
-                    [-2.148592, 51.979664],
-                    [-2.149459, 51.979312],
-                    [-2.149593, 51.979305],
-                    [-2.149807, 51.97936],
-                    [-2.149979, 51.979465],
-                    [-2.149871, 51.98094],
-                    [-2.150044, 51.980959],
-                    [-2.150461, 51.980897],
-                    [-2.150623, 51.980824],
-                    [-2.151307, 51.980673],
-                    [-2.151754, 51.980643],
-                    [-2.151866, 51.980672],
-                    [-2.152712, 51.980538],
-                    [-2.15289, 51.980552],
-                    [-2.153306, 51.980513],
-                    [-2.153648, 51.980153],
-                    [-2.154123, 51.980058],
-                    [-2.154583, 51.97887],
-                    [-2.15472, 51.978717],
-                    [-2.154802, 51.978401],
-                    [-2.154809, 51.978255],
-                    [-2.155132, 51.978249],
-                    [-2.155476, 51.976546],
-                    [-2.155925, 51.975352],
-                    [-2.156062, 51.975333],
-                    [-2.157088, 51.975428],
-                    [-2.158492, 51.975636],
-                    [-2.159428, 51.97592],
-                    [-2.159443, 51.975894],
-                    [-2.160052, 51.976023],
-                    [-2.160608, 51.976092],
-                    [-2.160668, 51.976135],
-                    [-2.161018, 51.977056],
-                    [-2.161258, 51.977167],
-                    [-2.161313, 51.977276],
-                    [-2.161317, 51.977454],
-                    [-2.161556, 51.977486],
-                    [-2.161725, 51.977501],
-                    [-2.161894, 51.977456],
-                    [-2.162442, 51.976889],
-                    [-2.162531, 51.97631],
-                    [-2.162577, 51.976301],
-                    [-2.163447, 51.976405],
-                    [-2.163471, 51.976367],
-                    [-2.16401, 51.976653],
-                    [-2.164128, 51.976621],
-                    [-2.167272, 51.976568],
-                    [-2.168226, 51.9771],
-                    [-2.168594, 51.977232],
-                    [-2.169181, 51.977543],
-                    [-2.169507, 51.977792],
-                    [-2.172738, 51.979106],
-                    [-2.173206, 51.979268],
-                    [-2.17344, 51.979459],
-                    [-2.175102, 51.983343],
-                    [-2.17535, 51.984033],
-                    [-2.174735, 51.984558],
-                    [-2.174255, 51.985186],
-                    [-2.174254, 51.98524],
-                    [-2.174502, 51.985679],
-                    [-2.174519, 51.985806],
-                    [-2.174496, 51.985847],
-                    [-2.174225, 51.985968],
-                    [-2.173538, 51.986126],
-                    [-2.173061, 51.986411],
-                    [-2.172728, 51.986489],
-                    [-2.172247, 51.986405],
-                    [-2.171755, 51.986454],
-                    [-2.171233, 51.986308],
-                    [-2.170991, 51.986295],
-                    [-2.170407, 51.986613],
-                    [-2.169999, 51.986701],
-                    [-2.169704, 51.986819],
-                    [-2.169556, 51.986949],
-                    [-2.169493, 51.987204],
-                    [-2.16934, 51.98732],
-                    [-2.169055, 51.987336],
-                    [-2.168509, 51.987113],
-                    [-2.16818, 51.987148],
-                    [-2.167988, 51.987232],
-                    [-2.167611, 51.987585],
-                    [-2.167332, 51.987731],
-                    [-2.167102, 51.987783],
-                    [-2.166597, 51.987741],
-                    [-2.166262, 51.987789],
-                    [-2.166136, 51.987868],
-                    [-2.166021, 51.988031],
-                    [-2.16596, 51.988041],
-                    [-2.165709, 51.988223],
-                    [-2.16539, 51.988348],
-                    [-2.165172, 51.988297],
-                    [-2.164143, 51.988231],
-                    [-2.163692, 51.988311],
-                    [-2.163276, 51.988343],
-                    [-2.162939, 51.988509],
-                    [-2.162627, 51.988578],
-                    [-2.162303, 51.988608],
-                    [-2.162138, 51.988712],
-                    [-2.161984, 51.988752],
-                    [-2.161272, 51.988693],
-                    [-2.160907, 51.988405]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2017-03-30",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400039,
-              "name": "Battle of Tewkesbury 1471",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000039",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000039"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000039"
-          }
-        },
-        {
-          "entry-date": "2017-06-01",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400040,
-          "name": "Battle of Towton, 1461",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000040",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.263726, 53.856821],
-                    [-1.261758, 53.855954],
-                    [-1.262766, 53.853169],
-                    [-1.263122, 53.852367],
-                    [-1.263708, 53.851272],
-                    [-1.264206, 53.851379],
-                    [-1.264268, 53.85105],
-                    [-1.265254, 53.85097],
-                    [-1.265284, 53.850747],
-                    [-1.265198, 53.850634],
-                    [-1.264994, 53.850542],
-                    [-1.264773, 53.850547],
-                    [-1.264485, 53.850314],
-                    [-1.264406, 53.850175],
-                    [-1.264387, 53.850039],
-                    [-1.264035, 53.850021],
-                    [-1.264041, 53.849989],
-                    [-1.26412, 53.849994],
-                    [-1.264231, 53.84953],
-                    [-1.264159, 53.849528],
-                    [-1.264359, 53.849161],
-                    [-1.265587, 53.848168],
-                    [-1.265608, 53.848097],
-                    [-1.26535, 53.848089],
-                    [-1.265324, 53.84813],
-                    [-1.264503, 53.848051],
-                    [-1.264333, 53.847983],
-                    [-1.25969, 53.847355],
-                    [-1.25183, 53.846308],
-                    [-1.251817, 53.846337],
-                    [-1.251649, 53.846314],
-                    [-1.251715, 53.84601],
-                    [-1.251967, 53.845545],
-                    [-1.251978, 53.845473],
-                    [-1.252698, 53.844783],
-                    [-1.252674, 53.844774],
-                    [-1.252746, 53.844628],
-                    [-1.252957, 53.844456],
-                    [-1.253313, 53.844059],
-                    [-1.254877, 53.842792],
-                    [-1.257513, 53.839367],
-                    [-1.257466, 53.839302],
-                    [-1.257551, 53.839304],
-                    [-1.258153, 53.838063],
-                    [-1.258289, 53.837963],
-                    [-1.258274, 53.837743],
-                    [-1.25818, 53.837676],
-                    [-1.258396, 53.837391],
-                    [-1.258456, 53.837177],
-                    [-1.258513, 53.836478],
-                    [-1.258594, 53.836034],
-                    [-1.258715, 53.835656],
-                    [-1.258725, 53.835203],
-                    [-1.25888, 53.835106],
-                    [-1.25906, 53.834859],
-                    [-1.259084, 53.834749],
-                    [-1.259567, 53.834152],
-                    [-1.260272, 53.834206],
-                    [-1.261826, 53.834236],
-                    [-1.26485, 53.834183],
-                    [-1.264868, 53.834117],
-                    [-1.264923, 53.833776],
-                    [-1.264895, 53.833489],
-                    [-1.264313, 53.832117],
-                    [-1.263833, 53.830871],
-                    [-1.263552, 53.830257],
-                    [-1.263338, 53.829923],
-                    [-1.263082, 53.829626],
-                    [-1.262221, 53.828903],
-                    [-1.261884, 53.828521],
-                    [-1.261466, 53.827835],
-                    [-1.261116, 53.826962],
-                    [-1.260938, 53.826632],
-                    [-1.260246, 53.825736],
-                    [-1.260071, 53.825422],
-                    [-1.259843, 53.824695],
-                    [-1.25931, 53.822675],
-                    [-1.259142, 53.822149],
-                    [-1.258886, 53.82154],
-                    [-1.258688, 53.820239],
-                    [-1.262973, 53.820129],
-                    [-1.263341, 53.820141],
-                    [-1.264933, 53.820497],
-                    [-1.267764, 53.820884],
-                    [-1.268982, 53.82118],
-                    [-1.272256, 53.822092],
-                    [-1.274632, 53.82288],
-                    [-1.275894, 53.823114],
-                    [-1.275748, 53.823404],
-                    [-1.276294, 53.823493],
-                    [-1.276278, 53.823527],
-                    [-1.278376, 53.823805],
-                    [-1.278275, 53.824585],
-                    [-1.278212, 53.824858],
-                    [-1.277599, 53.824854],
-                    [-1.276344, 53.824947],
-                    [-1.276352, 53.825028],
-                    [-1.275528, 53.825041],
-                    [-1.275458, 53.824849],
-                    [-1.274905, 53.824871],
-                    [-1.274926, 53.825336],
-                    [-1.275433, 53.825508],
-                    [-1.275859, 53.8255],
-                    [-1.276751, 53.825395],
-                    [-1.277235, 53.825434],
-                    [-1.277405, 53.826341],
-                    [-1.274615, 53.8267],
-                    [-1.274324, 53.826271],
-                    [-1.273469, 53.826412],
-                    [-1.273873, 53.82706],
-                    [-1.274216, 53.827035],
-                    [-1.274301, 53.827395],
-                    [-1.274754, 53.827364],
-                    [-1.274791, 53.827594],
-                    [-1.274759, 53.827597],
-                    [-1.274857, 53.828111],
-                    [-1.275309, 53.828047],
-                    [-1.275297, 53.827992],
-                    [-1.275355, 53.827962],
-                    [-1.278343, 53.827595],
-                    [-1.278328, 53.827522],
-                    [-1.279126, 53.827474],
-                    [-1.279138, 53.827528],
-                    [-1.279431, 53.827505],
-                    [-1.279461, 53.827702],
-                    [-1.279769, 53.827686],
-                    [-1.279745, 53.827525],
-                    [-1.281348, 53.827392],
-                    [-1.281329, 53.827352],
-                    [-1.281924, 53.827028],
-                    [-1.282858, 53.827581],
-                    [-1.28414, 53.828233],
-                    [-1.284337, 53.828363],
-                    [-1.284446, 53.828498],
-                    [-1.284484, 53.828703],
-                    [-1.284402, 53.8299],
-                    [-1.284437, 53.830129],
-                    [-1.284617, 53.830394],
-                    [-1.284974, 53.830723],
-                    [-1.285213, 53.830876],
-                    [-1.285708, 53.831512],
-                    [-1.285136, 53.831716],
-                    [-1.285123, 53.831784],
-                    [-1.284925, 53.831832],
-                    [-1.284922, 53.831901],
-                    [-1.284552, 53.831899],
-                    [-1.284634, 53.832493],
-                    [-1.28574, 53.83246],
-                    [-1.286424, 53.832342],
-                    [-1.286583, 53.832597],
-                    [-1.287047, 53.833101],
-                    [-1.287335, 53.833244],
-                    [-1.28782, 53.833372],
-                    [-1.288697, 53.833516],
-                    [-1.289665, 53.833543],
-                    [-1.291844, 53.833736],
-                    [-1.292564, 53.833818],
-                    [-1.293127, 53.833922],
-                    [-1.293563, 53.835728],
-                    [-1.293592, 53.836452],
-                    [-1.293509, 53.837368],
-                    [-1.292287, 53.837716],
-                    [-1.292152, 53.83778],
-                    [-1.29062, 53.838771],
-                    [-1.290665, 53.838904],
-                    [-1.291208, 53.839367],
-                    [-1.291742, 53.839695],
-                    [-1.291229, 53.83987],
-                    [-1.290501, 53.840058],
-                    [-1.289058, 53.840204],
-                    [-1.288432, 53.840214],
-                    [-1.283387, 53.839995],
-                    [-1.282107, 53.840047],
-                    [-1.282887, 53.841185],
-                    [-1.283611, 53.841563],
-                    [-1.283819, 53.841741],
-                    [-1.284407, 53.842145],
-                    [-1.287654, 53.843661],
-                    [-1.288192, 53.846063],
-                    [-1.288159, 53.846411],
-                    [-1.288027, 53.846824],
-                    [-1.287974, 53.847207],
-                    [-1.287977, 53.84772],
-                    [-1.288013, 53.847843],
-                    [-1.287218, 53.847923],
-                    [-1.286312, 53.847965],
-                    [-1.286216, 53.848006],
-                    [-1.285696, 53.849646],
-                    [-1.285092, 53.851401],
-                    [-1.283947, 53.85167],
-                    [-1.284405, 53.852348],
-                    [-1.28436, 53.852366],
-                    [-1.284373, 53.852447],
-                    [-1.284332, 53.852642],
-                    [-1.28436, 53.852648],
-                    [-1.284237, 53.852716],
-                    [-1.283969, 53.852785],
-                    [-1.283888, 53.852771],
-                    [-1.283957, 53.85343],
-                    [-1.283944, 53.853568],
-                    [-1.284266, 53.854103],
-                    [-1.284277, 53.854276],
-                    [-1.284071, 53.854271],
-                    [-1.283585, 53.854616],
-                    [-1.282477, 53.855232],
-                    [-1.281894, 53.855626],
-                    [-1.281429, 53.855842],
-                    [-1.280247, 53.856468],
-                    [-1.279874, 53.856603],
-                    [-1.279618, 53.856767],
-                    [-1.27892, 53.857063],
-                    [-1.278637, 53.856888],
-                    [-1.277631, 53.857183],
-                    [-1.276775, 53.857391],
-                    [-1.276264, 53.85756],
-                    [-1.274804, 53.857756],
-                    [-1.273808, 53.85798],
-                    [-1.272646, 53.858337],
-                    [-1.269381, 53.858961],
-                    [-1.266599, 53.857812],
-                    [-1.265936, 53.857562],
-                    [-1.263726, 53.856821]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2017-06-01",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400040,
-              "name": "Battle of Towton, 1461",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000040",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000040"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000040"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400041,
-          "name": "Battle of Winceby 1643",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000041",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-0.027349, 53.207732],
-                    [-0.025884, 53.20171],
-                    [-0.025577, 53.200804],
-                    [-0.025673, 53.20071],
-                    [-0.025787, 53.20029],
-                    [-0.025709, 53.199709],
-                    [-0.026277, 53.19826],
-                    [-0.026128, 53.197737],
-                    [-0.025996, 53.197527],
-                    [-0.026059, 53.197512],
-                    [-0.026321, 53.196771],
-                    [-0.025451, 53.196236],
-                    [-0.025183, 53.196289],
-                    [-0.024804, 53.196093],
-                    [-0.025239, 53.1955],
-                    [-0.025625, 53.195105],
-                    [-0.025764, 53.194855],
-                    [-0.026357, 53.194671],
-                    [-0.03666, 53.191993],
-                    [-0.037086, 53.192056],
-                    [-0.038552, 53.192448],
-                    [-0.041783, 53.193192],
-                    [-0.042231, 53.193431],
-                    [-0.042767, 53.193623],
-                    [-0.043885, 53.193712],
-                    [-0.044213, 53.193443],
-                    [-0.045539, 53.193288],
-                    [-0.045372, 53.19343],
-                    [-0.045216, 53.193796],
-                    [-0.044948, 53.194272],
-                    [-0.044541, 53.194757],
-                    [-0.04413, 53.195182],
-                    [-0.043748, 53.196401],
-                    [-0.042783, 53.19776],
-                    [-0.042575, 53.198386],
-                    [-0.042329, 53.198661],
-                    [-0.043184, 53.198842],
-                    [-0.043624, 53.198988],
-                    [-0.040594, 53.207603],
-                    [-0.037827, 53.207249],
-                    [-0.035141, 53.207108],
-                    [-0.034178, 53.207112],
-                    [-0.032242, 53.207204],
-                    [-0.027349, 53.207732]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400041,
-              "name": "Battle of Winceby 1643",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000041",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000041"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000041"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "1995-06-06",
-          "end-date": "",
-          "entity": 1400042,
-          "name": "Battle of Worcester 1651 with Powick Bridge 1642",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1000042",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.217404, 52.177344],
-                    [-2.217394, 52.176779],
-                    [-2.21733, 52.176464],
-                    [-2.217215, 52.176232],
-                    [-2.216535, 52.175241],
-                    [-2.216593, 52.175193],
-                    [-2.217025, 52.175036],
-                    [-2.217048, 52.175052],
-                    [-2.217076, 52.174857],
-                    [-2.21714, 52.174831],
-                    [-2.217234, 52.17472],
-                    [-2.217236, 52.174446],
-                    [-2.217329, 52.174421],
-                    [-2.217449, 52.174436],
-                    [-2.217503, 52.174393],
-                    [-2.217419, 52.174292],
-                    [-2.217528, 52.174242],
-                    [-2.217538, 52.174143],
-                    [-2.217668, 52.173999],
-                    [-2.217485, 52.173906],
-                    [-2.217495, 52.173834],
-                    [-2.217691, 52.173601],
-                    [-2.217633, 52.173552],
-                    [-2.21762, 52.173493],
-                    [-2.217805, 52.173351],
-                    [-2.217834, 52.173233],
-                    [-2.21873, 52.173187],
-                    [-2.218927, 52.173149],
-                    [-2.21936, 52.172991],
-                    [-2.218824, 52.172397],
-                    [-2.218736, 52.172211],
-                    [-2.218666, 52.172175],
-                    [-2.218457, 52.171679],
-                    [-2.218467, 52.171586],
-                    [-2.21839, 52.171447],
-                    [-2.218382, 52.171292],
-                    [-2.218277, 52.171098],
-                    [-2.218263, 52.170731],
-                    [-2.218349, 52.170466],
-                    [-2.218379, 52.170233],
-                    [-2.218486, 52.170096],
-                    [-2.218494, 52.169951],
-                    [-2.218669, 52.169637],
-                    [-2.218761, 52.169568],
-                    [-2.218841, 52.169446],
-                    [-2.217301, 52.169232],
-                    [-2.216984, 52.169215],
-                    [-2.217039, 52.169128],
-                    [-2.216389, 52.16878],
-                    [-2.214349, 52.168831],
-                    [-2.214009, 52.168799],
-                    [-2.213836, 52.168692],
-                    [-2.213401, 52.169271],
-                    [-2.209144, 52.17058],
-                    [-2.20994, 52.170954],
-                    [-2.210729, 52.171389],
-                    [-2.211294, 52.171769],
-                    [-2.211767, 52.172141],
-                    [-2.211965, 52.172252],
-                    [-2.212135, 52.172295],
-                    [-2.212366, 52.17248],
-                    [-2.212294, 52.172826],
-                    [-2.212299, 52.173285],
-                    [-2.211145, 52.173453],
-                    [-2.21119, 52.173568],
-                    [-2.209733, 52.173802],
-                    [-2.20936, 52.173828],
-                    [-2.208805, 52.174147],
-                    [-2.208667, 52.174254],
-                    [-2.208685, 52.174265],
-                    [-2.208103, 52.174672],
-                    [-2.207859, 52.175176],
-                    [-2.207922, 52.175369],
-                    [-2.208008, 52.175463],
-                    [-2.207995, 52.175543],
-                    [-2.208274, 52.175809],
-                    [-2.208495, 52.175921],
-                    [-2.208629, 52.176079],
-                    [-2.208977, 52.17626],
-                    [-2.209551, 52.176448],
-                    [-2.209904, 52.176622],
-                    [-2.210719, 52.176715],
-                    [-2.211461, 52.176904],
-                    [-2.211578, 52.177046],
-                    [-2.211779, 52.177046],
-                    [-2.211954, 52.177107],
-                    [-2.21219, 52.177112],
-                    [-2.212364, 52.177049],
-                    [-2.213539, 52.176987],
-                    [-2.213824, 52.176912],
-                    [-2.213959, 52.177188],
-                    [-2.213336, 52.177322],
-                    [-2.213654, 52.177974],
-                    [-2.213618, 52.178163],
-                    [-2.213459, 52.178368],
-                    [-2.21333, 52.178426],
-                    [-2.211955, 52.180535],
-                    [-2.212423, 52.180654],
-                    [-2.21164, 52.181636],
-                    [-2.210011, 52.181123],
-                    [-2.209349, 52.180812],
-                    [-2.209147, 52.180665],
-                    [-2.208994, 52.180683],
-                    [-2.208832, 52.18065],
-                    [-2.207263, 52.179558],
-                    [-2.206993, 52.179421],
-                    [-2.206367, 52.178984],
-                    [-2.206301, 52.178973],
-                    [-2.206348, 52.178927],
-                    [-2.203595, 52.177625],
-                    [-2.202291, 52.177059],
-                    [-2.202016, 52.176901],
-                    [-2.201069, 52.177398],
-                    [-2.200596, 52.177225],
-                    [-2.200219, 52.17702],
-                    [-2.199808, 52.176658],
-                    [-2.19897, 52.175696],
-                    [-2.198876, 52.175635],
-                    [-2.198417, 52.175111],
-                    [-2.197917, 52.174341],
-                    [-2.197564, 52.173872],
-                    [-2.196948, 52.173328],
-                    [-2.196808, 52.17315],
-                    [-2.196505, 52.172678],
-                    [-2.195412, 52.170751],
-                    [-2.195033, 52.170178],
-                    [-2.19476, 52.169831],
-                    [-2.195783, 52.1696],
-                    [-2.196067, 52.169474],
-                    [-2.198339, 52.168881],
-                    [-2.200315, 52.168306],
-                    [-2.202236, 52.168016],
-                    [-2.202872, 52.168007],
-                    [-2.203465, 52.167901],
-                    [-2.20352, 52.167741],
-                    [-2.203503, 52.167599],
-                    [-2.203634, 52.167446],
-                    [-2.20486, 52.166747],
-                    [-2.206146, 52.1662],
-                    [-2.209396, 52.165187],
-                    [-2.215455, 52.163506],
-                    [-2.215919, 52.16351],
-                    [-2.216158, 52.16355],
-                    [-2.216707, 52.163407],
-                    [-2.216923, 52.163196],
-                    [-2.219954, 52.162534],
-                    [-2.222397, 52.162321],
-                    [-2.223923, 52.162293],
-                    [-2.225114, 52.162368],
-                    [-2.226691, 52.162602],
-                    [-2.22741, 52.162765],
-                    [-2.228687, 52.163149],
-                    [-2.22866, 52.16319],
-                    [-2.22959, 52.163522],
-                    [-2.230083, 52.163648],
-                    [-2.23041, 52.163858],
-                    [-2.231994, 52.164475],
-                    [-2.232262, 52.164536],
-                    [-2.234594, 52.165439],
-                    [-2.234749, 52.165443],
-                    [-2.239437, 52.167399],
-                    [-2.240994, 52.16794],
-                    [-2.241274, 52.167997],
-                    [-2.241739, 52.168265],
-                    [-2.243221, 52.169581],
-                    [-2.244139, 52.170349],
-                    [-2.244012, 52.170644],
-                    [-2.243785, 52.170904],
-                    [-2.243425, 52.171167],
-                    [-2.242921, 52.171605],
-                    [-2.242558, 52.171815],
-                    [-2.242108, 52.171916],
-                    [-2.241889, 52.171933],
-                    [-2.239543, 52.172364],
-                    [-2.239199, 52.172377],
-                    [-2.239036, 52.172266],
-                    [-2.239459, 52.171892],
-                    [-2.239559, 52.171755],
-                    [-2.239168, 52.171668],
-                    [-2.238997, 52.171441],
-                    [-2.238255, 52.171501],
-                    [-2.237317, 52.171644],
-                    [-2.237223, 52.171684],
-                    [-2.236774, 52.171623],
-                    [-2.236303, 52.171789],
-                    [-2.236221, 52.171849],
-                    [-2.235773, 52.171983],
-                    [-2.235488, 52.172038],
-                    [-2.235207, 52.172029],
-                    [-2.235077, 52.172178],
-                    [-2.235316, 52.172747],
-                    [-2.235369, 52.173263],
-                    [-2.234595, 52.173],
-                    [-2.234484, 52.173191],
-                    [-2.234151, 52.173165],
-                    [-2.234391, 52.174412],
-                    [-2.23229, 52.173791],
-                    [-2.23248, 52.174668],
-                    [-2.228887, 52.175002],
-                    [-2.229134, 52.176028],
-                    [-2.228334, 52.17611],
-                    [-2.228277, 52.176082],
-                    [-2.228198, 52.175974],
-                    [-2.22806, 52.175964],
-                    [-2.22736, 52.176117],
-                    [-2.227209, 52.176086],
-                    [-2.225642, 52.176595],
-                    [-2.225359, 52.176626],
-                    [-2.22435, 52.176399],
-                    [-2.223417, 52.176683],
-                    [-2.22388, 52.177079],
-                    [-2.224009, 52.177231],
-                    [-2.224099, 52.177418],
-                    [-2.224044, 52.177937],
-                    [-2.223798, 52.177968],
-                    [-2.222763, 52.178239],
-                    [-2.222042, 52.17728],
-                    [-2.221843, 52.177333],
-                    [-2.221804, 52.177375],
-                    [-2.219959, 52.177909],
-                    [-2.220063, 52.178862],
-                    [-2.220056, 52.179011],
-                    [-2.219377, 52.179297],
-                    [-2.219308, 52.179347],
-                    [-2.219244, 52.17946],
-                    [-2.219276, 52.179628],
-                    [-2.219802, 52.181067],
-                    [-2.22012, 52.1816],
-                    [-2.219025, 52.181972],
-                    [-2.218945, 52.181929],
-                    [-2.218258, 52.181064],
-                    [-2.218021, 52.180506],
-                    [-2.217704, 52.179527],
-                    [-2.217594, 52.179053],
-                    [-2.217515, 52.17833],
-                    [-2.217416, 52.177785],
-                    [-2.217404, 52.177344]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "1995-06-06",
-              "end-date": "",
-              "entity": 1400042,
-              "name": "Battle of Worcester 1651 with Powick Bridge 1642",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1000042",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000042"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1000042"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "2018-01-31",
-          "end-date": "",
-          "entity": 1400043,
-          "name": "Battle of Winwick (also known as Battle of Red Bank) 1648",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1412878",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-2.610565, 53.447178],
-                    [-2.606558, 53.446447],
-                    [-2.606484, 53.445909],
-                    [-2.60488, 53.445988],
-                    [-2.603396, 53.445281],
-                    [-2.602654, 53.444134],
-                    [-2.602639, 53.442416],
-                    [-2.602444, 53.442419],
-                    [-2.602311, 53.442469],
-                    [-2.601967, 53.442789],
-                    [-2.601312, 53.443113],
-                    [-2.601068, 53.443164],
-                    [-2.600665, 53.443511],
-                    [-2.599993, 53.443383],
-                    [-2.59934, 53.443445],
-                    [-2.59901, 53.443399],
-                    [-2.598741, 53.443321],
-                    [-2.597091, 53.442529],
-                    [-2.596483, 53.442491],
-                    [-2.595897, 53.442356],
-                    [-2.595976, 53.442026],
-                    [-2.595845, 53.441988],
-                    [-2.595882, 53.441843],
-                    [-2.594811, 53.441754],
-                    [-2.594387, 53.441667],
-                    [-2.594596, 53.441513],
-                    [-2.595315, 53.440744],
-                    [-2.595469, 53.440502],
-                    [-2.595529, 53.440254],
-                    [-2.595507, 53.43964],
-                    [-2.595555, 53.43916],
-                    [-2.595635, 53.438662],
-                    [-2.59578, 53.438056],
-                    [-2.596881, 53.435562],
-                    [-2.597133, 53.434798],
-                    [-2.597593, 53.434055],
-                    [-2.597706, 53.433796],
-                    [-2.597801, 53.432948],
-                    [-2.597798, 53.432557],
-                    [-2.597735, 53.432235],
-                    [-2.597367, 53.431589],
-                    [-2.597271, 53.431299],
-                    [-2.597336, 53.431098],
-                    [-2.597217, 53.430909],
-                    [-2.597118, 53.430656],
-                    [-2.597249, 53.430233],
-                    [-2.59727, 53.430016],
-                    [-2.597655, 53.430003],
-                    [-2.597967, 53.430422],
-                    [-2.59875, 53.431051],
-                    [-2.602771, 53.430659],
-                    [-2.60457, 53.430431],
-                    [-2.604663, 53.430391],
-                    [-2.604697, 53.430517],
-                    [-2.604524, 53.430783],
-                    [-2.604355, 53.430982],
-                    [-2.604087, 53.431171],
-                    [-2.604217, 53.432025],
-                    [-2.606459, 53.433285],
-                    [-2.606333, 53.433419],
-                    [-2.60634, 53.433564],
-                    [-2.6064, 53.433743],
-                    [-2.606309, 53.433827],
-                    [-2.60633, 53.434062],
-                    [-2.60625, 53.434123],
-                    [-2.606438, 53.434129],
-                    [-2.606494, 53.434197],
-                    [-2.607749, 53.433766],
-                    [-2.609796, 53.435921],
-                    [-2.610182, 53.43643],
-                    [-2.61055, 53.436995],
-                    [-2.611065, 53.437832],
-                    [-2.611083, 53.43791],
-                    [-2.611275, 53.437882],
-                    [-2.611432, 53.43792],
-                    [-2.611533, 53.437999],
-                    [-2.611509, 53.43818],
-                    [-2.611553, 53.438275],
-                    [-2.611857, 53.438431],
-                    [-2.612271, 53.438528],
-                    [-2.615587, 53.441326],
-                    [-2.614971, 53.441587],
-                    [-2.614197, 53.441815],
-                    [-2.614474, 53.442583],
-                    [-2.614303, 53.442602],
-                    [-2.614361, 53.442791],
-                    [-2.614367, 53.443143],
-                    [-2.613432, 53.442919],
-                    [-2.611907, 53.442989],
-                    [-2.611938, 53.443179],
-                    [-2.611903, 53.443234],
-                    [-2.611897, 53.443412],
-                    [-2.611997, 53.443642],
-                    [-2.611973, 53.443702],
-                    [-2.611919, 53.443725],
-                    [-2.611916, 53.443767],
-                    [-2.611585, 53.443829],
-                    [-2.611447, 53.443747],
-                    [-2.611373, 53.443434],
-                    [-2.610907, 53.443465],
-                    [-2.611015, 53.443927],
-                    [-2.610831, 53.44396],
-                    [-2.61061, 53.443072],
-                    [-2.610406, 53.443138],
-                    [-2.610056, 53.442963],
-                    [-2.609845, 53.442976],
-                    [-2.609892, 53.443019],
-                    [-2.609995, 53.443403],
-                    [-2.610755, 53.446696],
-                    [-2.610565, 53.447178]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "2018-01-31",
-              "end-date": "",
-              "entity": 1400043,
-              "name": "Battle of Winwick (also known as Battle of Red Bank) 1648",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1412878",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1412878"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1412878"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "2013-07-15",
-          "end-date": "",
-          "entity": 1400044,
-          "name": "Battle of Lostwithiel 21 August 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1413619",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-4.637582, 50.425475],
-                    [-4.637489, 50.425474],
-                    [-4.637733, 50.424912],
-                    [-4.637697, 50.424558],
-                    [-4.637651, 50.424341],
-                    [-4.637397, 50.423788],
-                    [-4.637099, 50.42336],
-                    [-4.636583, 50.422926],
-                    [-4.635985, 50.422347],
-                    [-4.635737, 50.422242],
-                    [-4.635309, 50.421972],
-                    [-4.634863, 50.421164],
-                    [-4.634733, 50.420762],
-                    [-4.634682, 50.420101],
-                    [-4.634711, 50.419631],
-                    [-4.634673, 50.419153],
-                    [-4.634579, 50.418292],
-                    [-4.63453, 50.418237],
-                    [-4.634634, 50.418243],
-                    [-4.634619, 50.4181],
-                    [-4.634539, 50.417951],
-                    [-4.634713, 50.417519],
-                    [-4.634812, 50.417116],
-                    [-4.635084, 50.416619],
-                    [-4.635307, 50.416094],
-                    [-4.635402, 50.415773],
-                    [-4.635445, 50.415477],
-                    [-4.636132, 50.415509],
-                    [-4.63631, 50.414984],
-                    [-4.636439, 50.414827],
-                    [-4.636724, 50.414666],
-                    [-4.636694, 50.414438],
-                    [-4.636732, 50.414031],
-                    [-4.636869, 50.413409],
-                    [-4.636216, 50.41331],
-                    [-4.635515, 50.413443],
-                    [-4.634764, 50.413506],
-                    [-4.634748, 50.413115],
-                    [-4.634924, 50.412032],
-                    [-4.634805, 50.412014],
-                    [-4.634824, 50.411832],
-                    [-4.63503, 50.411154],
-                    [-4.635253, 50.410623],
-                    [-4.635325, 50.410485],
-                    [-4.635441, 50.410374],
-                    [-4.635358, 50.410307],
-                    [-4.63528, 50.409965],
-                    [-4.635251, 50.409964],
-                    [-4.635233, 50.409609],
-                    [-4.635316, 50.409064],
-                    [-4.635371, 50.409004],
-                    [-4.635774, 50.408887],
-                    [-4.636994, 50.407914],
-                    [-4.637538, 50.407388],
-                    [-4.637754, 50.407357],
-                    [-4.638048, 50.407393],
-                    [-4.638317, 50.407359],
-                    [-4.638474, 50.407308],
-                    [-4.638712, 50.40716],
-                    [-4.638905, 50.406896],
-                    [-4.639824, 50.405848],
-                    [-4.639957, 50.405704],
-                    [-4.640163, 50.40555],
-                    [-4.641822, 50.404746],
-                    [-4.643482, 50.404172],
-                    [-4.643982, 50.403961],
-                    [-4.644117, 50.403804],
-                    [-4.644382, 50.403284],
-                    [-4.644581, 50.402576],
-                    [-4.644671, 50.401803],
-                    [-4.644765, 50.401529],
-                    [-4.64505, 50.401178],
-                    [-4.646381, 50.399938],
-                    [-4.64651, 50.399682],
-                    [-4.646799, 50.399854],
-                    [-4.646933, 50.399993],
-                    [-4.647049, 50.400023],
-                    [-4.647324, 50.400361],
-                    [-4.647425, 50.40072],
-                    [-4.647373, 50.401298],
-                    [-4.647399, 50.401474],
-                    [-4.647514, 50.401804],
-                    [-4.647627, 50.401984],
-                    [-4.648309, 50.402519],
-                    [-4.648473, 50.402735],
-                    [-4.648995, 50.4032],
-                    [-4.649518, 50.403773],
-                    [-4.649664, 50.404024],
-                    [-4.650591, 50.404046],
-                    [-4.650493, 50.405374],
-                    [-4.650427, 50.405467],
-                    [-4.648029, 50.4049],
-                    [-4.647403, 50.406014],
-                    [-4.64681, 50.407159],
-                    [-4.646924, 50.407279],
-                    [-4.647019, 50.407311],
-                    [-4.647179, 50.407203],
-                    [-4.647446, 50.407126],
-                    [-4.647731, 50.40709],
-                    [-4.648006, 50.4071],
-                    [-4.648394, 50.407158],
-                    [-4.649379, 50.407387],
-                    [-4.650066, 50.407487],
-                    [-4.650199, 50.407531],
-                    [-4.65034, 50.407662],
-                    [-4.650452, 50.407945],
-                    [-4.650501, 50.408179],
-                    [-4.650405, 50.408289],
-                    [-4.649747, 50.408671],
-                    [-4.648761, 50.409329],
-                    [-4.648452, 50.409592],
-                    [-4.648184, 50.410013],
-                    [-4.647487, 50.411466],
-                    [-4.647326, 50.411867],
-                    [-4.647278, 50.412144],
-                    [-4.647218, 50.412192],
-                    [-4.647255, 50.412256],
-                    [-4.647228, 50.412471],
-                    [-4.647179, 50.412498],
-                    [-4.647186, 50.412624],
-                    [-4.647216, 50.412624],
-                    [-4.647128, 50.413405],
-                    [-4.646954, 50.414112],
-                    [-4.646762, 50.414708],
-                    [-4.64645, 50.415325],
-                    [-4.646212, 50.415718],
-                    [-4.645795, 50.416331],
-                    [-4.645734, 50.416369],
-                    [-4.64577, 50.416374],
-                    [-4.644825, 50.417768],
-                    [-4.645924, 50.417906],
-                    [-4.645895, 50.417965],
-                    [-4.646349, 50.418059],
-                    [-4.646339, 50.418077],
-                    [-4.646553, 50.418101],
-                    [-4.646636, 50.418059],
-                    [-4.647261, 50.418186],
-                    [-4.647127, 50.417948],
-                    [-4.647155, 50.417919],
-                    [-4.647451, 50.417844],
-                    [-4.647409, 50.417795],
-                    [-4.647771, 50.417666],
-                    [-4.647512, 50.417591],
-                    [-4.647558, 50.417413],
-                    [-4.647525, 50.417287],
-                    [-4.647619, 50.417182],
-                    [-4.647587, 50.417112],
-                    [-4.647618, 50.417034],
-                    [-4.648215, 50.416676],
-                    [-4.648955, 50.416286],
-                    [-4.649093, 50.416339],
-                    [-4.64915, 50.416397],
-                    [-4.649555, 50.41638],
-                    [-4.649768, 50.41657],
-                    [-4.65063, 50.417023],
-                    [-4.65178, 50.4179],
-                    [-4.652031, 50.417663],
-                    [-4.652526, 50.417329],
-                    [-4.652778, 50.417037],
-                    [-4.65337, 50.416468],
-                    [-4.653535, 50.416349],
-                    [-4.653785, 50.416243],
-                    [-4.654435, 50.416073],
-                    [-4.654824, 50.415926],
-                    [-4.656064, 50.415853],
-                    [-4.656183, 50.415856],
-                    [-4.656147, 50.417271],
-                    [-4.657226, 50.417133],
-                    [-4.657649, 50.41713],
-                    [-4.65806, 50.416956],
-                    [-4.658401, 50.417127],
-                    [-4.658592, 50.417268],
-                    [-4.658707, 50.417409],
-                    [-4.65906, 50.417552],
-                    [-4.659123, 50.417646],
-                    [-4.659012, 50.417758],
-                    [-4.6587, 50.417951],
-                    [-4.658124, 50.418235],
-                    [-4.658141, 50.418437],
-                    [-4.658066, 50.418661],
-                    [-4.657831, 50.418665],
-                    [-4.657773, 50.418834],
-                    [-4.65769, 50.418897],
-                    [-4.657482, 50.418949],
-                    [-4.656824, 50.419001],
-                    [-4.656764, 50.419382],
-                    [-4.656679, 50.41958],
-                    [-4.656614, 50.419646],
-                    [-4.656112, 50.419611],
-                    [-4.656075, 50.419635],
-                    [-4.655562, 50.420303],
-                    [-4.655049, 50.420809],
-                    [-4.654613, 50.421049],
-                    [-4.654511, 50.421036],
-                    [-4.654181, 50.420902],
-                    [-4.653874, 50.420876],
-                    [-4.653732, 50.421047],
-                    [-4.65368, 50.4212],
-                    [-4.654126, 50.421314],
-                    [-4.654317, 50.421408],
-                    [-4.654314, 50.421477],
-                    [-4.654236, 50.421583],
-                    [-4.65453, 50.421732],
-                    [-4.654879, 50.421824],
-                    [-4.654767, 50.422033],
-                    [-4.654408, 50.422447],
-                    [-4.654202, 50.422744],
-                    [-4.653926, 50.42282],
-                    [-4.651067, 50.422727],
-                    [-4.649845, 50.422739],
-                    [-4.6485, 50.422651],
-                    [-4.647275, 50.422411],
-                    [-4.647071, 50.422342],
-                    [-4.647209, 50.422168],
-                    [-4.648277, 50.42125],
-                    [-4.646612, 50.420568],
-                    [-4.646486, 50.420584],
-                    [-4.646332, 50.420734],
-                    [-4.646119, 50.420697],
-                    [-4.644853, 50.420061],
-                    [-4.64456, 50.419942],
-                    [-4.643796, 50.419726],
-                    [-4.6437, 50.419979],
-                    [-4.643499, 50.420789],
-                    [-4.643255, 50.422358],
-                    [-4.643165, 50.422747],
-                    [-4.642962, 50.423308],
-                    [-4.642751, 50.423569],
-                    [-4.64278, 50.423676],
-                    [-4.642688, 50.423999],
-                    [-4.642113, 50.424078],
-                    [-4.64056, 50.424516],
-                    [-4.639209, 50.425169],
-                    [-4.638515, 50.425253],
-                    [-4.637582, 50.425475]
-                  ]
-                ],
-                [
-                  [
-                    [-4.672132, 50.4248],
-                    [-4.671835, 50.425021],
-                    [-4.671458, 50.424856],
-                    [-4.671205, 50.42481],
-                    [-4.670762, 50.424825],
-                    [-4.670285, 50.42478],
-                    [-4.670046, 50.424239],
-                    [-4.669967, 50.424153],
-                    [-4.669907, 50.423912],
-                    [-4.669834, 50.423772],
-                    [-4.669491, 50.423618],
-                    [-4.668392, 50.423276],
-                    [-4.668507, 50.424188],
-                    [-4.668622, 50.424497],
-                    [-4.66795, 50.424164],
-                    [-4.66669, 50.423659],
-                    [-4.665947, 50.423263],
-                    [-4.66543, 50.422926],
-                    [-4.664983, 50.422507],
-                    [-4.664447, 50.421888],
-                    [-4.66429, 50.42165],
-                    [-4.663955, 50.420921],
-                    [-4.66388, 50.420557],
-                    [-4.663837, 50.420112],
-                    [-4.663891, 50.418653],
-                    [-4.664224, 50.418673],
-                    [-4.666843, 50.419109],
-                    [-4.66721, 50.418691],
-                    [-4.668091, 50.418383],
-                    [-4.668287, 50.4179],
-                    [-4.668584, 50.416941],
-                    [-4.668792, 50.416639],
-                    [-4.669019, 50.416433],
-                    [-4.669705, 50.416559],
-                    [-4.671114, 50.416909],
-                    [-4.672381, 50.41711],
-                    [-4.672584, 50.417084],
-                    [-4.673289, 50.417102],
-                    [-4.674758, 50.417471],
-                    [-4.674687, 50.418707],
-                    [-4.674784, 50.419583],
-                    [-4.674881, 50.419786],
-                    [-4.674987, 50.419902],
-                    [-4.675042, 50.420103],
-                    [-4.675141, 50.420291],
-                    [-4.67522, 50.420422],
-                    [-4.675375, 50.420557],
-                    [-4.675538, 50.420842],
-                    [-4.675573, 50.420951],
-                    [-4.675543, 50.421071],
-                    [-4.675581, 50.42124],
-                    [-4.674922, 50.421736],
-                    [-4.674817, 50.421783],
-                    [-4.674221, 50.422257],
-                    [-4.67465, 50.422556],
-                    [-4.677089, 50.423753],
-                    [-4.677418, 50.423886],
-                    [-4.678187, 50.424083],
-                    [-4.678589, 50.424312],
-                    [-4.678737, 50.424487],
-                    [-4.678937, 50.42461],
-                    [-4.679012, 50.424611],
-                    [-4.679074, 50.42469],
-                    [-4.678917, 50.424926],
-                    [-4.678704, 50.425146],
-                    [-4.678339, 50.425327],
-                    [-4.67793, 50.425646],
-                    [-4.677422, 50.425881],
-                    [-4.676783, 50.426272],
-                    [-4.67654, 50.426311],
-                    [-4.676344, 50.426301],
-                    [-4.674954, 50.425751],
-                    [-4.672132, 50.4248]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "2013-07-15",
-              "end-date": "",
-              "entity": 1400044,
-              "name": "Battle of Lostwithiel 21 August 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1413619",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413619"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413619"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "2013-07-15",
-          "end-date": "",
-          "entity": 1400045,
-          "name": "Battle of Lostwithiel 31 August - 1 September 1644",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1413762",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-4.664135, 50.381245],
-                    [-4.663868, 50.380263],
-                    [-4.663418, 50.379504],
-                    [-4.663199, 50.379069],
-                    [-4.662411, 50.377951],
-                    [-4.662379, 50.377951],
-                    [-4.661905, 50.377526],
-                    [-4.66149, 50.377106],
-                    [-4.661198, 50.376711],
-                    [-4.660945, 50.376281],
-                    [-4.660798, 50.375942],
-                    [-4.66073, 50.375522],
-                    [-4.660697, 50.37547],
-                    [-4.660662, 50.374939],
-                    [-4.660686, 50.374898],
-                    [-4.66063, 50.374573],
-                    [-4.660516, 50.374248],
-                    [-4.66042, 50.374131],
-                    [-4.660377, 50.374007],
-                    [-4.660476, 50.373691],
-                    [-4.661211, 50.372517],
-                    [-4.661378, 50.372158],
-                    [-4.661367, 50.372013],
-                    [-4.661059, 50.371349],
-                    [-4.66161, 50.370976],
-                    [-4.662058, 50.370736],
-                    [-4.662557, 50.370542],
-                    [-4.663254, 50.370354],
-                    [-4.664479, 50.370159],
-                    [-4.664531, 50.37],
-                    [-4.664464, 50.369751],
-                    [-4.663919, 50.368987],
-                    [-4.6638, 50.36889],
-                    [-4.663905, 50.368841],
-                    [-4.663943, 50.368741],
-                    [-4.663842, 50.368446],
-                    [-4.663777, 50.368308],
-                    [-4.66345, 50.367973],
-                    [-4.663169, 50.367861],
-                    [-4.662246, 50.367776],
-                    [-4.661514, 50.367584],
-                    [-4.661067, 50.367365],
-                    [-4.660529, 50.366968],
-                    [-4.660351, 50.366766],
-                    [-4.660245, 50.366758],
-                    [-4.660178, 50.366848],
-                    [-4.659928, 50.366825],
-                    [-4.659478, 50.366417],
-                    [-4.658954, 50.366126],
-                    [-4.658849, 50.366037],
-                    [-4.658782, 50.365866],
-                    [-4.658739, 50.365632],
-                    [-4.658938, 50.365298],
-                    [-4.65898, 50.365175],
-                    [-4.658984, 50.365196],
-                    [-4.658985, 50.36516],
-                    [-4.658958, 50.365062],
-                    [-4.65882, 50.364894],
-                    [-4.658666, 50.364844],
-                    [-4.658617, 50.36479],
-                    [-4.658575, 50.364577],
-                    [-4.658451, 50.364293],
-                    [-4.658366, 50.364199],
-                    [-4.658225, 50.364142],
-                    [-4.658137, 50.364055],
-                    [-4.658098, 50.363796],
-                    [-4.658108, 50.363583],
-                    [-4.658165, 50.363448],
-                    [-4.657962, 50.363457],
-                    [-4.65764, 50.36317],
-                    [-4.656401, 50.362409],
-                    [-4.657406, 50.362176],
-                    [-4.658611, 50.361697],
-                    [-4.659392, 50.361491],
-                    [-4.662303, 50.361011],
-                    [-4.66289, 50.360832],
-                    [-4.663896, 50.360625],
-                    [-4.66498, 50.360518],
-                    [-4.665553, 50.360296],
-                    [-4.666732, 50.360211],
-                    [-4.667069, 50.360093],
-                    [-4.66717, 50.360185],
-                    [-4.667505, 50.360383],
-                    [-4.667723, 50.360446],
-                    [-4.667975, 50.360424],
-                    [-4.668847, 50.360206],
-                    [-4.669801, 50.359943],
-                    [-4.669458, 50.360037],
-                    [-4.669973, 50.359876],
-                    [-4.670465, 50.359658],
-                    [-4.670585, 50.359682],
-                    [-4.670741, 50.360185],
-                    [-4.670752, 50.360317],
-                    [-4.671448, 50.360265],
-                    [-4.671678, 50.360321],
-                    [-4.671853, 50.360534],
-                    [-4.672661, 50.361082],
-                    [-4.673332, 50.361275],
-                    [-4.673429, 50.361523],
-                    [-4.673508, 50.361582],
-                    [-4.673589, 50.361708],
-                    [-4.673675, 50.361878],
-                    [-4.673862, 50.362423],
-                    [-4.673839, 50.362592],
-                    [-4.673708, 50.362818],
-                    [-4.674128, 50.363082],
-                    [-4.673497, 50.363805],
-                    [-4.673389, 50.363967],
-                    [-4.673257, 50.364275],
-                    [-4.673345, 50.3644],
-                    [-4.673914, 50.364948],
-                    [-4.674073, 50.365177],
-                    [-4.67401, 50.365207],
-                    [-4.674192, 50.36596],
-                    [-4.674148, 50.366149],
-                    [-4.673933, 50.36653],
-                    [-4.673885, 50.366722],
-                    [-4.673982, 50.366792],
-                    [-4.673916, 50.366838],
-                    [-4.674279, 50.367025],
-                    [-4.674291, 50.367209],
-                    [-4.674253, 50.367272],
-                    [-4.674443, 50.367528],
-                    [-4.674614, 50.367636],
-                    [-4.67454, 50.367848],
-                    [-4.674343, 50.368171],
-                    [-4.675418, 50.368256],
-                    [-4.675494, 50.368303],
-                    [-4.675517, 50.368436],
-                    [-4.675614, 50.369299],
-                    [-4.675604, 50.369586],
-                    [-4.675527, 50.370082],
-                    [-4.675634, 50.370089],
-                    [-4.6752, 50.370816],
-                    [-4.675063, 50.37121],
-                    [-4.67515, 50.371217],
-                    [-4.674264, 50.372253],
-                    [-4.672647, 50.372437],
-                    [-4.67266, 50.372502],
-                    [-4.672611, 50.372747],
-                    [-4.672472, 50.372817],
-                    [-4.672438, 50.372944],
-                    [-4.672628, 50.373797],
-                    [-4.672758, 50.374009],
-                    [-4.673421, 50.373992],
-                    [-4.673372, 50.374125],
-                    [-4.674212, 50.374342],
-                    [-4.675238, 50.375011],
-                    [-4.675514, 50.375137],
-                    [-4.676136, 50.375333],
-                    [-4.676439, 50.375463],
-                    [-4.675985, 50.376134],
-                    [-4.675553, 50.376656],
-                    [-4.674856, 50.377379],
-                    [-4.674809, 50.377369],
-                    [-4.674708, 50.377521],
-                    [-4.674329, 50.378293],
-                    [-4.67428, 50.378826],
-                    [-4.674309, 50.378834],
-                    [-4.674331, 50.379177],
-                    [-4.674478, 50.379571],
-                    [-4.674688, 50.379832],
-                    [-4.67463, 50.379902],
-                    [-4.67316, 50.380242],
-                    [-4.672747, 50.379659],
-                    [-4.672466, 50.379351],
-                    [-4.672039, 50.379544],
-                    [-4.671531, 50.379875],
-                    [-4.670485, 50.380387],
-                    [-4.669935, 50.380722],
-                    [-4.669318, 50.381203],
-                    [-4.669305, 50.381264],
-                    [-4.66924, 50.381271],
-                    [-4.668823, 50.381537],
-                    [-4.668645, 50.381594],
-                    [-4.668576, 50.381589],
-                    [-4.668566, 50.381402],
-                    [-4.668086, 50.381413],
-                    [-4.668105, 50.381232],
-                    [-4.667885, 50.381091],
-                    [-4.667136, 50.381382],
-                    [-4.664356, 50.382066],
-                    [-4.664227, 50.381733],
-                    [-4.664135, 50.381245]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "2013-07-15",
-              "end-date": "",
-              "entity": 1400045,
-              "name": "Battle of Lostwithiel 31 August - 1 September 1644",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1413762",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413762"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413762"
-          }
-        },
-        {
-          "entry-date": "2021-08-31",
-          "start-date": "2013-10-07",
-          "end-date": "",
-          "entity": 1400046,
-          "name": "Battle of Edgcote 1469",
-          "dataset": "battlefield",
-          "typology": "geography",
-          "reference": "1413782",
-          "prefix": "battlefield",
-          "organisation-entity": "16",
-          "geojson": {
-            "geometry": {
-              "type": "MultiPolygon",
-              "coordinates": [
-                [
-                  [
-                    [-1.260209, 52.129685],
-                    [-1.260354, 52.129596],
-                    [-1.26054, 52.12922],
-                    [-1.260243, 52.128874],
-                    [-1.259931, 52.128666],
-                    [-1.259761, 52.128598],
-                    [-1.259626, 52.12861],
-                    [-1.259073, 52.128778],
-                    [-1.258766, 52.128813],
-                    [-1.258583, 52.128899],
-                    [-1.258162, 52.128931],
-                    [-1.25762, 52.129058],
-                    [-1.257029, 52.129126],
-                    [-1.25694, 52.129181],
-                    [-1.256883, 52.129268],
-                    [-1.256717, 52.129302],
-                    [-1.254846, 52.128933],
-                    [-1.25467, 52.128766],
-                    [-1.25467, 52.128602],
-                    [-1.254177, 52.12856],
-                    [-1.253621, 52.128577],
-                    [-1.253266, 52.128513],
-                    [-1.253173, 52.128455],
-                    [-1.253161, 52.128392],
-                    [-1.253333, 52.128309],
-                    [-1.253336, 52.128235],
-                    [-1.25328, 52.12813],
-                    [-1.25319, 52.128085],
-                    [-1.25236, 52.128185],
-                    [-1.252246, 52.128275],
-                    [-1.251978, 52.128628],
-                    [-1.251688, 52.128786],
-                    [-1.251033, 52.128691],
-                    [-1.250288, 52.128787],
-                    [-1.249652, 52.128723],
-                    [-1.249334, 52.128746],
-                    [-1.249251, 52.128688],
-                    [-1.249265, 52.128597],
-                    [-1.249238, 52.128549],
-                    [-1.249137, 52.128471],
-                    [-1.248931, 52.128504],
-                    [-1.248553, 52.128651],
-                    [-1.248232, 52.128654],
-                    [-1.248121, 52.128603],
-                    [-1.248023, 52.128467],
-                    [-1.247661, 52.12827],
-                    [-1.247689, 52.128179],
-                    [-1.247885, 52.128081],
-                    [-1.247917, 52.128011],
-                    [-1.247673, 52.127698],
-                    [-1.247633, 52.127596],
-                    [-1.24762, 52.127345],
-                    [-1.247506, 52.127272],
-                    [-1.24735, 52.12726],
-                    [-1.247115, 52.127382],
-                    [-1.246929, 52.127386],
-                    [-1.246876, 52.127341],
-                    [-1.246841, 52.127148],
-                    [-1.246636, 52.127014],
-                    [-1.246503, 52.127023],
-                    [-1.246005, 52.127197],
-                    [-1.245631, 52.127129],
-                    [-1.245493, 52.127054],
-                    [-1.245373, 52.126876],
-                    [-1.245291, 52.126815],
-                    [-1.245282, 52.126726],
-                    [-1.244971, 52.126673],
-                    [-1.244756, 52.126665],
-                    [-1.244445, 52.12684],
-                    [-1.244422, 52.126827],
-                    [-1.244508, 52.126944],
-                    [-1.244428, 52.126924],
-                    [-1.244312, 52.126799],
-                    [-1.24415, 52.12676],
-                    [-1.244143, 52.126663],
-                    [-1.243448, 52.1267],
-                    [-1.242646, 52.126681],
-                    [-1.24214, 52.126579],
-                    [-1.241856, 52.126434],
-                    [-1.241765, 52.126365],
-                    [-1.241295, 52.125649],
-                    [-1.240983, 52.125079],
-                    [-1.240931, 52.125037],
-                    [-1.240641, 52.124947],
-                    [-1.240603, 52.12489],
-                    [-1.240378, 52.124244],
-                    [-1.239861, 52.12226],
-                    [-1.23971, 52.12195],
-                    [-1.239648, 52.12192],
-                    [-1.239569, 52.121734],
-                    [-1.239417, 52.121591],
-                    [-1.238773, 52.121231],
-                    [-1.238547, 52.121036],
-                    [-1.237831, 52.120205],
-                    [-1.237135, 52.119317],
-                    [-1.236772, 52.118965],
-                    [-1.236689, 52.118847],
-                    [-1.236265, 52.117859],
-                    [-1.23541, 52.116087],
-                    [-1.235096, 52.114626],
-                    [-1.234942, 52.113739],
-                    [-1.234893, 52.113621],
-                    [-1.23418, 52.112619],
-                    [-1.232879, 52.111148],
-                    [-1.231881, 52.1105],
-                    [-1.231722, 52.110305],
-                    [-1.231624, 52.110288],
-                    [-1.231357, 52.109822],
-                    [-1.231261, 52.109586],
-                    [-1.230725, 52.108008],
-                    [-1.232748, 52.107645],
-                    [-1.233198, 52.107657],
-                    [-1.236041, 52.108475],
-                    [-1.237488, 52.10874],
-                    [-1.239874, 52.109354],
-                    [-1.2407, 52.109504],
-                    [-1.240907, 52.109212],
-                    [-1.240923, 52.108772],
-                    [-1.241178, 52.10781],
-                    [-1.241241, 52.107708],
-                    [-1.24133, 52.107355],
-                    [-1.24102, 52.106582],
-                    [-1.241015, 52.106514],
-                    [-1.241107, 52.106298],
-                    [-1.241404, 52.106073],
-                    [-1.24193, 52.105905],
-                    [-1.242466, 52.105689],
-                    [-1.243316, 52.105546],
-                    [-1.243312, 52.105523],
-                    [-1.243562, 52.105547],
-                    [-1.243908, 52.10551],
-                    [-1.245576, 52.105094],
-                    [-1.246054, 52.105472],
-                    [-1.246357, 52.105609],
-                    [-1.246849, 52.105913],
-                    [-1.247479, 52.106146],
-                    [-1.24763, 52.106165],
-                    [-1.248026, 52.106292],
-                    [-1.248707, 52.106653],
-                    [-1.24871, 52.106765],
-                    [-1.252089, 52.11067],
-                    [-1.252118, 52.11075],
-                    [-1.252306, 52.110754],
-                    [-1.252405, 52.110897],
-                    [-1.253528, 52.110807],
-                    [-1.253988, 52.110596],
-                    [-1.254001, 52.110532],
-                    [-1.255009, 52.110152],
-                    [-1.256367, 52.110171],
-                    [-1.256525, 52.109941],
-                    [-1.257615, 52.110823],
-                    [-1.258122, 52.111323],
-                    [-1.25928, 52.112099],
-                    [-1.25973, 52.112328],
-                    [-1.25894, 52.113684],
-                    [-1.258132, 52.114565],
-                    [-1.25795, 52.114707],
-                    [-1.25781, 52.114758],
-                    [-1.258834, 52.114845],
-                    [-1.260772, 52.115454],
-                    [-1.261482, 52.115527],
-                    [-1.261825, 52.115615],
-                    [-1.261774, 52.115644],
-                    [-1.26176, 52.115724],
-                    [-1.261711, 52.115728],
-                    [-1.261484, 52.116033],
-                    [-1.261594, 52.116187],
-                    [-1.261097, 52.117295],
-                    [-1.260909, 52.117909],
-                    [-1.260703, 52.118844],
-                    [-1.260771, 52.118843],
-                    [-1.260796, 52.118974],
-                    [-1.261036, 52.119302],
-                    [-1.261012, 52.119463],
-                    [-1.260811, 52.119776],
-                    [-1.260569, 52.119914],
-                    [-1.260468, 52.119874],
-                    [-1.260259, 52.120223],
-                    [-1.26014, 52.120349],
-                    [-1.26013, 52.120417],
-                    [-1.260028, 52.120508],
-                    [-1.261211, 52.120835],
-                    [-1.261788, 52.120825],
-                    [-1.261838, 52.120946],
-                    [-1.261915, 52.121001],
-                    [-1.262329, 52.121912],
-                    [-1.262763, 52.122591],
-                    [-1.263563, 52.12416],
-                    [-1.264095, 52.12541],
-                    [-1.264021, 52.125559],
-                    [-1.264101, 52.125824],
-                    [-1.264055, 52.126072],
-                    [-1.264082, 52.126108],
-                    [-1.264142, 52.126115],
-                    [-1.264107, 52.126635],
-                    [-1.264203, 52.126704],
-                    [-1.264368, 52.126727],
-                    [-1.264555, 52.126709],
-                    [-1.264406, 52.126943],
-                    [-1.264309, 52.127257],
-                    [-1.264332, 52.127474],
-                    [-1.264446, 52.127879],
-                    [-1.264396, 52.12813],
-                    [-1.264317, 52.128292],
-                    [-1.264084, 52.128486],
-                    [-1.263036, 52.129106],
-                    [-1.262996, 52.129192],
-                    [-1.26211, 52.129519],
-                    [-1.26203, 52.129469],
-                    [-1.261814, 52.129523],
-                    [-1.261743, 52.12966],
-                    [-1.26159, 52.129738],
-                    [-1.261183, 52.129761],
-                    [-1.260856, 52.129691],
-                    [-1.260471, 52.129683],
-                    [-1.260344, 52.129748],
-                    [-1.260364, 52.129882],
-                    [-1.260277, 52.129917],
-                    [-1.26005, 52.129843],
-                    [-1.260063, 52.129796],
-                    [-1.260209, 52.129685]
-                  ]
-                ]
-              ]
-            },
-            "type": "Feature",
-            "properties": {
-              "entry-date": "2021-08-31",
-              "start-date": "2013-10-07",
-              "end-date": "",
-              "entity": 1400046,
-              "name": "Battle of Edgcote 1469",
-              "dataset": "battlefield",
-              "typology": "geography",
-              "reference": "1413782",
-              "prefix": "battlefield",
-              "organisation-entity": "16",
-              "json": {
-                "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413782"
-              }
-            }
-          },
-          "json": {
-            "documentation-url": "https://historicengland.org.uk/listing/the-list/list-entry/1413782"
-          }
-        }
-      ]
-      
-
-    const purpleOptions = { color: 'purple' } 
-
-    return (
-        <div>
-            {/* {(battleData[0] !== undefined)? */}
-      <MapContainer center={[54.236, -4.54]} zoom={6}>
-      <TileLayer
-        attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-        url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-      />
-      {/* this under maps and displayes the battlefield. I tryed to have it in a different 
-      component, but for some reason it dosent like it */}
-   
-      <Polygon pathOptions={purpleOptions} positions={coordinates
-        .map(coordinateArray=> coordinateArray.reverse())} />
-    ))
-    
-    {/*the code bellow maps and renders the markers, this lives in a different component called MarkerElement  */}
-    {entities.map((entity, index) => (
-      <MarkerElement entity={entity} />
-    ))}
-    
-    </MapContainer>
-        </div>
-    )
-}
-
-export default SimpleMap
